### 2021-05-09 #3269 @whedon
[PRE REVIEW]: pySYD: Automated measurements of global asteroseismic parameters
**Submitting author:** @ashleychontos (<a href="http://orcid.org/0000-0003-1125-2564">Ashley Chontos</a>)
**Repository:** <a href="https://github.com/ashleychontos/pySYD" target ="_blank">https://github.com/ashleychontos/pySYD</a>
**Version:** v0.7.5
