### 2021-05-09 #3269 @arfon
:wave: @ashleychontos, thanks for your submission to JOSS. We have a reasonable number of submissions in our backlog right now so I'm going to mark this submission as _waitlisted_ until an editor with appropriate experience has time to take on this.

One thing you might want to work on while we're waiting to send this out for review is that I couldn't see any tests covering the functionality of the software. I would strongly encourage you to add a test suite for your package - this will almost certainly be flagged by any reviewers during the review process (https://joss.readthedocs.io/en/latest/review_criteria.html#tests)
