### 2021-05-17 #3269 @ashleychontos
@arfon hello and thank your quick response with comments and suggestions!

Re: the test suite:
I have thought about adding a separate test feature (i.e. one that does not depend on input data, science, etc.), as it would also be helpful for folks interested in developing the software. We do have an example star (with output) available in the [documentation](https://pysyd.readthedocs.io/en/latest/quickstart.html#summary). We have additional examples [here](https://pysyd.readthedocs.io/en/latest/examples.html), including detailed descriptions about the output figures as well as how to interpret them. In your experience(s), do you think this would suffice or should we consider other options? Thank you again!
