### 2021-05-17 #3269 @arfon
> We have additional examples here, including detailed descriptions about the output figures as well as how to interpret them. In your experience(s), do you think this would suffice or should we consider other options?

I think these are helpful, but if I were reviewing this submission, I'd be looking for a way to verify the functionality of the files in [this directory](https://github.com/ashleychontos/pySYD/tree/master/pysyd). As we describe in https://joss.readthedocs.io/en/latest/review_criteria.html#tests, there are multiple ways to do this, and example inputs/outputs is the _minimum_ we allow but I would [encourage you to go further](https://docs.python-guide.org/writing/tests/) :-)
