+++ pysyd/__init__.py
@@ -9,7 +9,7 @@ from .utils import *
 
 __all__ = ['cli', 'functions', 'pipeline', 'models', 'target', 'plots', 'utils']
 
-__version__ = '0.4.9'
+__version__ = '0.7.7'
 
 _ROOT = os.path.abspath(os.getcwd())
 TODODIR = os.path.join(_ROOT, 'info', 'todo.txt')
