### 2021-05-27 1cbfe0c
Some new CLI stuff
