+++ pysyd/cli.py
@@ -13,8 +13,6 @@ def main():
                                      prog='pySYD',
     )
 
-    sub_parser = parser.add_subparsers(title='subcommands', dest='subcommand')
-
     # In the parent parser, we define arguments and options common to all subcommands
     parent_parser = argparse.ArgumentParser(add_help=False)
     parent_parser.add_argument('-file', '--file', '-list', '--list', '-todo', '--todo',
@@ -46,75 +44,75 @@ def main():
                                default='results/',
     )
 
-    # Setting up
-    parser_setup = sub_parser.add_parser('setup', parents=[parent_parser],
-                                         description='Easy setup for directories and files')
-    parser_setup.set_defaults(func=pipeline.setup)
+    # Main options 
+    main_parser = argparse.ArgumentParser(add_help=False)
 
-    # Run pySYD
-    parser_run = sub_parser.add_parser('run', parents=[parent_parser],
-                                       description='Run pySYD')
-    parser_run.add_argument('-bg', '--bg', '-fitbg', '--fitbg', '-background', '--background',
+    main_parser.add_argument('-bg', '--bg', '-fitbg', '--fitbg', '-background', '--background',
                             dest='background',
                             help='Turn off the background fitting process (although this is not recommended)',
                             default=True, 
                             action='store_false',
     )
-    parser_run.add_argument('-ex', '--ex', '-findex', '--findex', '-excess', '--excess',
+    main_parser.add_argument('-cad', '--cad', '-cadence', '--cadence', 
+                            dest='cadence',
+                            help='Cadence of time series (in seconds), which will automatically be calculated when time series data is available.',
+                            type=int,
+                            default=0, 
+    )
+    main_parser.add_argument('-ex', '--ex', '-findex', '--findex', '-excess', '--excess',
                             dest='excess',
                             help='Turn off the find excess module. This is only recommended when a list of numaxes or a list of stellar parameters (to estimate the numaxes) are provided.',
                             default=True, 
                             action='store_false',
     )
-    parser_run.add_argument('-kc', '--kc', '-kepcorr', '--kepcorr',
+    main_parser.add_argument('-kc', '--kc', '-kepcorr', '--kepcorr',
                             dest='kepcorr',
                             help='Turn on Kepler short-cadence artefact corrections',
                             default=False, 
                             action='store_true',
     )
-    parser_run.add_argument('-nt', '--nt', '-nthread', '--nthread', '-nthreads', '--nthreads',
-                            dest='n_threads',
-                            help='Number of processes to run in parallel',
-                            type=int,
-                            default=0,
+    main_parser.add_argument('-nyq', '--nyq', '-nyquist', '--nyquist', 
+                            dest='nyquist',
+                            help='Nyquist frequency of power spectrum. Relevant for when the time series is not provided.',
+                            default=None,
     )
-    parser_run.add_argument('-ofa', '--ofa', '-of_actual', '--of_actual',
+    main_parser.add_argument('-ofa', '--ofa', '-of_actual', '--of_actual',
                             dest='of_actual',
                             help='Provide the actual oversampling factor of the power spectrum (provided there is no light curve data). Default is `0`, which means it is calculated from the time series data.',
                             type=int,
                             default=0,
     )
-    parser_run.add_argument('-ofn', '--ofn', '-of_new', '--of_new',
+    main_parser.add_argument('-ofn', '--ofn', '-of_new', '--of_new',
                             dest='of_new',
                             help='The desired oversampling factor for the newly-computed power spectrum. Default is `5`.',
                             type=int,
                             default=5,
     )
-    parser_run.add_argument('-os', '--os', '-over', '--over', '-oversample', '--oversample',
[…]
