### 2021-05-29 68eaeb5
fixed CLI options
