+++ pysyd/utils.py
@@ -17,11 +17,11 @@ from pysyd.models import *
 def get_info(args, parallel=False, stars=None, verbose=False, show=False, save=True, G=6.67428e-8, 
              teff_sun=5777.0, mass_sun=1.9891e33, radius_sun=6.95508e10, dnu_sun=135.1, 
              numax_sun=3090.0, width_sun=1300.0, tau_sun=[5.2e6,1.8e5,1.7e4,2.5e3,280.0,80.0], 
-             tau_sun_single=[3.8e6, 2.5e5, 1.5e5, 1.0e5, 230., 70.], kepcorr=False, groups=[],
+             tau_sun_single=[3.8e6,2.5e5,1.5e5,1.0e5,230.,70.], kepcorr=False, groups=[],
              oversample=True, of_actual=0, of_new=5):
     """
-    Loads todo.txt, sets up file paths, loads in any available star information,
-    and sets up matplotlib params.
+    Loads todo.txt, sets up file paths, loads in any available star information, saves the 
+    relevant parameters for each of the two main routines and sets the plotting parameters.
 
     Parameters
     ----------
@@ -71,7 +71,7 @@ def get_info(args, parallel=False, stars=None, verbose=False, show=False, save=T
     # Set file paths
     for star in args.stars:
         params[star] = {}
-        params[star]['path'] = '%s/%d/' % (args.outdir,star)
+        params[star]['path'] = '%s/%d/'%(args.outdir,star)
     args.params = params
 
     # Initialise parameters for the find excess routine
@@ -87,7 +87,7 @@ def get_info(args, parallel=False, stars=None, verbose=False, show=False, save=T
 
 def check_inputs(args):
     """ 
-    Make sure the command line inputs are the proper lengths and/or types
+    Make sure the command line inputs are the proper lengths and types
 
     Parameters
     ----------
@@ -115,23 +115,23 @@ def get_excess_params(args, n_trials=3, step=0.25, binning=0.005, smooth_width=1
     ----------
     args : argparse.Namespace
         the command line arguments
-    findex : dict
-        the parameters of the find excess routine
-    step : float
+    args.step : float
         TODO: Write description. Default value is `0.25`.
-    binning : float
+    args.binning : float
         logarithmic binning width. Default value is `0.005`.
-    n_trials : int
+    args.n_trials : int
         the number of trials. Default value is `3`.
-    lower : float
-        the lower frequency bound. Default value is `10.0`.
-    upper : float
-        the upper frequency bound. Default value is `4000.0`.
+    args.lower_x : float
+        the lower frequency bound (in muHz). Default value is `10.0` muHz.
+    args.upper_x : float
+        the upper frequency bound (in muHz). Default value is `4000.0` muHz.
 
     Returns
     -------
     args : argparse.Namespace
         the updated command line arguments
+    args.findex : Dict[str,object]
+        the parameters of the find excess routine
 
     """
 
@@ -163,42 +163,40 @@ def get_background_params(args, box_filter=2.5, mc_iter=1, ind_width=50, n_rms=2
     ----------
     args : argparse.Namespace
         the command line arguments
-
-    Uses
-    ----
-    fitbg : dict
-        the parameters relevant for the fit background routine, which is saved to args.fitbg
-    box_filter : args.box_filter
+    args.box_filter : float
         the size of the 1D box smoothing filter (in $\mu$Hz). Default value is `2.5`.
-    ind_width : int
+    args.ind_width : int
         the independent average smoothing width. Default value is `50`.
-    n_rms : int
+    args.n_rms : int
         number of data points to estimate red noise contributions. Default value is `20`.
-    n_peaks : int
+    args.n_peaks : int
[…]
