### 2021-06-01 01bc5c5
updating docs with new output in example
