+++ pysyd/functions.py
@@ -10,6 +10,22 @@ def set_seed(star, lower=1, upper=10**7, size=1):
     For Kepler targets that require a correction via CLI (--kc), a random seed is generated
     from U~[1,10^7] and stored in stars_info.csv for reproducible results in later runs.
 
+    Parameters
+    ----------
+    star : target.Target
+        the pySYD pipeline object
+    lower : int 
+        lower limit for random seed value. Default value is `1`.
+    upper : int
+        upper limit for random seed value. Default value is `10**7`.
+    size : int
+        number of seed values returned. Default value is `1`.
+
+    Returns
+    -------
+    star : target.Target
+        the pySYD pipeline object
+        
     """
 
     seed = list(np.random.randint(lower,high=upper,size=size))
@@ -36,13 +52,22 @@ def remove_artefact(star, lcp=1.0/(29.4244*60*1e-6), lf_lower = [240.0, 500.0], 
 
     Parameters
     ----------
-    frequency : np.ndarray
-        the frequency of the power spectrum
-    power : np.ndarray
-        the power of the power spectrum
+    star : target.Target
+        the pySYD pipeline object
     lcp : float
-        TODO: Write description. Default value is `1/(29.4244*60*1e-6)`.
-
+        long cadence period in Msec
+    lf_lower : List[float]
+        lower limit of low frequency artefact
+    lf_upper : List[float]
+        upper limit of low frequency artefact
+    hf_lower : List[float]
+        lower limit of high frequency artefact
+    hf_upper : List[float]
+        upper limit of high frequency artefact
+    Returns
+    -------
+    star : target.Target
+        the pySYD pipeline object
     """
 
     if star.params[star.name]['seed'] is None:
@@ -82,7 +107,7 @@ def remove_artefact(star, lcp=1.0/(29.4244*60*1e-6), lf_lower = [240.0, 500.0], 
     return star
 
 
-def gaussian_bounds(x, y, guesses, best_x=None, sigma=None):
+def gaussian_bounds(x, y, guesses, sigma=None):
     """Get the bounds for the parameters of a Gaussian fit to the data.
 
     Parameters
@@ -91,16 +116,17 @@ def gaussian_bounds(x, y, guesses, best_x=None, sigma=None):
         the x values of the data
     y : np.ndarray
         the y values of the data
-    best_x : Optional[float]
-        TODO: Write description. Default value is `None`.
+    guesses : List[float]
+        initial guesses for a gaussian fit
     sigma : Optional[float]
-        TODO: Write description. Default value is `None`.
+        sigma from a standard Normal distribution 
 
     Returns
     -------
     bb : List[Tuple]
         list of parameter bounds of a Gaussian fit to the data
     """
+
     offset, amp, center, width = guesses
     if sigma is None:
         sigma = (max(x)-min(x))/8.0/np.sqrt(8.0*np.log(2.0))
@@ -214,6 +240,7 @@ def mean_smooth_ind(x, y, width):
     se : np.ndarray
         standard error
     """
+
     step = width-1
     j=0
     
@@ -235,114 +262,6 @@ def mean_smooth_ind(x, y, width):
     sy = sy[(sy != 0.0)]
     se[(se == 0.0)] = np.median(se)
     return sx, sy, se
-
-
-def smooth(array, width, params, method='box', mode=None, fft=False, silent=False):
-    """Smooths using a variety of methods. TODO: Write description.
-
-    Parameters
-    ----------
-    array : np.ndarray
-        the data
-    TODO: Add parameters
-
[…]
