+++ pysyd/target.py
@@ -68,6 +68,10 @@ class Target:
         2) estimates the stellar background contributions before estimating the
            global asteroseismic parameters
 
+        Returns
+        ----------
+        None
+
         """
         # Run the find excess routine
         if self.params[self.name]['excess']:
@@ -84,6 +88,10 @@ class Target:
         """
         Automatically finds power excess due to solar-like oscillations using a
         frequency-resolved, collapsed autocorrelation function (ACF).
+
+        Returns
+        ----------
+        None
 
         """
         # Make sure the binning is specified, otherwise it cannot run
@@ -120,16 +128,24 @@ class Target:
 
     def collapsed_acf(self, b, j=0, start=0, max_iterations=5000, max_snr=100.):
         """
-        TODO
+        Computes a collapsed autocorrelation function (ACF).
 
         Parameters
         ----------
         b : int
             the trial number
         j : int
+            index at which to start storing the cumulative sum and mean of ACF. Default value is `0`.
         start : int
+            index at which to start masking the frequency and power spectrum. Default value is `0`.
         max_iterations : int
+            maximum number of interations to try in curve fitting routine. Default value is `5000`.
         max_snr : float
+            maximum SNR corresponding to power excess. Default value is `100.0`.
+
+        Returns
+        ----------
+        None
 
         """
         # Computes a collapsed ACF using different "box" (or bin) sizes
@@ -194,6 +210,10 @@ class Target:
         Perform a fit to the granulation background and measures the frequency of maximum power (numax),
         the large frequency separation (dnu) and oscillation amplitude.
 
+        Returns
+        ----------
+        None
+
         """
         self.fitbg['results'][self.name] = {'numax_smooth':[],'A_smooth':[],'numax_gauss':[],'A_gauss':[],
                                             'FWHM':[],'dnu':[],'white':[]}
@@ -234,7 +254,7 @@ class Target:
     def first_step(self):
         """
         The first step in the background fitting, which determines the best-fit stellar 
-        contribution model (i.e. number of Harvey-like components) and corrects for this
+        contribution model (i.e. number of Harvey-like components) and corrects for this.
 
         TODO: implement more robust criterion (i.e. BIC or AIC) and also include the simplest model.
         before estimating numax and dnu.
@@ -246,6 +266,11 @@ class Target:
                      4) takes the autocorrelation (using ffts) of the masked power spectrum that contains the power excess
                      5) selects the peak (via -npeaks, default=10) closest to the expected spacing based on the calculated numax
                      6) fits Gaussian to the "cutout" peak of the ACF, where center is dnu
+        
+        Returns
+        ----------
+        result : bool
+            will return `True` if the model converges, otherwise `False`.
 
         """
         # Estimate white noise level
@@ -283,6 +308,10 @@ class Target:
         when the args.mciter > 1, where args.mciter = 200 is typically sufficient for 
         estimating an uncertainty.
 
+        Returns
+        ----------
+        None
+
         """
         # Randomize power spectrum to get uncertainty on measured values
         self.random_pow = (np.random.chisquare(2, len(self.frequency))*self.power)/2.
@@ -316,6 +345,10 @@ class Target:
         Estimate the white noise level (in muHz) by taking a mean over a region 
         in the power spectrum near the nyquist frequency.
 
+        Returns
+        ----------
+        None
+
         """
         if self.nyquist < 400.0:
             mask = (self.frequency > 200.0) & (self.frequency < 270.0)
@@ -336,6 +369,10 @@ class Target:
         Estimates amplitude of red noise components by using a smoothed version of the power
[…]
