### 2021-06-02 beef2ca
added low SNR detection
