+++ docs/examples.rst
@@ -27,9 +27,9 @@ KIC 1435467 is our least evolved example star, with numax ~1400 muHz.
 | **Top left:** Original time series.  
 | **Top middle:** Original power spectrum (white) and heavily smoothed power spectrum (green). The latter is used as an initial (crude) background fit to search for oscillations.  
 | **Top right:** Power spectrum after correcting the crude background fit.  
-**Bottom left:** Frequency-resolved, collapsed autocorrelation function of the background-corrected power spectrum using a small step size. This step size is optimized for low-frequency oscillators. The green line is a Gaussian fit to the data, which provides the initial numax estimate.  
-**Bottom middle:** Same as bottom left but for the medium step size (optimized for subgiant stars).  
-**Bottom right:** Same as bottom left but for the large step size (optimized for main-sequence stars).
+| **Bottom left:** Frequency-resolved, collapsed autocorrelation function of the background-corrected power spectrum using a small step size. This step size is optimized for low-frequency oscillators. The green line is a Gaussian fit to the data, which provides the initial numax estimate.  
+| **Bottom middle:** Same as bottom left but for the medium step size (optimized for subgiant stars).  
+| **Bottom right:** Same as bottom left but for the large step size (optimized for main-sequence stars).
 
 
 ``fit_background`` results:
@@ -38,15 +38,15 @@ KIC 1435467 is our least evolved example star, with numax ~1400 muHz.
   :width: 600
   :alt: Fit background output plot for KIC 1435467.
   
-**Top left:** Original time series. 
-**Top middle:** Original power spectrum (white), lightly smoothed power spectrum (red), and binned power spectrum (green). Blue lines show initial guesses of the fit to the granulation background. The grey region is excluded from the background fit based on the numax estimate provided to the module.
-**Top right:** Same as top middle but now showing the best fit background model (blue) and a heavily smoothed version of the power spectrum (yellow)
-**Center left:** Background corrected, heavily smoothed power spectrum (white). The blue line shows a Gaussian fit to the data (used to calculate numax_gaussian) and the red square is the peak of the smoothed, background corrected power excess (numax_smoothed).
-**Center:** Lightly smoothed, background corrected power spectrum centered on numax. 
-**Center right:** Autocorrelation function of the data in the center panel. The red dotted line shows the estimate Dnu value given the input numax value, and the red region shows the extracted ACF peak that will be used to measure Dnu. The yellow line shows the Gaussian weighting function used to define the red region.
-**Bottom left:** ACF peak extracted in the center right panel (white) and a Gaussian fit to that peak (green). The center of the Gaussian is the estimate of Dnu.
-**Bottom middle:** Echelle diagram of the background corrected power spectrum using the measured Dnu value.
-**Bottom right:** Echelle diagram collapsed along the frequency direction.
+| **Top left:** Original time series. 
+| **Top middle:** Original power spectrum (white), lightly smoothed power spectrum (red), and binned power spectrum (green). Blue lines show initial guesses of the fit to the granulation background. The grey region is excluded from the background fit based on the numax estimate provided to the module.
+| **Top right:** Same as top middle but now showing the best fit background model (blue) and a heavily smoothed version of the power spectrum (yellow)
+| **Center left:** Background corrected, heavily smoothed power spectrum (white). The blue line shows a Gaussian fit to the data (used to calculate numax_gaussian) and the red square is the peak of the smoothed, background corrected power excess (numax_smoothed).
+| **Center:** Lightly smoothed, background corrected power spectrum centered on numax. 
[…]
