+++ docs/examples.rst
@@ -30,8 +30,7 @@ KIC 1435467 is our least evolved example star, with numax ~1400 muHz.
 | **Bottom left:** Frequency-resolved, collapsed autocorrelation function of the background-corrected power spectrum using a small step size. This step size is optimized for low-frequency oscillators. The green line is a Gaussian fit to the data, which provides the initial numax estimate.  
 | **Bottom middle:** Same as bottom left but for the medium step size (optimized for subgiant stars).  
 | **Bottom right:** Same as bottom left but for the large step size (optimized for main-sequence stars).
-
-
+|
 
 ``fit_background`` results:
 
@@ -48,7 +47,7 @@ KIC 1435467 is our least evolved example star, with numax ~1400 muHz.
 | **Bottom left:** ACF peak extracted in the center right panel (white) and a Gaussian fit to that peak (green). The center of the Gaussian is the estimate of Dnu.
 | **Bottom middle:** Echelle diagram of the background corrected power spectrum using the measured Dnu value.
 | **Bottom right:** Echelle diagram collapsed along the frequency direction.
-
+|
 
 
 ``sampling`` results:
