+++ docs/examples.rst
@@ -1,4 +1,5 @@
 .. _examples:
+.. math::
 
 Examples
 ########
