### 2021-06-04 #3331 @whedon
[REVIEW]: pySYD: Automated measurements of global asteroseismic parameters
**Submitting author:** @ashleychontos (<a href="http://orcid.org/0000-0003-1125-2564">Ashley Chontos</a>)
**Repository:** https://github.com/ashleychontos/pySYD
**Branch with paper.md** (empty if default branch): master
**Version:** v6.10.0
