+++ README.md
@@ -16,7 +16,7 @@
 
 Written by Ashley Chontos. Developed by Ashley Chontos, Daniel Huber, Maryum Sayeed and Pavadol Yamsiri.
 
-We thank Dennis Stello, Jie Yu, Marc Hon, and other users of the `SYD` pipeline for discussion and suggestions which helped the development of this code.
+We thank Dennis Stello, Jie Yu, Marc Hon, and other users of the IDL `SYD` pipeline for discussion and suggestions which helped the development of this code.
 
 Please cite our [JOSS paper](https://github.com/openjournals/joss-papers/blob/joss.03331/joss.03331/10.21105.joss.03331.pdf) as well as the [original publication](https://ui.adsabs.harvard.edu/abs/2009CoAst.160...74H/abstract) discussing the methodology if you make use of `pySYD` in your work.
 
