### 2021-06-07 #3331 @benjaminpope
### Intro Comments:
- Thanks for inviting me to review the pySYD release - this is an important piece of software that will bring asteroseismology from the closed-source dark ages to an open source renaissance (I hope!). IDL-SYD is well-used and well-known, but together with some other major pipelines has not been available open-source, which means that asteroseismology as a field suffers from reproducibility and inaccessibility. pySYD will be an important step forward that, I am pretty sure, will enable a great deal of exciting science.
- The release of pySYD as documented so far isn't quite where we need to be for JOSS or for optimal use for the community. In particular, there are two high level issues:
	- The testing of the pySYD software against IDL-SYD is not documented in this paper, or cited from elsewhere, so there is actually a gap in reproducibility introduced.
	- The documentation of pySYD is not quite at the stage where it will be optimally user-friendly, and we probably need to have a bit more hand-holding with the examples to help people like me pick it up and play.
- It is also not clear to me whether this is intended as a purely command-line tool, or whether the API is sufficiently useful for working with other code, for example in a Jupyter notebook or as part of a larger script. If this is beyond the scope of the project it is fine, but I didn't get a sense from the paper or the documentation what kind of workflow was expected, supported, or possible.

### Installation:
- pip install works, and pysyd setup works too.
- Could pysyd setup put data in an install directory, rather than in the current directory?

### Unit tests:
- As far as I can tell, there are no unit tests or continuous integration solutions, at least as-is on the GitHub repo. The JOSS review requires automated tests.

### Repository:
- There are no community guidelines as required by JOSS.

### Examples:
- We get nice and fast performance from pysyd run -star 1435467 -show -verbose as expected from the Getting Started / Example Fit part of the readthedocs. On this page, you should include the expected output though - so we know if the graphs are working ok.
- The output from this example differs from the one on the page! I understand this is because of the differing -mc flag, but I'm not sure why one command is shown in full (with no output), and the other output is shown in full (but no command).
- The Quickstart example should have an example output. Also not clear how this is different from Example Fit.
- The examples are nice graphs... but supposing we did pip install, it's not immediately obvious where the data live, and no commands are given to reproduce the examples. So after "you are ready to test out the software!" there should be a few worked examples showing commands and outputs!
- The figures take a very long time to render and close, at least on my MacBook Pro. Perhaps the examples might like to have -save flags by default so that these don't cause an issue.

### Manuscript:
- Summary - this probably needs to be reworked. Currently it is a summary of _asteroseismology_, but not of pySYD. You can cut down some of the explanation of physics if necessary, but certainly need to add a bit about what pySYD is and what you use it for.
- In first paragraphs - there is a comment that SYD has been tested against closed-source pipelines. It is my understanding that SYD itself is not yet open source, and that this very paper is the open-source release - so you want to clarify whether SYD too is closed-source.
- Should cite Borucki and Ricker papers for Kepler and TESS citations.
- The comparison of pySYD and IDL-SYD shows nonzero scatter... but this is ostensibly a _translation_ of the IDL-SYD pipeline, so I would expect them to be even closer. This is important because if reproducibility of legacy Kepler results is a goal, systematics in IDL-SYD should also be reproduced! If it is a reworking of the same ideas, but not a translation, this will differ. There should probably be some comments about this. 0.5% is not quite trivial, even if the uncertainties are typically of order ~ %, if you average over many stars.
- The IDL-SYD to pySYD benchmarks are not open source - are these linked anywhere? This comparison is not reproducible - we don't know which stars these are, or what the _exact_ time baselines are (3-11 months is quite a range). I think this comparison is an important part of the paper, given pySYD is about reproducibility and extension of the IDL-SYD pipeline.
- I think it should say earlier in the paper that this is about solar-like oscillations, as opposed to (say) classical pulsations.
