### 2021-06-09 #3331 @danhey
Hi all, this has been a pleasure to review! It's a great piece of software, and I have no doubt it will be extremely useful for asteroseismology. Overall, I believe pySYD satisfies the JOSS criteria for publication after a few points have been addressed.

I agree with all of Ben's comments, especially regarding the distinction between a command-line vs. API tool. It appears to me that the functionality is there for using pySYD from say, a Jupyter notebook, but this has not been documented.
I will try to structure my review in the same format as Ben's.

### Installation
- Installation is fine from a fresh Conda environment. However, `tqdm` appears to be a requirement that is not listed in pySYD.

### Unit tests
- I agree with Ben here, there don't appear to be any unit tests! It's important to have unit tests, especially for a tool that will hopefully become a standard in asteroseismology. I suggest at the minimum, you include tests for the core functionality (i.e., input a star -> ensure the outputs are exactly the same). If you need some inspiration on how testing is done for an astronomy Python package, Lightkurve does a very good job of it. Check out their [tests](https://github.com/lightkurve/lightkurve/tree/main/tests), and their [GitHub actions workflow](https://github.com/lightkurve/lightkurve/blob/main/.github/workflows/lightkurve-tests.yml).

### Repository
- No community guidelines.
- Perhaps a minor oversight but pySYD doesn't seem to specify which Python version(s) it is compatible with. I know many astronomers still use Python 2.7 for compatibility, will pySYD work with that?

### Examples and functionality
- The examples all work fine for me.
- it is not clear in the documentation that the plots *must* be closed for the pipeline to proceed. i.e., when I ran the “pysyd run -star 1435467 -show -verbose” command, I had to close the plots as they appeared for the pipeline to continue. They should either automatically close or a prompt should be displayed for continuing.
- Would it be prudent to include an example for the TESS data? And on that note, how does PySyd deal with significant gaps in the data? This is kind of a new issue because the Kepler data was so good! I’ve noticed that there’s a `pysyd.utils.stitch_data()` function for gapped data, perhaps a warning should also be thrown when this function is used? This could be fixed in one step: provide an example for a TESS star that has a large data gap in it!

### Manuscript
- Overall the manuscript is well written and clear, I like it!
- The scatter in the IDL-SYD and pySYD comparison plots also concerns me a bit. If the goal is to completely reproduce the functionality of the IDL version then there should be little to no scatter. Do we know what causes these deviations, or, are they purely the result of random sampling? And which pipeline output would we expect to give the more accurate result?
- Regarding the open sourcing of the benchmarks:
> Providing the power spectra to reproduce this figure would be about ~300 MB of data and therefore probably too large to provide directly on our GitHub repo.

Only a suggestion, but would it be easier to instead write a small utility function in pySYD that calculates the power spectrum given an input light curve? This would also be useful for the general functionality of PySYD, allowing people to supply a light curve. It would also benefit reproducibility: I have seen a lot of different power spectra normalizations in my time ...
