### 2021-06-11 887ba81
Update examples.rst
