+++ docs/advanced.rst
@@ -1,12 +1,12 @@
+.. _cli:
+
+Command Line Interface
+######################
+
 .. _advanced:
 
 Advanced Usage
 ##############
-
-.. _cli:
-
-Command Line Interface
-======================
 
 By default, initialization of the ``pySYD`` command will use the paths:
 
@@ -15,100 +15,275 @@ By default, initialization of the ``pySYD`` command will use the paths:
 - ``INPDIR`` : '~/path_to_put_pysyd_stuff/data'
 - ``OUTDIR`` : '~/path_to_put_pysyd_stuff/results'
 
-which is based on the absolute path of the current working directory. All of these paths should be ready to go
-if you followed the suggestions in :ref:`structure`, used our ``setup`` feature or ran the :ref:`sixtyseconds`
-challenge.
+which is based on the absolute path of the current working directory (or however you choose to set it up). All of these paths should be ready to go
+if you followed the suggestions in :ref:`structure` or used our ``setup`` feature.
 
-Parent Parser
-+++++++++++++
-
-The ``pySYD`` command line feature has three subcommands: ``setup``, ``run`` and ``parallel``, the former which should only really
-be used once after installation. Command line options inherent to both subcommands include:
-
-- ``-file``, ``--file``, ``-list``, ``--list``, ``-todo``, ``--todo`` : path to text file that contains the list of stars to process
-   * dest = ``args.file``
-   * type = string
-   * default = ``TODODIR``
-- ``-in``, ``--in``, ``-input``, ``--input``, ``-inpdir``, ``--inpdir`` : path to input data
-   * dest = ``args.inpdir``
-   * type = string
-   * default = ``INPDIR``
-- ``-info``, ``--info``, ``-information``, ``--information`` : path to the csv containing star information
-   * dest = ``args.info``
-   * type = string
-   * default = ``INFODIR``
-- ``-out``, ``--out``, ``-output``, ``--output``, ``-outdir``, ``--outdir`` : path that results are saved to
-   * dest = ``args.outdir``
-   * type = string
-   * default = ``OUTDIR``
-- ``-verbose``, ``--verbose`` : turn on verbose output
-   * dest = ``args.verbose``
-   * type = boolean
-   * default = ``False``
-   * action = ``store_true``
-   
+There are currently three modes that ``pySYD`` can operate in via command line: ``setup``, ``run`` and ``parallel``. These are discussed in more detail below, including what commands and options are available in each mode.
 
 Setup
-+++++
+****
+
+Inherits=[parent_parser]
 
 Initializes ``pysyd.pipeline.setup`` for quick and easy setup of directories, files and examples. 
 
-
 Run
-+++
+***
+
+Inherits=[parent_parser,main_parser]
 
 The main pySYD pipeline function initialized through ``pysyd.pipeline.run`` occurs in two main steps: 
-``find_excess`` and ``fit_background``. CLI options relevant to pre-pipeline initialization or both
-submodules include:
+``find_excess`` and ``fit_background``.
 
-- ``-bg``, ``--bg``, ``-fitbg``, ``--fitbg``, ``-background``, ``--background`` : turn off the background fitting procedure
-   * dest = ``args.background``
-   * type = boolean
-   * default = ``True``
-   * action = ``store_false``
-- ``-ex``, ``--ex``, ``-findex``, ``--findex``, ``-excess``, ``--excess`` : turn off the find excess module
-   * dest = ``args.background``
-   * type = boolean
-   * default = ``True``
-   * action = ``store_false``
[…]
