+++ requirements.txt
@@ -2,4 +2,4 @@ astropy>=4.0.0
 matplotlib>=1.5.3
 numpy>=1.16
 pandas>=1.0.5
-scipy>=1.5.0
+scipy==1.5.1
