+++ pysyd/utils.py
@@ -2,6 +2,7 @@ import os
 import glob
 import numpy as np
 import pandas as pd
+from tqdm import tqdm
 from itertools import chain
 from astropy.io import ascii
 import multiprocessing as mp
@@ -14,11 +15,7 @@ from pysyd.models import *
 
 
 
-def get_info(args, parallel=False, stars=None, verbose=False, show=False, save=True, G=6.67428e-8, 
-             teff_sun=5777.0, mass_sun=1.9891e33, radius_sun=6.95508e10, dnu_sun=135.1, 
-             numax_sun=3090.0, width_sun=1300.0, tau_sun=[5.2e6,1.8e5,1.7e4,2.5e3,280.0,80.0], 
-             tau_sun_single=[3.8e6,2.5e5,1.5e5,1.0e5,230.,70.], kepcorr=False, groups=[],
-             oversample=True, of_actual=0, of_new=5):
+def get_info_all(args, parallel, CLI=True):
     """
     Loads todo.txt, sets up file paths, loads in any available star information, saves the 
     relevant parameters for each of the two main routines and sets the plotting parameters.
@@ -26,88 +23,65 @@ def get_info(args, parallel=False, stars=None, verbose=False, show=False, save=T
     Parameters
     ----------
     args : argparse.Namespace
-        command line arguments
-    star_info : str
-        the file path to the star_info.csv file containing star information. Default value is `'info/star_info.csv'`
-    params : dict
-        the pipeline parameters, which is saved to args.params
+        command-line arguments
+    parallel : bool
+        if pysyd will be running in parallel mode
+    CLI : bool, optional
+        if CLI is not being used (i.e. `False`), the modules draw default values from a different location
 
     Returns
     -------
     args : argparse.Namespace
-        the updated command line arguments
+        the updated command-line arguments
+
     """
-
-    params={}
-    # Open star list
-    if args.stars is None or args.stars == []:
-        with open(args.todo, "r") as f:
-            args.stars = [int(float(line.strip().split()[0])) for line in f.readlines()]
-    check_inputs(args)
-    params['inpdir'] = args.inpdir
-    params['outdir'] = args.outdir
-    if parallel:
-        todo = np.array(args.stars)
-        args.verbose = False
-        args.show = False
-        if args.n_threads == 0:
-            args.n_threads = mp.cpu_count()
-        if len(todo) < args.n_threads:
-            args.n_threads = len(todo)
-        # divide stars into groups set by number of cpus/nthreads available
-        digitized = np.digitize(np.arange(len(todo))%args.n_threads,np.arange(args.n_threads))
-        groups = np.array([todo[digitized == i] for i in range(1, args.n_threads+1)], dtype=object)
-    else:
-        groups = np.array([])
-
-    # Adding constants and the star list
-    params.update({
-        'numax_sun':3090.0, 'width_sun':1300.0, 'stars':args.stars, 'G':6.67428e-8, 
-        'show':args.show, 'oversample':args.oversample, 'tau_sun':[5.2e6,1.8e5,1.7e4,2.5e3,280.0,80.0], 
-        'tau_sun_single': [3.8e6, 2.5e5, 1.5e5, 1.0e5, 230., 70.], 'radius_sun': 6.95508e10, 'save':args.save, 
-        'teff_sun':5777.0, 'mass_sun':1.9891e33, 'dnu_sun':135.1, 'kepcorr': args.kepcorr, 'groups':groups, 
-        'of_actual':args.of_actual, 'of_new':args.of_new,
-    })
-    # Set file paths
-    for star in args.stars:
-        params[star] = {}
-        params[star]['path'] = '%s/%d/'%(args.outdir,star)
-    args.params = params
-
-    # Initialise parameters for the find excess routine
[…]
