+++ pysyd/target.py
@@ -1,6 +1,5 @@
 import numpy as np
 from tqdm import tqdm
-from scipy.stats import chisquare
 from scipy.signal import find_peaks
 from scipy.optimize import curve_fit
 from scipy.interpolate import InterpolatedUnivariateSpline
@@ -33,8 +32,6 @@ class Target:
         the parameters of the fit background routine
     verbose : bool
         if true, turns on the verbose output
-    oversample : bool
-        if true, uses an oversampled power spectrum for the first iteration to estimate parameters (default is `False`)
 
     Parameters
     ----------
@@ -52,13 +49,9 @@ class Target:
         self.params = args.params
         self.findex = args.findex
         self.fitbg = args.fitbg
+        self.globe = args.globe
         self.verbose = args.verbose
-        self.oversample = args.oversample
         self = utils.load_data(self, args)
-        if self.ps:
-            self.run = 1
-        else:
-            self.run = 0
 
 
     def run_syd(self):
@@ -73,11 +66,11 @@ class Target:
         if self.params[self.name]['excess']:
             self = utils.get_findex(self)
             self.find_excess()
-        # Run the fit background routine
+        # Run the global fitting routine
         if self.params[self.name]['background']:
             if utils.check_fitbg(self):
                 self = utils.get_fitbg(self)
-                self.fit_background()
+                self.fit_global()
 
 
     def find_excess(self):
@@ -88,13 +81,16 @@ class Target:
         """
         # Make sure the binning is specified, otherwise it cannot run
         if self.findex['binning'] is not None:
-            self.bin_freq, self.bin_pow = functions.bin_data(self.freq, self.pow, self.findex)
+            self.bin_freq, self.bin_pow, _ = functions.bin_data(self.freq, self.pow, width=self.findex['binning'], log=True, mode=self.findex['mode'])
             if self.verbose:
                 print('----------------------------------------------------')
                 print('Running find_excess module:')
                 print('PS binned to %d datapoints' % len(self.bin_freq))
+                print('Binned freq res: %.2f muHz'%(self.bin_freq[1]-self.bin_freq[0]))
             # Smooth the binned power spectrum for a rough estimate of background
-            boxsize = int(np.ceil(float(self.findex['smooth_width'])/(self.bin_freq[1]-self.bin_freq[0])))
+            boxsize=int(np.ceil(self.findex['smooth_width']/(self.bin_freq[1]-self.bin_freq[0])))
+            if boxsize >= len(self.bin_freq):
+                boxsize=int(np.ceil(len(self.bin_freq)/10.))
             sp = convolve(self.bin_pow, Box1DKernel(boxsize))
             smooth_freq = self.bin_freq[int(boxsize/2):-int(boxsize/2)]
             smooth_pow = sp[int(boxsize/2):-int(boxsize/2)]
@@ -113,12 +109,12 @@ class Target:
             # Select trial that resulted with the highest SNR detection
             self.findex['results'][self.name]['best'] = self.compare.index(max(self.compare))+1
             if self.verbose:
-                print('selecting model %d' % self.findex['results'][self.name]['best'])
+                print('selecting model %d'%self.findex['results'][self.name]['best'])
             utils.save_findex(self)
             plots.plot_excess(self)
 
 
-    def collapsed_acf(self, b, j=0, start=0, max_iterations=5000, max_snr=100.):
+    def collapsed_acf(self, b, start=0, max_iterations=5000, max_snr=100.):
         """
         TODO
 
@@ -126,51 +122,42 @@ class Target:
         ----------
         b : int
             the trial number
-        j : int
         start : int
+            what index of the frequency array to start with, which is `0` by default.
         max_iterations : int
[…]
