+++ pysyd/target.py
@@ -62,7 +62,7 @@ class Target:
            global asteroseismic parameters
 
         Returns
-        ----------
+        -------
         None
 
         """
@@ -242,65 +242,9 @@ class Target:
                     utils.verbose_output(self)
 
 
-    def first_step(self):
-        """
-        The first step in the background fitting, which determines the best-fit stellar 
-        contribution model (i.e. number of Harvey-like components) and corrects for this.
-
-        TODO: implement more robust criterion (i.e. BIC or AIC) and also include the simplest model.
-        before estimating numax and dnu.
-
-        single_step: The main iteration of the background fitting, which operates in the following steps:
-                     1) determines the best-fit model (i.e. number of Harvey components) using a reduced chi-sq analysis
-                     2) corrects for stellar background contributions by dividing the power spectrum by the best-fit model
-                     3) estimates two values for numax by fitting a Gaussian and by using a heavy smoothing filter
-                     4) takes the autocorrelation (using ffts) of the masked power spectrum that contains the power excess
-                     5) selects the peak (via -npeaks, default=10) closest to the expected spacing based on the calculated numax
-                     6) fits Gaussian to the "cutout" peak of the ACF, where center is dnu
-        
-        Returns
-        ----------
-        result : bool
-            will return `True` if the model converges, otherwise `False`.
-
-        """
-        while self.i < self.fitbg['mc_iter']:
-            # Requires convergence of background fit before going to the next step 
-            if self.fit_background():
-                self.get_numax()
-                self.get_dnu()
-                # First step?
-                if self.i == 0:
-                    # Plot results
-                    plots.plot_background(self)
-                    if self.fitbg['mc_iter'] > 1:
-                        # Switch to critically-sampled PS if sampling
-                        mask = np.ma.getmask(np.ma.masked_inside(self.freq_cs, self.params[self.name]['bg_mask'][0], self.params[self.name]['bg_mask'][1]))
-                        self.frequency, self.power = np.copy(self.freq_cs[mask]), np.copy(self.pow_cs[mask])
-                        self.resolution = self.frequency[1]-self.frequency[0]
-                        if self.verbose:
-                            print('----------------------------------------------------\nRunning sampling routine:')
-                            self.pbar = tqdm(total=self.fitbg['mc_iter'])
-                            self.pbar.update(1)
-                else:
-                    if self.verbose:
-                        self.pbar.update(1)
-                self.i += 1
-                if self.i == self.fitbg['mc_iter'] and self.fitbg['mc_iter'] > 1:
-                    self.pbar.close()
-        # Save results of second module
-        utils.save_fitbg(self)
-        if self.fitbg['mc_iter'] > 1:
-            # Plot results if sampling
-            plots.plot_samples(self)
-        if self.verbose:
-            # Print results
-            utils.verbose_output(self)
-
-
     def fit_background(self):
         """
-        Fits for stellar background contribution due to granulation. 
+        Fits the stellar background contribution due to granulation. 
 
         Returns
         -------
@@ -342,7 +286,7 @@ class Target:
         in the power spectrum near the nyquist frequency.
 
         Returns
-        ----------
+        -------
         None
 
         """
@@ -368,15 +312,15 @@ class Target:
         Estimates amplitude of red noise components by using a smoothed version of the power
         spectrum with the power excess region masked out. This will take the mean of a specified 
         number of points (via -nrms, default=20) for each Harvey-like component.
-        
-        Returns
-        ----------
-        None
 
         Parameters
         ----------
         a : List[float]
             initial guesses for the amplitudes of all Harvey components
+
+        Returns
[…]
