+++ pysyd/target.py
@@ -83,7 +83,7 @@ class Target:
         frequency-resolved, collapsed autocorrelation function (ACF).
 
         Returns
-        ----------
+        -------
         None
 
         """
@@ -144,7 +144,7 @@ class Target:
             maximum SNR corresponding to power excess. Default value is `100.0`.
 
         Returns
-        ----------
+        -------
         None
 
         """
@@ -202,44 +202,42 @@ class Target:
         the large frequency separation (dnu) and oscillation amplitude.
 
         Returns
-        ----------
+        -------
         None
 
         """
-        self.fitbg['results'][self.name] = {'numax_smooth':[],'A_smooth':[],'numax_gauss':[],'A_gauss':[],
-                                            'FWHM':[],'dnu':[],'white':[]}
+        while self.i < self.fitbg['mc_iter']:
+            # Requires convergence of background fit before going to the next step 
+            if self.fit_background():
+                self.get_numax()
+                self.get_dnu()
+                # First step?
+                if self.i == 0:
+                    # Plot results
+                    plots.plot_background(self)
+                    if self.fitbg['mc_iter'] > 1:
+                        # Switch to critically-sampled PS if sampling
+                        mask = np.ma.getmask(np.ma.masked_inside(self.freq_cs, self.params[self.name]['bg_mask'][0], self.params[self.name]['bg_mask'][1]))
+                        self.frequency, self.power = np.copy(self.freq_cs[mask]), np.copy(self.pow_cs[mask])
+                        self.resolution = self.frequency[1]-self.frequency[0]
+                        if self.verbose:
+                            print('----------------------------------------------------\nRunning sampling routine:')
+                            self.pbar = tqdm(total=self.fitbg['mc_iter'])
+                            self.pbar.update(1)
+                else:
+                    if self.verbose:
+                        self.pbar.update(1)
+                self.i += 1
+                if self.i == self.fitbg['mc_iter'] and self.fitbg['mc_iter'] > 1:
+                    self.pbar.close()
+        # Save results of second module
+        utils.save_fitbg(self)
+        if self.fitbg['mc_iter'] > 1:
+            # Plot results if sampling
+            plots.plot_samples(self)
         if self.verbose:
-            print('----------------------------------------------------')
-            print('Running fit_background module:')
-            print('PS binned to %d data points' % len(self.bin_freq))
-
-        # Run first iteration (which has different steps than any other n>1 runs)
-        good = self.first_step()
-        if not good:
-            pass
-        else:
-            # If sampling is enabled (i.e., args.mciter > 1), a progress bar is created w/ verbose output
-            if self.fitbg['mc_iter'] > 1:
-                if self.verbose:
-                    print('----------------------------------------------------')
-                    print('Running sampling routine:')
-                    self.pbar = tqdm(total=self.fitbg['mc_iter'])
-                    self.pbar.update(1)
-                self.i = 1
-                # Continue to sample while the number of successful steps is less than args.mciter
-                while self.i < self.fitbg['mc_iter']:
-                    self.sampling_step()
-                utils.save_fitbg(self)
-                plots.plot_samples(self)
-                if self.verbose:
-                    # Print results with uncertainties
-                    utils.verbose_output(self, sampling=True)
-            # Single iteration
-            else:
-                utils.save_fitbg(self)
-                if self.verbose:
-                    # Print results without uncertainties
[…]
