### 2021-07-01 5658ee7
Create cli.rst
