+++ docs/cli.rst
@@ -1,0 +1,262 @@
+.. _cli:
+
+Command Line Interface
+************************
+
+.. _parentparse:
+
+Parent Parser
+===============
+
+Higher level functionality of the software. All three modes inherent the parent parser.
+
+- ``-file``, ``--file``, ``-list``, ``--list``, ``-todo``, ``--todo``
+   * dest = ``args.file``
+   * help = Path to text file that contains the list of stars to process (convenient for running many stars).
+   * type = ``str``
+   * default = ``TODODIR``
+- ``-in``, ``--in``, ``-input``, ``--input``, ``-inpdir``, ``--inpdir``
+   * dest = ``args.inpdir``
+   * help = Path to input data
+   * type = ``str``
+   * default = ``INPDIR``
+- ``-info``, ``--info``, ``-information``, ``--information`` 
+   * dest = ``args.info``
+   * help = Path to the csv containing star information (although not required).
+   * type = ``str``
+   * default = ``INFODIR``
+- ``-out``, ``--out``, ``-output``, ``--output``, ``-outdir``, ``--outdir``
+   * dest = ``args.outdir``
+   * help = Path that results are saved to
+   * type = ``str``
+   * default = ``OUTDIR``
+- ``-verbose``, ``--verbose``
+   * dest = ``args.verbose``
+   * help = Turn on verbose output
+   * type = ``bool``
+   * default = ``False``
+   * action = ``store_true``
+
+.. _mainparse:
+
+Main Parser
+=============
+
+Accesses all science-related functions and is therefore for both ``run`` and ``parallel`` modes.
+
+- ``-bg``, ``--bg``, ``-fitbg``, ``--fitbg``, ``-background``, ``--background``
+   * dest = ``args.background``
+   * help = Turn off the background fitting procedure
+   * type = ``bool``
+   * default = ``True``
+   * action = ``store_false``
+- ``-cad``, ``--cad``, ``-cadence``, ``--cadence``
+   * dest = ``args.cadence``
+   * help = cadence of time series (used to calculate nyquist frequency), which will automatically be calculated when time series data is available. 
+   * type = ``int``
+   * default = `0`
+   * unit = seconds
+- ``-ex``, ``--ex``, ``-findex``, ``--findex``, ``-excess``, ``--excess``
+   * dest = ``args.background``
+   * help = turn off find excess module
+   * type = ``bool``
+   * default = ``True``
+   * action = ``store_false``
+- ``-kc``, ``--kc``, ``-kepcorr``, ``--kepcorr``
+   * dest = ``args.kepcorr``
+   * help = turn on the *Kepler* short-cadence artefact correction module
+   * type = ``bool``
+   * default = ``False``
+   * action = ``store_true``
+- ``-nyq``, ``--nyq``, ``-nyquist``, ``--nyquist``
+   * dest = ``args.nyquist``
+   * help = nyquist frequency of the power spectrum (relevant for when the time series is not provided) 
+   * type = ``float``
+   * default = ``None``
+   * unit = muHz
+- ``-ofa``, ``--ofa``, ``-of_actual``, ``--of_actual``
+   * dest = ``args.of_actual``
+   * help = The oversampling factor of the provided power spectrum. Default is `0`, which means it is calculated from the time series data. Note: This needs to be provided if there is no time series data!
+   * type = ``int``
+   * default = `0`
+- ``-ofn``, ``--ofn``, ``-of_new``, ``--of_new``
+   * dest = ``args.of_new``
+   * help = The new oversampling factor to use in the first iterations of both modules. Default is `5` (see performance for more details).
[…]
