### 2021-07-01 215014f
Update cli.rst
