+++ docs/cli.rst
@@ -5,142 +5,134 @@ Command Line Interface
 
 .. _parentparse:
 
+
 Parent Parser
 ===============
 
 Higher level functionality of the software. All three modes inherent the parent parser.
 
-- ``-file``, ``--file``, ``-list``, ``--list``, ``-todo``, ``--todo``
+- ``--file``, ``--list``, ``--todo``
    * dest = ``args.file``
    * help = Path to text file that contains the list of stars to process (convenient for running many stars).
    * type = ``str``
    * default = ``TODODIR``
-- ``-in``, ``--in``, ``-input``, ``--input``, ``-inpdir``, ``--inpdir``
+- ``--in``, ``--input``, ``--inpdir``
    * dest = ``args.inpdir``
    * help = Path to input data
    * type = ``str``
    * default = ``INPDIR``
-- ``-info``, ``--info``, ``-information``, ``--information`` 
+- ``--info``, ``--information`` 
    * dest = ``args.info``
    * help = Path to the csv containing star information (although not required).
    * type = ``str``
    * default = ``INFODIR``
-- ``-out``, ``--out``, ``-output``, ``--output``, ``-outdir``, ``--outdir``
+- ``--out``, ``--output``, ``--outdir``
    * dest = ``args.outdir``
    * help = Path that results are saved to
    * type = ``str``
    * default = ``OUTDIR``
-- ``-verbose``, ``--verbose``
+- ``--verbose``, ``-v``
    * dest = ``args.verbose``
    * help = Turn on verbose output
    * type = ``bool``
    * default = ``False``
    * action = ``store_true``
 
+
 .. _mainparse:
+
 
 Main Parser
 =============
 
 Accesses all science-related functions and is therefore for both ``run`` and ``parallel`` modes.
 
-- ``-bg``, ``--bg``, ``-fitbg``, ``--fitbg``, ``-background``, ``--background``
+-  ``--bg``, ``--fitbg``, ``--background``, ``-b``
    * dest = ``args.background``
    * help = Turn off the background fitting procedure
    * type = ``bool``
    * default = ``True``
    * action = ``store_false``
-- ``-cad``, ``--cad``, ``-cadence``, ``--cadence``
-   * dest = ``args.cadence``
-   * help = cadence of time series (used to calculate nyquist frequency), which will automatically be calculated when time series data is available. 
-   * type = ``int``
-   * default = `0`
-   * unit = seconds
-- ``-ex``, ``--ex``, ``-findex``, ``--findex``, ``-excess``, ``--excess``
+- ``--ex``, ``--findex``, ``--excess``, ``-x``
    * dest = ``args.background``
    * help = turn off find excess module
    * type = ``bool``
    * default = ``True``
    * action = ``store_false``
-- ``-kc``, ``--kc``, ``-kepcorr``, ``--kepcorr``
+- ``--kc``, ``--kepcorr``, ``-k``
    * dest = ``args.kepcorr``
    * help = turn on the *Kepler* short-cadence artefact correction module
    * type = ``bool``
    * default = ``False``
    * action = ``store_true``
-- ``-nyq``, ``--nyq``, ``-nyquist``, ``--nyquist``
-   * dest = ``args.nyquist``
-   * help = nyquist frequency of the power spectrum (relevant for when the time series is not provided) 
-   * type = ``float``
-   * default = ``None``
-   * unit = muHz
-- ``-ofa``, ``--ofa``, ``-of_actual``, ``--of_actual``
+- ``--ofa``, ``--of_actual``
    * dest = ``args.of_actual``
[…]
