### 2021-07-01 fd98f7d
Update cli.rst
