+++ docs/overview.rst
@@ -1,18 +1,120 @@
 .. _overview:
 
-Overview
-########
 
-A ``pySYD`` pipeline ``Target`` class object has two main methods:
+Initialization of ``pySYD`` via command line will use the following paths:
 
-#. The find excess routine (via ``Target.find_excess``) :
+- ``TODODIR`` : '~/path_to_put_pysyd_stuff/info/todo.txt'
+- ``INFODIR`` : '~/path_to_put_pysyd_stuff/info/star_info.csv'
+- ``INPDIR`` : '~/path_to_put_pysyd_stuff/data'
+- ``OUTDIR`` : '~/path_to_put_pysyd_stuff/results'
+
+which by default, is the absolute path of the current working directory (or however you choose to set it up). All of these paths should be ready to go
+if you followed the suggestions in :ref:`structure` or used our ``setup`` feature.
+
+
+Modes
+******
+
+There are currently four main ``pySYD`` modes: 
+
+#. ``setup`` : Initializes ``pysyd.pipeline.setup`` for quick and easy setup of directories, files and examples. This mode only
+   inherits higher level functionality and has limited CLI (see :ref:`parent parser<parentparse>` below).
+
+#. ``load`` : Initializes ``pysyd.pipeline.load`` to load in data for a single target. Because this does handle data, this has 
+   full access to both the :ref:`parent<parentparse>` and :ref:`main parser<mainparse>`.
+
+#. ``run`` : The main pySYD pipeline function is initialized through ``pysyd.pipeline.run`` and runs the two core modules 
+   (i.e. ``find_excess`` and ``fit_background``) for each star consecutively. This mode operates using most CLI options, inheriting
+   both the :ref:`parent<parentparse>` and :ref:`main parser<mainparse>` options.
+
+#. ``parallel`` : Operates the same way as the previous mode, but processes stars simultaneously in parallel. Based on the number of threads
+   available, stars are separated into groups (where the number of groups is exactly equal to the number of threads). This mode uses all CLI
+   options, including the number of threads to use for parallelization (:ref:`see here<parallel>`).
+
+=========================
+
+
+Recommended Structure
+***********************
+
+We recommend using the following structure under three main directories, which is discussed 
+in more detail below:
+
+#. **info/** : [optional input] directory to provide prior information on the processed stars
+#. **data/** : [required input] directory containing the light curves and power spectra
+#. **results/** : [optional output] directory for result figures and files
+
+Input
+********
+
+Info/
+++++++++
+
+There are two main files provided:
+
+#. **info/todo.txt** : text file with one star ID per line, which must match the data ID. If no stars are specified via command line, the star(s) listed in the text file will be processed by default. This is recommended when running a large number of stars.
+
+#. **info/star_info.csv** : contains individual star information. Star IDs are crossmatched with this list and therefore, do not need to be in any particular order. In order to read information in properly, it **must** contain the following columns:
+
+   * "stars" [int] : star IDs that should exactly match the star provided via command line or in todo.txt
+   * "radius" [float] : stellar radius (units: solar radii)
+   * "teff" [float] : effective temperature (units: K)
+   * "logg" [float] : surface gravity (units: dex)
+   * "numax" [float] : the frequency corresponding to maximum power (units: muHz)
+   * "lower_ex" [float] : lower frequency limit to use in the find_excess module (units: muHz)
+   * "upper_ex" [float] : upper frequency limit to use in the find_excess module (units: muHz)
+   * "lower_bg" [float] : lower frequency limit to use in the fit_background module (units: muHz)
+   * "upper_bg" [float] : upper frequency limit to use in the fit_background module (units: muHz)
[…]
