### 2021-07-01 cf5c5a9
Update cli.rst
