+++ docs/advanced.rst
@@ -3,293 +3,3 @@
 Advanced Usage
 #################
 
-Initialization of ``pySYD`` via command line will use the following paths:
-
-- ``TODODIR`` : '~/path_to_put_pysyd_stuff/info/todo.txt'
-- ``INFODIR`` : '~/path_to_put_pysyd_stuff/info/star_info.csv'
-- ``INPDIR`` : '~/path_to_put_pysyd_stuff/data'
-- ``OUTDIR`` : '~/path_to_put_pysyd_stuff/results'
-
-which by default, is the absolute path of the current working directory (or however you choose to set it up). All of these paths should be ready to go
-if you followed the suggestions in :ref:`structure` or used our ``setup`` feature.
-
-Modes
-******
-
-There are currently three modes that ``pySYD`` can operate in via command line: 
-
-#. ``setup`` : Initializes ``pysyd.pipeline.setup`` for quick and easy setup of directories, files and examples. This mode only
-   inherits higher level functionality and has limited CLI (see :ref:`parent parser<parentparse>` below).
-
-#. ``run`` : The main pySYD pipeline function is initialized through ``pysyd.pipeline.run`` and runs the two core modules 
-   (i.e. ``find_excess`` and ``fit_background``) for each star consecutively. This mode operates using most CLI options, inheriting
-   both the :ref:`parent<parentparse>` and :ref:`main parser<mainparse>` options.
-
-#. ``parallel`` : Operates the same way as the previous mode, but processes stars simultaneously in parallel. Based on the number of threads
-   available, stars are separated into groups (where the number of groups is exactly equal to the number of threads). This mode uses all CLI
-   options, including the number of threads to use for parallelization (:ref:`see here<parallel>`).
-
-=======
-
-.. _cli:
-
-Command Line Interface
-************************
-
-.. _parentparse:
-
-Parent Parser
-===============
-
-Higher level functionality of the software. All three modes inherent the parent parser.
-
-- ``-file``, ``--file``, ``-list``, ``--list``, ``-todo``, ``--todo``
-   * dest = ``args.file``
-   * help = Path to text file that contains the list of stars to process (convenient for running many stars).
-   * type = ``str``
-   * default = ``TODODIR``
-- ``-in``, ``--in``, ``-input``, ``--input``, ``-inpdir``, ``--inpdir``
-   * dest = ``args.inpdir``
-   * help = Path to input data
-   * type = ``str``
-   * default = ``INPDIR``
-- ``-info``, ``--info``, ``-information``, ``--information`` 
-   * dest = ``args.info``
-   * help = Path to the csv containing star information (although not required).
-   * type = ``str``
-   * default = ``INFODIR``
-- ``-out``, ``--out``, ``-output``, ``--output``, ``-outdir``, ``--outdir``
-   * dest = ``args.outdir``
-   * help = Path that results are saved to
-   * type = ``str``
-   * default = ``OUTDIR``
-- ``-verbose``, ``--verbose``
-   * dest = ``args.verbose``
-   * help = Turn on verbose output
-   * type = ``bool``
-   * default = ``False``
-   * action = ``store_true``
-
-.. _mainparse:
-
-Main Parser
-=============
-
-Accesses all science-related functions and is therefore for both ``run`` and ``parallel`` modes.
-
-- ``-bg``, ``--bg``, ``-fitbg``, ``--fitbg``, ``-background``, ``--background``
-   * dest = ``args.background``
-   * help = Turn off the background fitting procedure
-   * type = ``bool``
-   * default = ``True``
[…]
