### 2021-07-01 e5646a0
Update cli.rst
