+++ docs/cli.rst
@@ -3,8 +3,8 @@
 Command Line Interface
 ************************
 
-In order to maximize the performance of the software, we have included many optional commands to help identify the
-best possible asteroseismic parameters in even the lowest SNR cases. Examples of these optional features are shown 
+In order to maximize on the performance of the software, we have included many optional commands to help identify the
+best possible asteroseismic properties in even the lowest SNR cases. Examples of these optional features are shown 
 in advanced usage (ref).
 
 .. _parentparse:
