### 2021-07-01 #3331 @ashleychontos
@mbobra apologies in advance since this is all of our (the authors) first time preparing anything like this. First and foremost, what is the most appropriate way to address and comment on either/both of the reviews (similar to a resubmission with AAS)? Following are comments from the reviews that we could use some clarification on before proceeding.

1. Re: the unit test. The scope of unit tests is more clear for say, a squared function, that can    easily assert that the result is indeed the squared value of the input.  However, it is not as clear to us on how we could implement this for a more complex software package like pySYD. @benjaminpope brought up a fair point about the documentation examples not matching the output, which was true and actually the result of an older version of both the software and documentation. We (the authors) have discussed the possible implementation of such test suite but were first curious that if the examples in the documentation matched the output verbatim, if this would be sufficient?
2. Re: the JOSS community guidelines. On our documentation homepage, it provides both an invitation for code contributions as well as the link to report any issues with the software. Should this content be a page in its own or highlighted in some other way that meets the JOSS requirements?

If the above test is not sufficient, we have discussed two other alternatives and would like to hear any comments or suggestions (@danhey included). The first is providing enough examples that cover all optional arguments to make sure no error is produced when they are all called (additionally, a plot could be produced to show the before and after when those arguments are used). The second idea was to have a counter that runs through all functions and if successful, asserts that the number returned equals the number of functions in the pipeline. This would also account for optional functions that are not required for the software to run successfully.
