### 2021-07-01 #3331 @benjaminpope
@ashleychontos thanks! So as a gold standard example, @dfm did this for the `exoplanet` review: https://github.com/openjournals/joss-reviews/issues/3285#issuecomment-859043010, in which he links to some GitHub Issues that resolve each of the reviews as a checklist. So you might like to create a checklist, in which you respond point by point to each of the reviews, or you might like to create a separate issue for each point, or something like that.

If the examples in the documentation match the output verbatim that will be good! It will also be good to have richer output - for example, figures (and autosave figures, rather than click-through) *and* terminal output, so that we can test that it reproduces the examples exactly. (And a GitHub Action to test one or more examples, even with just a checksum, would meet the continuous testing requirement).
