### 2021-07-01 #3331 @danhey
Agree with @benjaminpope. I also think a good unit test for GitHub Actions would be to just simply run pySYD on a test star for which numax and dnu is well known, and assert that the output of the pipeline is `np.isclose()` to what you expect. This means that if you make some changes to the code, you will instantly see if the new output is significantly different, and the test will fail.
