### 2021-07-02 #3331 @mbobra
> First and foremost, what is the most appropriate way to address and comment on either/both of the reviews (similar to a resubmission with AAS)?

👋 @ashleychontos It looks like you already got your answer -- but in any case, I will provide a rather vague response that as long as the reviewers feel (1) you have addressed their concerns and, (2) the software adheres to the JOSS guideline, I will accept the submission!
