### 2021-07-02 45f1729
Updates for JOSS review comments
