+++ pysyd/target.py
@@ -1,5 +1,7 @@
+import pickle
 import numpy as np
 from tqdm import tqdm
+import matplotlib.pyplot as plt
 from scipy.signal import find_peaks
 from scipy.optimize import curve_fit
 from scipy.interpolate import InterpolatedUnivariateSpline
@@ -13,38 +15,40 @@ from pysyd import plots
 
 
 class Target:
-    """
-    A pySYD pipeline target. Initialization stores all the relevant information and
-    checks/loads in data for the given target. pySYD no longer requires BOTH the time
-    series data and the power spectrum, but requires additional information via CLI if
-    the former is not provided i.e. cadence or nyquist frequency, the oversampling
-    factor (if relevant), etc.
-
-    Attributes
-    ----------
-    star : int
-        the star ID
-    params : Dict[str,object]
-        the pipeline parameters
-    findex : Dict[str,object]
-        the parameters of the find excess routine
-    fitbg : Dict[str,object]
-        the parameters of the fit background routine
-    verbose : bool
-        if true, turns on the verbose output
-
-    Parameters
-    ----------
-    args : argparse.Namespace
-        the parsed and updated command line arguments
-
-    Methods
-    -------
-    TODO: Add methods
-
-    """
 
     def __init__(self, star, args):
+        """
+        A pySYD pipeline target. Initialization stores all the relevant information and
+        checks/loads in data for the given target. pySYD no longer requires BOTH the time
+        series data and the power spectrum, but requires additional information via CLI if
+        the former is not provided i.e. cadence or nyquist frequency, the oversampling
+        factor (if relevant), etc.
+    
+        Attributes
+        ----------
+        star : int
+            the star ID
+        params : Dict[str,object]
+            the pipeline parameters
+        findex : Dict[str,object]
+            the parameters of the find excess routine
+        fitbg : Dict[str,object]
+            the parameters relevant for the background-fitting procedure
+        globe : Dict[str,object]
+            parameters relevant for estimating global asteroseismic parameters numax and dnu
+        verbose : bool
+            if true, turns on the verbose output
+    
+        Parameters
+        ----------
+        args : argparse.Namespace
+            the parsed and updated command line arguments
+
+        Methods
+        -------
+        TODO: Add methods
+
+        """
         self.name = star
         self.params = args.params
         self.findex = args.findex
@@ -75,6 +79,18 @@ class Target:
             if utils.check_fitbg(self):
                 self = utils.get_fitbg(self)
                 self.fit_global()
+#        if self.params['show'] and self.pickles != []:
+#            for p, pkl in enumerate(self.pickles):
+#                with open(pkl,'rb') as file:
+#                    fig = pickle.load(file)
+#                plt.show(block=False)
+        if self.params['show']:
+            if self.verbose:
+                print('Displaying figures...')
+            plt.show(block=False)
+            if self.verbose:
+                input('------------ please press RETURN to exit -------------')
+                input('------------------------------------------------------')
 
 
     def find_excess(self):
@@ -91,7 +107,7 @@ class Target:
         if self.findex['binning'] is not None:
             self.bin_freq, self.bin_pow, _ = functions.bin_data(self.freq, self.pow, width=self.findex['binning'], log=True, mode=self.findex['mode'])
             if self.verbose:
-                print('----------------------------------------------------')
+                print('------------------------------------------------------')
                 print('Running find_excess module:')
                 print('PS binned to %d datapoints' % len(self.bin_freq))
                 print('Binned freq res: %.2f muHz'%(self.bin_freq[1]-self.bin_freq[0]))
@@ -120,6 +136,7 @@ class Target:
                 print('selecting model %d'%self.findex['results'][self.name]['best'])
[…]
