### 2021-07-03 1949c59
Adding Borucki+2011 and Ricker+2015 citations
