+++ paper/paper.bib
@@ -83,6 +83,37 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{borucki2010,
+   author = {{Borucki}, W.~J. and {Koch}, D. and {Basri}, G. and {Batalha}, N. and 
+	{Brown}, T. and {Caldwell}, D. and {Caldwell}, J. and {Christensen-Dalsgaard}, J. and 
+	{Cochran}, W.~D. and {DeVore}, E. and {Dunham}, E.~W. and {Dupree}, A.~K. and 
+	{Gautier}, T.~N. and {Geary}, J.~C. and {Gilliland}, R. and 
+	{Gould}, A. and {Howell}, S.~B. and {Jenkins}, J.~M. and {Kondo}, Y. and 
+	{Latham}, D.~W. and {Marcy}, G.~W. and {Meibom}, S. and {Kjeldsen}, H. and 
+	{Lissauer}, J.~J. and {Monet}, D.~G. and {Morrison}, D. and 
+	{Sasselov}, D. and {Tarter}, J. and {Boss}, A. and {Brownlee}, D. and 
+	{Owen}, T. and {Buzasi}, D. and {Charbonneau}, D. and {Doyle}, L. and 
+	{Fortney}, J. and {Ford}, E.~B. and {Holman}, M.~J. and {Seager}, S. and 
+	{Steffen}, J.~H. and {Welsh}, W.~F. and {Rowe}, J. and {Anderson}, H. and 
+	{Buchhave}, L. and {Ciardi}, D. and {Walkowicz}, L. and {Sherry}, W. and 
+	{Horch}, E. and {Isaacson}, H. and {Everett}, M.~E. and {Fischer}, D. and 
+	{Torres}, G. and {Johnson}, J.~A. and {Endl}, M. and {MacQueen}, P. and 
+	{Bryson}, S.~T. and {Dotson}, J. and {Haas}, M. and {Kolodziejczak}, J. and 
+	{Van Cleve}, J. and {Chandrasekaran}, H. and {Twicken}, J.~D. and 
+	{Quintana}, E.~V. and {Clarke}, B.~D. and {Allen}, C. and {Li}, J. and 
+	{Wu}, H. and {Tenenbaum}, P. and {Verner}, E. and {Bruhweiler}, F. and 
+	{Barnes}, J. and {Prsa}, A.},
+    title = "{Kepler Planet-Detection Mission: Introduction and First Results}",
+  journal = {Science},
+     year = 2010,
+    month = feb,
+   volume = 327,
+    pages = {977},
+      doi = {10.1126/science.1185402},
+   adsurl = {http://adsabs.harvard.edu/abs/2010Sci...327..977B},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @article{chaplin2014,
    author = {{Chaplin}, W.~J. and {Basu}, S. and {Huber}, D. and {Serenelli}, A. and 
 	{Casagrande}, L. and {Silva Aguirre}, V. and {Ball}, W.~H. and 
@@ -332,6 +363,36 @@ archivePrefix = {arXiv},
  doi           = {10.1038/s41586-020-2649-2},
  publisher     = {Springer Science and Business Media {LLC}},
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+@ARTICLE{ricker2015,
+   author = {{Ricker}, G.~R. and {Winn}, J.~N. and {Vanderspek}, R. and {Latham}, D.~W. and 
[…]
