### 2021-07-03 c6f51fc
Borucki+2010 and Ricker+2015 citations
