### 2021-07-05 3d15609
bug fixes for batch runs from todo.txt
