### 2021-07-28 cbebc88
Create CITATION.cff
Using TESS SciCon II poster DOI for now until JOSS DOI assigned
