### 2021-07-28 9fcef3a
Update performance.rst
