### 2021-07-28 f4ed88e
Update performance.rst
