### 2021-07-28 ebe2c32
Update performance.rst
