+++ docs/performance.rst
@@ -13,6 +13,7 @@ The resulting values are compared for the two methods below for the frequency of
 .. image:: figures/new.png
 
 There residuals show no strong systematics to within <0.5% in Dnu and <~1% in numax, which is smaller than the typical 
-random uncertainties. This confirms that the open-source python implementation of the SYD pipeline presented here provides consistent results with the legacy IDL version that has been used in the literature.
+random uncertainties. This confirms that the open-source python package pySYD provides consistent results with the legacy 
+IDL version that has been used extensively in the literature.
 
 *** NOTE **** Add tutorial or jupyter notebook to reproduce this figure.
