### 2021-07-29 a814208
Fixes bug that produces an IndexError when reducing the number of Harvey-like components used in the background model
