+++ pysyd/utils.py
@@ -198,7 +198,7 @@ def get_groups(args, parallel=False):
 
 
 def get_excess_params(args, CLI, n_trials=3, step=0.25, binning=0.005, smooth_width=50.0, 
-                      lower_ex=10.0, upper_ex=4000.0, mode='mean'):
+                      mode='mean'):
     """
     Get the parameters for the find excess routine.
 
@@ -206,10 +206,6 @@ def get_excess_params(args, CLI, n_trials=3, step=0.25, binning=0.005, smooth_wi
     ----------
     args : argparse.Namespace
         the command line arguments
-    lower_ex : float, optional
-        the lower frequency bound (in muHz). Default value is `10.0` muHz.
-    upper_ex : float, optional
-        the upper frequency bound (in muHz). Default value is `4000.0` muHz.
     step : float, optional
         TODO: Write description. Default value is `0.25`.
     binning : float, optional
@@ -229,8 +225,6 @@ def get_excess_params(args, CLI, n_trials=3, step=0.25, binning=0.005, smooth_wi
     """
     if CLI:
         findex = {
-            'lower_ex': lower_ex,
-            'upper_ex': upper_ex,
             'step': args.step,
             'binning': args.binning,
             'mode': args.mode,
@@ -240,8 +234,6 @@ def get_excess_params(args, CLI, n_trials=3, step=0.25, binning=0.005, smooth_wi
         }
     else:
         findex = {
-            'lower_ex': lower_ex,
-            'upper_ex': upper_ex,
             'step': step,
             'binning': binning,
             'mode': mode,
@@ -254,9 +246,8 @@ def get_excess_params(args, CLI, n_trials=3, step=0.25, binning=0.005, smooth_wi
     return args
 
 
-def get_background_params(args, CLI, lower_bg=10.0, upper_bg=4000.0, ind_width=20.0,   
-                          box_filter=1.0, n_rms=20, mc_iter=1, samples=False, n_laws=None,
-                          use='bic',):
+def get_background_params(args, CLI, ind_width=20.0, box_filter=1.0, n_rms=20, mc_iter=1, 
+                          samples=False, n_laws=None, ab=False, use='bic',):
     """
     Get the parameters for the background-fitting routine.
 
@@ -264,10 +255,6 @@ def get_background_params(args, CLI, lower_bg=10.0, upper_bg=4000.0, ind_width=2
     ----------
     args : argparse.Namespace
         the command line arguments
-    lower_bg : float, optional
-        the lower frequency bound (in muHz). Default value is `10.0` muHz.
-    upper_bg : float, optional
-        the upper frequency bound (in muHz). Default value is `4000.0` muHz.
     box_filter : float
         the size of the 1D box smoothing filter (in muHz). Default value is `1.0`.
     ind_width : float
@@ -276,6 +263,8 @@ def get_background_params(args, CLI, lower_bg=10.0, upper_bg=4000.0, ind_width=2
         number of data points to estimate red noise contributions. Default value is `20`.
     use : str
         which metric to use (i.e. bic or aic) for model selection. Default is `'bic'`.
+    ab : bool
+        use {a,b} parametrization for Harvey models. Default is `False`.
     n_laws : int
         force number of Harvey-like components in background fit. Default value is `None`.
     mc_iter : int
@@ -313,12 +302,11 @@ def get_background_params(args, CLI, lower_bg=10.0, upper_bg=4000.0, ind_width=2
     """
     if CLI:
         fitbg = {
-            'lower_bg': lower_bg,
-            'upper_bg': upper_bg,
             'ind_width': args.ind_width,
             'box_filter': args.box_filter,
[…]
