### 2021-07-30 3b1ee13
Fixed binning issue in find excess and some other minor things (including a version update)
