+++ docs/advanced.rst
@@ -1,0 +1,99 @@
+.. _advanced:
+.. code-block:: bash
+
+Advanced Usage
+#################
+
+=======
+
+Below are examples of how to use specific ``pySYD`` features, as well as plots showing results before and after their usage.
+
+``--dnu``
+++++++++
+.. tabularcolumns:: |c|c||
+
++-------------------------------------------------+---------------------------------------------------------+
+| Before                                          | After                                                   |
++=================================================+=========================================================+
+| :bash:`pysyd run --star 9512063 --numax 843`    | :bash:`pysyd run --star 9512063 --numax 843 --dnu 49.54`|
++-------------------------------------------------+---------------------------------------------------------+
+| .. figure:: figures_advanced/9512063_before.png | .. figure:: figures_advanced/9512063_after.png          |
+|    :scale: 50 %                                 |    :scale: 50 %                                         |
++-------------------------------------------------+---------------------------------------------------------+
+
+``--ew``
+++++++++
+.. tabularcolumns:: |c|c||
+
++-------------------------------------------------+---------------------------------------------------------+
+| Before                                          | After                                                   |
++=================================================+=========================================================+
+| :bash:`pysyd run --star 9542776 --numax 900 --dnu 39.62`    | :bash:`pysyd run --star 9542776 --numax 900 --dnu 39.62 --ew 1.5`|
++-------------------------------------------------+---------------------------------------------------------+
+| .. figure:: figures_advanced/9542776_before.png | .. figure:: figures_advanced/9542776_after.png          |
+|    :scale: 50 %                                 |    :scale: 50 %                                         |
++-------------------------------------------------+---------------------------------------------------------+
+
+``--kc``
+++++++++
+.. tabularcolumns:: |c|c||
++-------------------------------------------------+-------------------------------------------------------+
+| Before                                          | After                                                 |
++=================================================+=======================================================+
+| :bash:`pysyd run --star 11769801`    | :bash:`pysyd run --star 11769801 --ux 3500`     |
++-------------------------------------------------+-------------------------------------------------------+
+| .. figure:: figures_advanced/8045442_before.png | .. figure:: figures_advanced/8045442_after.png        |
+|    :scale: 50 %                                 |    :scale: 50 %                                       |
++-------------------------------------------------+-------------------------------------------------------+
+
+``--lp``
+++++++++
+.. tabularcolumns:: |c|c||
+
++--------------------------------------------------------+-----------------------------------------------------------------+
+| Before                                                 | After                                                           |
++========================================================+=================================================================+
+| :bash:`pysyd run --star 10731424 --numax 750 --dnu 42` | :bash:`pysyd run --star 10731424 --numax 750 --dnu 42 --lp 480` |
++-------------------------------------------------+------------------------------------------------------------------------+
+| .. figure:: figures_advanced/10731424_after.png        | .. figure:: figures_advanced/10731424_after.png                 |
+|    :scale: 50 %                                        |    :scale: 50 %                                                 |
++--------------------------------------------------------+-----------------------------------------------------------------+
+
+``--npeaks``
+++++++++
+.. tabularcolumns:: |c|c||
+
++--------------------------------------------------------------------+-------------------------------------------------------------------------------+
+| Before                                                             | After                                                                         |
++====================================================================+===============================================================================+
+| :bash:`pysyd run --star 9455860 --numax 1487 --ew 1.2  --dnu 70.9` | :bash:`pysyd run --star 9455860 --numax 1487 --ew 1.2 --npeaks 20 --dnu 70.9` |
++--------------------------------------------------------------------+-------------------------------------------------------------------------------+
+| .. figure:: figures_advanced/9455860_after.png                     | .. figure:: figures_advanced/9455860_after.png                                |
[…]
