+++ docs/advanced.rst
@@ -1,7 +1,7 @@
 .. _advanced:
-.. role:: bash
+.. role:: bash(code)
    :language: bash
-   
+
 Advanced Usage
 #################
 
@@ -9,40 +9,45 @@ Advanced Usage
 
 Below are examples of how to use specific ``pySYD`` features, as well as plots showing results before and after their usage.
 
+- `kc`_ 
+- `npeaks`_ 
+
 ``--dnu``
 ++++++++
 
-+-------------------------------------------------+---------------------------------------------------------+
-| Before                                          | After                                                   |
-+=================================================+=========================================================+
-| :bash:`pysyd run --star 9512063 --numax 843`    | :bash:`pysyd run --star 9512063 --numax 843 --dnu 49.54`|
-+-------------------------------------------------+---------------------------------------------------------+
-| .. figure:: figures_advanced/9512063_before.png | .. figure:: figures_advanced/9512063_after.png          |
-|    :scale: 50 %                                 |    :scale: 50 %                                         |
-+-------------------------------------------------+---------------------------------------------------------+
++------------------------------------------------+---------------------------------------------------------+
+| Before                                         | After                                                   |
++================================================+=========================================================+
+|:bash:`pysyd run --star 9512063 --numax 843`    |:bash:`pysyd run --star 9512063 --numax 843 --dnu 49.54` |
++-----------------------------------------------------+----------------------------------------------------+
+| .. figure:: figures_advanced/9512063_before.png| .. figure:: figures_advanced/9512063_after.png          |
+|    :scale: 50 %                                |    :scale: 50 %                                         |
++------------------------------------------------+---------------------------------------------------------+
 
 ``--ew``
 ++++++++
 
-+-------------------------------------------------+---------------------------------------------------------+
-| Before                                          | After                                                   |
-+=================================================+=========================================================+
-| :bash:`pysyd run --star 9542776 --numax 900 --dnu 39.62`    | :bash:`pysyd run --star 9542776 --numax 900 --dnu 39.62 --ew 1.5`|
-+-------------------------------------------------+---------------------------------------------------------+
-| .. figure:: figures_advanced/9542776_before.png | .. figure:: figures_advanced/9542776_after.png          |
-|    :scale: 50 %                                 |    :scale: 50 %                                         |
-+-------------------------------------------------+---------------------------------------------------------+
++----------------------------------------------------------+------------------------------------------------------------------+
+| Before                                                   | After                                                            |
++==========================================================+==================================================================+
+| :bash:`pysyd run --star 9542776 --numax 900 --dnu 39.62` | :bash:`pysyd run --star 9542776 --numax 900 --dnu 39.62 --ew 1.5`|
++----------------------------------------------------------+------------------------------------------------------------------+
+| .. figure:: figures_advanced/9542776_before.png          | .. figure:: figures_advanced/9542776_after.png                   |
+|    :scale: 50 %                                          |    :scale: 50 %                                                  |
++----------------------------------------------------------+------------------------------------------------------------------+
 
 ``--kc``
 ++++++++
 +-------------------------------------------------+-------------------------------------------------------+
 | Before                                          | After                                                 |
 +=================================================+=======================================================+
-| :bash:`pysyd run --star 11769801`    | :bash:`pysyd run --star 11769801 --ux 3500`     |
+| :bash:`pysyd run --star 11769801`  			  | :bash:`pysyd run --star 11769801 --ux 3500`           |
 +-------------------------------------------------+-------------------------------------------------------+
 | .. figure:: figures_advanced/8045442_before.png | .. figure:: figures_advanced/8045442_after.png        |
 |    :scale: 50 %                                 |    :scale: 50 %                                       |
[…]
