+++ docs/advanced.rst
@@ -9,20 +9,20 @@ Advanced Usage
 
 Below are examples of how to use specific ``pySYD`` features, as well as plots showing results before and after their usage.
 
-- `kc`_ 
-- `npeaks`_ 
+- `kc`
+- `npeaks`
 
 ``--dnu``
 ++++++++
 
-+------------------------------------------------+---------------------------------------------------------+
-| Before                                         | After                                                   |
-+================================================+=========================================================+
-|:bash:`pysyd run --star 9512063 --numax 843`    |:bash:`pysyd run --star 9512063 --numax 843 --dnu 49.54` |
-+-----------------------------------------------------+----------------------------------------------------+
-| .. figure:: figures_advanced/9512063_before.png| .. figure:: figures_advanced/9512063_after.png          |
-|    :scale: 50 %                                |    :scale: 50 %                                         |
-+------------------------------------------------+---------------------------------------------------------+
++-------------------------------------------------+---------------------------------------------------------+
+| Before                                          | After                                                   |
++=================================================+=========================================================+
+|:bash:`pysyd run --star 9512063 --numax 843`     |:bash:`pysyd run --star 9512063 --numax 843 --dnu 49.54` |
++-------------------------------------------------+---------------------------------------------------------+
+| .. figure:: figures_advanced/9512063_before.png | .. figure:: figures_advanced/9512063_after.png          |
+|    :scale: 50 %                                 |    :scale: 50 %                                         |
++-------------------------------------------------+---------------------------------------------------------+
 
 ``--ew``
 ++++++++
@@ -47,7 +47,7 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 |    :scale: 50 %                                 |    :scale: 50 %                                       |
 +-------------------------------------------------+-------------------------------------------------------+
 | useful when an artefact is present in *Kepler* data that prevents ``pysyd`` in finding numax            |
-+--------------------------------------------------+------------------------------------------------------+
++-------------------------------------------------+-------------------------------------------------------+
 
 ``--lp``
 ++++++++
@@ -56,7 +56,7 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 | Before                                                 | After                                                           |
 +========================================================+=================================================================+
 | :bash:`pysyd run --star 10731424 --numax 750 --dnu 42` | :bash:`pysyd run --star 10731424 --numax 750 --dnu 42 --lp 480` |
-+-------------------------------------------------+------------------------------------------------------------------------+
++--------------------------------------------------------+-----------------------------------------------------------------+
 | .. figure:: figures_advanced/10731424_after.png        | .. figure:: figures_advanced/10731424_after.png                 |
 |    :scale: 50 %                                        |    :scale: 50 %                                                 |
 +--------------------------------------------------------+-----------------------------------------------------------------+
@@ -73,7 +73,7 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 |    :scale: 50 %                                                    |    :scale: 50 %                                                               |
 +--------------------------------------------------------------------+-------------------------------------------------------------------------------+
 | useful when ``pysyd`` doesn't find the correct peak in the autocorrelation function 				    											 |
-+-------------------------------------------------+--------------------------------------------------------------------------------------------------+
++--------------------------------------------------------------------+-------------------------------------------------------------------------------+
 
 ``--numax``
 ++++++++
