+++ docs/advanced.rst
@@ -40,6 +40,7 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 
 ``--kc``
 ++++++++
+
 +-------------------------------------------------+-------------------------------------------------------+
 | Before                                          | After                                                 |
 +=================================================+=======================================================+
@@ -47,6 +48,8 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 +-------------------------------------------------+-------------------------------------------------------+
 | .. figure:: figures_advanced/8045442_before.png | .. figure:: figures_advanced/8045442_after.png        |
 |    :scale: 50 %                                 |    :scale: 50 %                                       |
++-------------------------------------------------+-------------------------------------------------------+
+| useful when an artefact is present in *Kepler* data that prevents ``pysyd`` in finding numax            |
 +-------------------------------------------------+-------------------------------------------------------+
 
 
