+++ docs/advanced.rst
@@ -57,7 +57,7 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 +========================================================+=================================================================+
 | :bash:`pysyd run --star 10731424 --numax 750 --dnu 42` | :bash:`pysyd run --star 10731424 --numax 750 --dnu 42 --lp 480` |
 +--------------------------------------------------------+-----------------------------------------------------------------+
-| .. figure:: figures_advanced/10731424_after.png        | .. figure:: figures_advanced/10731424_after.png                 |
+| .. figure:: figures_advanced/10731424_before.png       | .. figure:: figures_advanced/10731424_after.png                 |
 |    :scale: 50 %                                        |    :scale: 50 %                                                 |
 +--------------------------------------------------------+-----------------------------------------------------------------+
 
@@ -70,7 +70,7 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 +====================================================================+===============================================================================+
 | :bash:`pysyd run --star 9455860 --numax 1487 --ew 1.2  --dnu 70.9` | :bash:`pysyd run --star 9455860 --numax 1487 --ew 1.2 --npeaks 20 --dnu 70.9` |
 +--------------------------------------------------------------------+-------------------------------------------------------------------------------+
-| .. figure:: figures_advanced/9455860_after.png                     | .. figure:: figures_advanced/9455860_after.png                                |
+| .. figure:: figures_advanced/9455860_before.png                    | .. figure:: figures_advanced/9455860_after.png                                |
 |    :scale: 50 %                                                    |    :scale: 50 %                                                               |
 +--------------------------------------------------------------------+-------------------------------------------------------------------------------+
 
@@ -91,12 +91,12 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 ``--ux``
 ++++++++
 
-+-------------------------------------------------+-------------------------------------------------------+
-| Before                                          | After                                                 |
-+=================================================+=======================================================+
-| :bash:`pysyd run --star 11769801`               | :bash:`pysyd run --star 11769801 -ux 3500`            |
-+-------------------------------------------------+-------------------------------------------------------+
-| .. figure:: figures_advanced/11769801_after.png | .. figure:: figures_advanced/11769801_after.png       |
-|    :scale: 50 %                                 |    :scale: 50 %                                       |
-+-------------------------------------------------+-------------------------------------------------------+
++--------------------------------------------------+-------------------------------------------------------+
+| Before                                           | After                                                 |
++==================================================+=======================================================+
+| :bash:`pysyd run --star 11769801`                | :bash:`pysyd run --star 11769801 -ux 3500`            |
++--------------------------------------------------+-------------------------------------------------------+
+| .. figure:: figures_advanced/11769801_before.png | .. figure:: figures_advanced/11769801_after.png       |
+|    :scale: 50 %                                  |    :scale: 50 %                                       |
++--------------------------------------------------+-------------------------------------------------------+
 
