+++ docs/advanced.rst
@@ -36,6 +36,19 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 +----------------------------------------------------------+------------------------------------------------------------------+
 
 
+``--ie``
+++++++++
+
++----------------------------------------------------------+------------------------------------------------------------------+
+| Before                                                   | After                                                            |
++==========================================================+==================================================================+
+| :bash:`pysyd run 3112889 --numax 871.52 --dnu 53.2`      | :bash:`pysyd run --star 3112889 --numax 871.52 --dnu 53.2 --ie`  |
++----------------------------------------------------------+------------------------------------------------------------------+
+| .. figure:: figures_advanced/3112889_before.png          | .. figure:: figures_advanced/3112889_after.png                   |
+|    :scale: 50 %                                          |    :scale: 50 %                                                  |
++----------------------------------------------------------+------------------------------------------------------------------+
+
+
 ``--kc``
 ++++++++
 
