### 2021-07-31 39703c6
npeaks default was not documented as 5 - changed this in utils.py.
