+++ pysyd/utils.py
@@ -272,7 +272,7 @@ def get_background_params(args, CLI, ind_width=20.0, box_filter=1.0, n_rms=20, m
     samples : bool
         if true, will save the monte carlo samples to a csv. Default value is `False`.
     args.n_peaks : int
-        the number of peaks to select. Default value is `10`.
+        the number of peaks to select. Default value is `5`.
     args.force : float
         if not false (i.e. non-zero) will force dnu to be the equal to this value. 
     args.clip : bool
