+++ pysyd/functions.py
@@ -75,7 +75,7 @@ def remove_artefact(star, lcp=1.0/(29.4244*60*1e-6), lf_lower=[240.0,500.0], lf_
     # LC period in Msec -> 1/LC ~muHz
     artefact = (1.0+np.arange(14))*lcp
     # Estimate white noise
-    white = np.mean(star.power[(star.frequency >= star.nyquist-100.0)&(star.frequency <= star.nyquist-50.0)])
+    white = np.mean(star.power[(star.frequency >= max(self.frequency)-100.0)&(star.frequency <= max(self.frequency)-50.0)])
 
     np.random.seed(int(star.params[star.name]['seed']))
     # Routine 1: remove 1/LC artefacts by subtracting +/- 5 muHz given each artefact
