### 2021-07-31 e8f0d33
change advanced figures location; added description for each feature
