### 2021-07-31 e7a5269
added descriptions for each feature
