+++ docs/advanced.rst
@@ -25,12 +25,14 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 +-------------------------------------------------+---------------------------------------------------------+
 
 
-``--ew``
+``--ew: excess width``
 ++++++++
 
 +----------------------------------------------------------+------------------------------------------------------------------+
 | Before                                                   | After                                                            |
 +==========================================================+==================================================================+
+| Changed the excess width in the background corrected power spectrum used to calculate the ACF (and hence dnu).              |
++----------------------------------------------------------+------------------------------------------------------------------+
 | :bash:`pysyd run --star 9542776 --numax 900 --dnu 39.62` | :bash:`pysyd run --star 9542776 --numax 900 --dnu 39.62 --ew 1.5`|
 +----------------------------------------------------------+------------------------------------------------------------------+
 | .. figure:: figures/advanced/9542776_before.png          | .. figure:: figures/advanced/9542776_after.png                   |
@@ -38,12 +40,14 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 +----------------------------------------------------------+------------------------------------------------------------------+
 
 
-``--ie``
+``--ie: smooth echelle``
 ++++++++
 
 +----------------------------------------------------------+------------------------------------------------------------------+
 | Before                                                   | After                                                            |
 +==========================================================+==================================================================+
+| Smooth echelle diagram by turning on the interpolation, in order to distinguish the modes for low SNR cases                 |
++----------------------------------------------------------+------------------------------------------------------------------+
 | :bash:`pysyd run 3112889 --numax 871.52 --dnu 53.2`      | :bash:`pysyd run --star 3112889 --numax 871.52 --dnu 53.2 --ie`  |
 +----------------------------------------------------------+------------------------------------------------------------------+
 | .. figure:: figures/advanced/3112889_before.png          | .. figure:: figures/advanced/3112889_after.png                   |
@@ -51,12 +55,14 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 +----------------------------------------------------------+------------------------------------------------------------------+
 
 
-``--kc``
+``--kc: *Kepler* correction``
 ++++++++
 
 +-------------------------------------------------+-------------------------------------------------------+
 | Before                                          | After                                                 |
 +=================================================+=======================================================+
+| Remove *Kepler* artefacts from the power spectrum for an accurate numax estimate                        |
++-------------------------------------------------+-------------------------------------------------------+
 | :bash:`pysyd run --star 8045442 --numax 550`    | :bash:`pysyd run --star 8045442 --numax 550 --kc`     |
 +-------------------------------------------------+-------------------------------------------------------+
 | .. figure:: figures/advanced/8045442_before.png | .. figure:: figures/advanced/8045442_after.png        |
@@ -64,25 +70,29 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 +-------------------------------------------------+-------------------------------------------------------+
 
 
-``--lp``
-++++++++
-
-+--------------------------------------------------------+-----------------------------------------------------------------+
-| Before                                                 | After                                                           |
-+========================================================+=================================================================+
-| :bash:`pysyd run --star 10731424 --numax 750 --dnu 42` | :bash:`pysyd run --star 10731424 --numax 750 --dnu 42 --lp 480` |
-+--------------------------------------------------------+-----------------------------------------------------------------+
-| .. figure:: figures/advanced/10731424_before.png       | .. figure:: figures/advanced/10731424_after.png                 |
-|    :width: 600                                         |    :width: 600                                                  |
-+--------------------------------------------------------+-----------------------------------------------------------------+
-
-
-``--npeaks``
+``--lp: lower frequency (zoomed PS)``
 ++++++++
 
[…]
