### 2021-07-31 f480542
added descriptions for each feature
