### 2021-07-31 38f67e7
added descriptions for each feature
