### 2021-07-31 d935be2
added descriptions for each feature
