+++ docs/advanced.rst
@@ -103,16 +103,16 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 ``--numax``
 ++++++++
 
-+-------------------------------------------------+-------------------------------------------------------+
-| Before                                          | After                                                 |
-+=================================================+=======================================================+
-| Set the numax value if pySYD chooses the wrong excess in the power spectrum.  	                      |
-+-------------------------------------------------+-------------------------------------------------------+
-| :bash:`pysyd run --star 5791521`                | :bash:`pysyd run --star 5791521  --numax 670`         |
-+-------------------------------------------------+-------------------------------------------------------+
-| .. figure:: figures/advanced/5791521_before.png | .. figure:: figures/advanced/5791521_after.png        |
-|    :width: 600                                  |    :width: 600                                        |
-+-------------------------------------------------+-------------------------------------------------------+
++--------------------------------------------------+-------------------------------------------------------+
+| Before                                           | After                                                 |
++==================================================+=======================================================+
+| Set the numax value if pySYD chooses the wrong excess in the power spectrum.                             |
++--------------------------------------------------+-------------------------------------------------------+
+| :bash:`pysyd run --star 5791521`                 | :bash:`pysyd run --star 5791521 --numax 670`          |
++--------------------------------------------------+-------------------------------------------------------+
+| .. figure:: figures/advanced/5791521_before.png  | .. figure:: figures/advanced/5791521_after.png        |
+|    :width: 600                                   |    :width: 600                                        |
++--------------------------------------------------+-------------------------------------------------------+
 
 
 ``--ux: upper frequency (PS)``
