+++ paper/copy/paper.md
@@ -1,1 +1,127 @@
-dfa
+---
+title: "`pySYD`: Automated measurements of global asteroseismic parameters"
+tags:
+  - Python
+  - astronomy
+  - stellar astrophysics
+  - asteroseismology
+  - stellar oscillations
+  - fundamental stellar properties
+  - solar-like oscillations
+  - global asteroseismology
+authors:
+ - name: Ashley Chontos
+   orcid: 0000-0003-1125-2564
+   affiliation: "1, 2"
+ - name: Daniel Huber
+   orcid: 0000-0001-8832-4488
+   affiliation: 1
+ - name: Maryum Sayeed 
+   orcid: 0000-0001-6180-8482
+   affiliation: 1
+ - name: Pavadol Yamsiri
+   affiliation: 3
+affiliations:
+ - name: Institute for Astronomy, University of Hawai'i, 2680 Woodlawn Drive, Honolulu, HI 96822, USA
+   index: 1
+ - name: NSF Graduate Research Fellow
+   index: 2
+ - name: Sydney Institute for Astronomy, School of Physics, University of Sydney, NSW 2006, Australia
+   index: 3
+date: 6 May 2021
+bibliography: paper.bib
+---
+
+# Summary
+
+Asteroseismology, the study of stellar oscillations, is a powerful tool for studying the interiors of stars
+and determining their fundamental properties [@aerts2021]. For stars with temperatures that are similar to 
+the Sun, turbulent near-surface convection excites sound waves that propagate within the stellar cavity 
+[@bedding2014]. These waves penetrate into different depths within the star and therefore provide powerful 
+constraints on stellar interiors that would otherwise be inaccessible. Asteroseismology is well-established 
+in astronomy as the gold standard for characterizing fundamental properties like masses, radii, densities, 
+and ages for single stars, which has broad impacts on several fields in astronomy. For example, ages of stars 
+are important to reconstruct the formation history of the Milky Way (so-called galactic archeology). For 
+exoplanets that are discovered indirectly through changes in stellar observables, precise and accurate stellar 
+masses and radii are critical for learning about the planets that orbit them. 
+
+`pySYD` is a Python package for detecting solar-like oscillations and measuring global asteroseismic parameters given an input
+power spectrum. Derived parameters include seismic observables $\rm \nu_{max}$ and $\Delta\nu$, which are directly tied
+to the surface gravity and density of the star [@kjeldsen1995]. `pySYD` also performs a maximum likelihood estimation for 
+determining the stellar background contribution, characterizing different time scales and amplitudes of granulation 
+processes. 
+
+# Statement of Need
+
+The NASA space telescopes *Kepler*, K2 and TESS have recently provided very large databases of high-precision light curves 
+of stars. By detecting brightness variations due to solar-like oscillations, these light curves allow the application of 
+asteroseismology to large numbers of stars, which requires automated software tools to efficiently extract observables. 
+Several tools have been developed for asteroseismic analyses [e.g., `A2Z`, @mathur2010; `COR`, @mosser2009; `OCT`, @hekker2010; 
+`SYD`, @huber2009], but most of them are closed-source and thus inaccessible to the general astronomy community. Some open-source 
+tools exist [e.g., `DIAMONDS` and `FAMED`, @corsaro2014; `PBjam`, @nielsen2021; `lightkurve`, @lightkurve], but they are either 
+optimized for smaller samples of stars or have not yet been extensively tested against closed-source tools.
+
+`pySYD` is adapted from the framework of the IDL-based `SYD` pipeline [@huber2009; hereafter referred to as `IDL-SYD`], which has 
+been used frequently to measure asteroseismic parameters for *Kepler* stars and has been extensively tested against other closed-source 
[…]
