+++ paper/paper.md
@@ -34,94 +34,119 @@ bibliography: paper.bib
 
 # Summary
 
-Asteroseismology, the study of stellar oscillations, is a powerful tool for studying the interiors of stars
-and determining their fundamental properties [@aerts2021]. For stars with temperatures that are similar to 
-the Sun, turbulent near-surface convection excites sound waves that propagate within the stellar cavity 
-[@bedding2014]. These waves penetrate into different depths within the star and therefore provide powerful 
-constraints on stellar interiors that would otherwise be inaccessible. Asteroseismology is well-established 
-in astronomy as the gold standard for characterizing fundamental properties like masses, radii, densities, 
-and ages for single stars, which has broad impacts on several fields in astronomy. For example, ages of stars 
-are important to reconstruct the formation history of the Milky Way (so-called galactic archeology). For 
-exoplanets that are discovered indirectly through changes in stellar observables, precise and accurate stellar 
-masses and radii are critical for learning about the planets that orbit them. 
-
-`pySYD` is a Python package for detecting solar-like oscillations and measuring global asteroseismic parameters given an input
-power spectrum. Derived parameters include seismic observables $\rm \nu_{max}$ and $\Delta\nu$, which are directly tied
-to the surface gravity and density of the star [@kjeldsen1995]. `pySYD` also performs a maximum likelihood estimation for 
-determining the stellar background contribution, characterizing different time scales and amplitudes of granulation 
-processes. 
+Asteroseismology, the study of stellar oscillations, is a powerful tool
+for studying the interiors of stars and determining their fundamental
+properties [@aerts2021]. For stars with temperatures that are similar
+to the Sun, turbulent near-surface convection excites sound waves that
+propagate within the stellar cavity [@bedding2014]. These waves
+penetrate into different depths within the star and therefore provide
+powerful constraints on stellar interiors that would otherwise be
+inaccessible. Asteroseismology is well-established in astronomy as the
+gold standard for characterizing fundamental properties like masses,
+radii, densities, and ages for single stars, which has broad impacts on
+several fields in astronomy. For example, ages of stars are important to
+reconstruct the formation history of the Milky Way (so-called galactic
+archaeology). For exoplanets that are discovered indirectly through
+changes in stellar observables, precise and accurate stellar masses and
+radii are critical for learning about the planets that orbit them.
 
 # Statement of Need
 
-The NASA space telescopes *Kepler*, K2 and TESS have recently provided very large databases of high-precision light curves 
-of stars. By detecting brightness variations due to solar-like oscillations, these light curves allow the application of 
-asteroseismology to large numbers of stars, which requires automated software tools to efficiently extract observables. 
-Several tools have been developed for asteroseismic analyses [e.g., `A2Z`, @mathur2010; `COR`, @mosser2009; `OCT`, @hekker2010; 
-`SYD`, @huber2009], but most of them are closed-source and thus inaccessible to the general astronomy community. Some open-source 
-tools exist [e.g., `DIAMONDS` and `FAMED`, @corsaro2014; `PBjam`, @nielsen2021; `lightkurve`, @lightkurve], but they are either 
-optimized for smaller samples of stars or have not yet been extensively tested against closed-source tools.
+The NASA space telescopes *Kepler*, K2 and TESS have recently
+provided very large databases of high-precision light curves of stars.
+By detecting brightness variations due to stellar oscillations, these
[…]
