+++ paper/paper.md
@@ -28,7 +28,7 @@ affiliations:
    index: 2
  - name: Sydney Institute for Astronomy, School of Physics, University of Sydney, NSW 2006, Australia
    index: 3
-date: 6 May 2021
+date: 1 Aug 2021
 bibliography: paper.bib
 ---
 
