+++ README.md
@@ -18,7 +18,7 @@ Written & developed by Ashley Chontos, Daniel Huber, Maryum Sayeed and Pavadol Y
 
 We thank Dennis Stello, Jie Yu, Marc Hon, and other users of the IDL `SYD` pipeline for discussion and suggestions which helped the development of this code.
 
-Please cite our [JOSS paper](https://joss.theoj.org/papers/6465a9dd3141c207175f200c7f891f1e) as well as the [original publication](https://ui.adsabs.harvard.edu/abs/2009CoAst.160...74H/abstract) discussing the methodology if you make use of `pySYD` in your work.
+Please cite our [JOSS paper](https://arxiv.org/abs/2108.00582) if you make use of `pySYD` in your work.
 
 ## Documentation
 
