### 2021-08-03 3c8054d
added FAQ section; changed table width of 1st column
