+++ docs/advanced.rst
@@ -114,16 +114,16 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 ``--lp: lower frequency (zoomed PS)``
 ++++++++
 
-+-------------------------------------------------------------------------------+-------------------------------------------------------------------------------+
-| Before                                                             			| After                                                                         |
-+===============================================================================+===============================================================================+
-| Set the lower frequency limit in zoomed in power spectrum; useful when an artefact is present close to the excess and cannot be removed otherwise.            |
-+-------------------------------------------------------------------------------+-------------------------------------------------------------------------------+
-| :bash:`pysyd run --star 10731424 --numax 750 --dnu 42`           	     		| :bash:`pysyd run --star 10731424 --numax 750 --dnu 42 --lp 480`               |
-+-------------------------------------------------------------------------------+-------------------------------------------------------------------------------+
-| .. figure:: figures/advanced/10731424_before.png                              | .. figure:: figures/advanced/10731424_after.png                               |
-|    :width: 600                                                                |    :width: 600                                                                |
-+-------------------------------------------------------------------------------+-------------------------------------------------------------------------------+
++--------------------------------------------------------------------------+--------------------------------------------------------------------------+
+| Before                                                                   | After                                                                    |
++==========================================================================+==========================================================================+
+| Set the lower frequency limit in zoomed in power spectrum; useful when an artefact is present close to the excess and cannot be removed otherwise.  |
++--------------------------------------------------------------------------+--------------------------------------------------------------------------+
+| :bash:`pysyd run --star 10731424 --numax 750 --dnu 42`                   | :bash:`pysyd run --star 10731424 --numax 750 --dnu 42 --lp 480`          |
++--------------------------------------------------------------------------+--------------------------------------------------------------------------+
+| .. figure:: figures/advanced/10731424_before.png                         | .. figure:: figures/advanced/10731424_after.png                          |
+|    :width: 600                                                           |    :width: 600                                                           |
++--------------------------------------------------------------------------+--------------------------------------------------------------------------+
 
 
 ``--npeaks: number of peaks``
