+++ docs/advanced.rst
@@ -74,7 +74,7 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 +==================================================================+==================================================================+
 | Changed the excess width in the background corrected power spectrum used to calculate the ACF (and hence dnu).                      |
 +------------------------------------------------------------------+------------------------------------------------------------------+
-| :bash:`pysyd run --star 9542776 --numax 900 --dnu 39.62`         | :bash:`pysyd run --star 9542776 --numax 900 --dnu 39.62 --ew 1.5`|
+| :bash:`pysyd run --star 9542776 --numax 900`                     | :bash:`pysyd run --star 9542776 --numax 900 --ew 1.5`            |
 +------------------------------------------------------------------+------------------------------------------------------------------+
 | .. figure:: figures/advanced/9542776_before.png                  | .. figure:: figures/advanced/9542776_after.png                   |
 |    :width: 600                                                   |    :width: 600                                                   |
@@ -119,7 +119,7 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 +==========================================================================+==========================================================================+
 | Set the lower frequency limit in zoomed in power spectrum; useful when an artefact is present close to the excess and cannot be removed otherwise.  |
 +--------------------------------------------------------------------------+--------------------------------------------------------------------------+
-| :bash:`pysyd run --star 10731424 --numax 750 --dnu 42`                   | :bash:`pysyd run --star 10731424 --numax 750 --dnu 42 --lp 480`          |
+| :bash:`pysyd run --star 10731424 --numax 750`                            | :bash:`pysyd run --star 10731424 --numax 750 --lp 490`                   |
 +--------------------------------------------------------------------------+--------------------------------------------------------------------------+
 | .. figure:: figures/advanced/10731424_before.png                         | .. figure:: figures/advanced/10731424_after.png                          |
 |    :width: 600                                                           |    :width: 600                                                           |
@@ -129,16 +129,16 @@ Below are examples of how to use specific ``pySYD`` features, as well as plots s
 ``--npeaks: number of peaks``
 ++++++++
 
-+-------------------------------------------------------------------------------+-------------------------------------------------------------------------------+
-| Before                                                            		    | After                                                                         |
-+===============================================================================+===============================================================================+
-| Change the number of peaks chosen in ACF; useful in low SNR cases where the spectrum is noisy and ACF has many peaks close to the expected dnu.               |
-+-------------------------------------------------------------------------------+-------------------------------------------------------------------------------+
-| :bash:`pysyd run --star 9455860 --numax 1487 --ew 1.2  --dnu 70.9` 			| :bash:`pysyd run --star 9455860 --numax 1487 --ew 1.2 --npeaks 20 --dnu 70.9` |
-+-------------------------------------------------------------------------------+-------------------------------------------------------------------------------+
-| .. figure:: figures/advanced/9455860_before.png                    			| .. figure:: figures/advanced/9455860_after.png                                |
-|    :width: 600                                                    		    |    :width: 600                                                                |
-+-------------------------------------------------------------------------------+-------------------------------------------------------------------------------+
++--------------------------------------------------------------------------+--------------------------------------------------------------------------+
+| Before                                                                   | After                                                                    |
++==========================================================================+==========================================================================+
+| Change the number of peaks chosen in ACF; useful in low SNR cases where the spectrum is noisy and ACF has many peaks close to the expected dnu.     |
++--------------------------------------------------------------------------+--------------------------------------------------------------------------+
+| :bash:`pysyd run --star 9455860`                                         | :bash:`pysyd run --star 9455860 --npeaks 10`                             |
++--------------------------------------------------------------------------+--------------------------------------------------------------------------+
[…]
