### 2021-08-03 cfd583c
changed low SNR example; changed command and results for npeaks, ew & lp
