### 2021-08-03 d964ca2
Update examples.rst
