+++ docs/examples.rst
@@ -17,47 +17,153 @@ Below are examples of medium to high signal-to-noise (SNR) detections in three s
 KIC 1435467
 +++++++++++
 
-KIC 1435467 is our least evolved example star, with numax ~1400 muHz.
+KIC 1435467 is our least evolved example star, with numax ~1400 muHz. The following command:
 
-``find_excess`` results:
+.. code-block::
+
+    $ pysyd run --star 1435467 --ux 5000 --ie -sv
+
+    ------------------------------------------------------
+    Target: 1435467
+    ------------------------------------------------------
+    # LIGHT CURVE: 37919 lines of data read
+    # Time series cadence: 59 seconds
+    # POWER SPECTRUM: 99518 lines of data read
+    # PS is oversampled by a factor of 5
+    # PS resolution: 0.426868 muHz
+    ------------------------------------------------------
+    Running find_excess module:
+    PS binned to 97 datapoints
+    power excess trial 1: numax = 1424.92 +/- 82.81
+    S/N: 2.85
+    power excess trial 2: numax = 1464.24 +/- 68.35
+    S/N: 6.27
+    power excess trial 3: numax = 1441.51 +/- 98.61
+    S/N: 22.32
+    selecting model 3
+    ------------------------------------------------------
+    Running fit_background module:
+    PS binned to 419 data points
+    Comparing 3 different models:
+    Model 0: Flat white-noise term only
+    Model 1: White noise term + 1 Harvey-like component(s)
+    Model 2: White noise term + 2 Harvey-like component(s)
+    Based on BIC statistic: model 1
+     **background-corrected PS saved**
+    ----------------------------------------------------
+    Output parameters:
+    numax_smooth: 1307.75 muHz
+    A_smooth: 1.77 ppm^2/muHz
+    numax_gauss: 1364.81 muHz
+    A_gauss: 1.45 ppm^2/muHz
+    FWHM: 280.26 muHz
+    dnu: 70.71 muHz
+    white: 1.42 ppm^2/muHz
+    tau_1: 219.15 s
+    sigma_1: 89.21 ppm
+    ----------------------------------------------------
+     - displaying figures
+     - press RETURN to exit
+     - combining results into single csv file
+    ------------------------------------------------------
+
+
+runs KIC 1435467 using the default method, which first runs ``find_excess`` followed by ``fit_global``.
+
+For the additional commands used:
+ - ``--ux 5000`` is the upper frequency bound of the power spectrum use during the first module 
+   (i.e. x, --lx would be the same but the lower bound). This is  strictly for computation purposes 
+   and does not alter or change the original power spectrum in any way.
+ - ``--ie`` turns the bicubic interpolation on when plotting the echelle diagram. This is 
+   particularly helpful for lower SNR examples like this. 
+ - ``-sv`` == `-s` + `-v` -> single hashes are reserved for boolean arguments, which correspond to 
+   ``show`` and ``verbose``, respectively. Since ``pySYD`` is optimized for many stars, both of these
+   options are ``False`` by default.
+   
+As you can read in the text output, the example started with n=2 Harvey-like components but reduced to 1 
+based on the BIC statistic. 
+
+The ``find_excess`` output figure:
 
 .. image:: figures/1435467_excess.png
   :width: 600
   :alt: Find excess output plot for KIC 1435467.
-  
-| **Top left:** Original time series.  
-| **Top middle:** Original power spectrum (white) and heavily smoothed power spectrum (green). The latter is used as an initial (crude) background fit to search for oscillations.  
-| **Top right:** Power spectrum after correcting the crude background fit.  
[…]
