### 2021-08-03 72ba9eb
Update examples.rst
