### 2021-08-03 8414207
Update examples.rst
