+++ docs/advanced.rst
@@ -2,12 +2,15 @@
 .. role:: bash(code)
    :language: bash
 
+FAQs & Advanced Usage
+########################
+
 FAQs
-#################
+*******
 
 #. I'm getting an error. What do I do?
 	
-	Please submit a pull request and include the version of ``pySYD`` used, the command you ran, screenshot of the error and any other details you can provide. The GitHub page for ``pySYD`` can be found here: `https://github.com/ashleychontos/pySYD`.
+	Please submit a pull request `here <https://github.com/ashleychontos/pySYD/pulls>`_. Be sure to include the version of ``pySYD`` used, the command you ran, screenshot of the error and any other details you can provide. The GitHub page for ``pySYD`` can be found here: `https://github.com/ashleychontos/pySYD`.
 
 #. What is the difference between ``--lx/--ux``, ``--lb/--ub`` and ``--lp/--up``?
 	
@@ -18,35 +21,35 @@ FAQs
 #. What's the difference between ``--bf``, ``--sw``, ``--sm``, ``--sp`` and ``--se``?
 	
 	- ``--bf``: smooths the power spectrum to calculate red noise components *after* the power excess has been masked out.
-	- ``--sw``: smooths the power spectrum in the first module when ``pySYD`` searches for the power excess. This value does not affect the power spectrum in the second module. The default is 50 muHz.
+	- ``--sw``: smooths the power spectrum in the first module when ``pySYD`` searches for the power excess. This value does not affect the power spectrum in the second module. The default value is 50 muHz.
 	- ``--sm``: smooths the power spectrum in the second module when ``pySYD`` measures numax as the center of the Gaussian around the power excess.
-	- ``--sp``: smooths the power spectrum in panel 5 that is used to calculate the ACF in panel 6. The default is 2.5 muHz.
+	- ``--sp``: smooths the power spectrum in panel 5 that is used to calculate the ACF in panel 6. The default value is 2.5 muHz.
 	- ``--se``: smooths the echelle diagram in panel 8. To use this argument, specify the smoothing box filter in muHz.
 
 #. I'm not happy with ``pySYD``'s measured dnu. How can I change the dnu value?
 	
-	The best way to change the dnu measurement is to specify a value of dnu using the `--dnu` flag. If the ACF contains many peaks around its higher amplitude peaks due to low signal-to-noise data, you can smooth the power spectrum that is used to calculate the ACF by changing the `--sp` flag. Alternatively, if the peak corresponding to the best dnu measurement is not within the five highest peaks, you can change the number of peaks found in the ACF by specifying `--npeaks`. 
+	The best way to change the dnu measurement is to specify a value of dnu using the ``--dnu`` flag (ie. ``--dnu 42.1``). If the ACF contains many peaks around its higher amplitude peaks due to low signal-to-noise data, you can smooth the power spectrum that is used to calculate the ACF by changing the ``--sp`` flag (ie. ``--sp 3.5``). Alternatively, if the peak corresponding to the best dnu measurement is not one of the five highest peaks, you can change the number of peaks found in the ACF by specifying ``--npeaks`` (ie. ``--npeaks 10``). 
 	
-	If the power spectrum contains mixed modes (common for evolved stars) that mask the dnu measurement, you can remove these mixed modes by specifying the lower and upper bound of where the mixed modes are found in the echelle diagram using the following flags: `--le` and `--ue`. Refer to the advanced usage section below for an example on how to use remove mixed modes. Coming soon!
+	If the power spectrum contains mixed modes that mask the dnu measurement, you can remove these mixed modes by specifying the lower and upper bound of where the mixed modes are found in the echelle diagram using the following flags: ``--le`` and ``--ue``. Refer to the advanced usage section below for an example on how to remove mixed modes. **This feature is coming soon!**
 
[…]
