### 2021-08-03 9b4c6d3
Update examples.rst
