### 2021-08-03 0a89b76
Update examples.rst
