### 2021-08-11 593c207
Fixing the smoothing for evolved stars with low numax
