+++ .github/workflows/joss-pdf.yml
@@ -1,19 +1,0 @@
-name: Compile JOSS paper
-on: [push]
-jobs:
-  paper:
-    runs-on: ubuntu-latest
-    name: Compile paper
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Build JOSS pdf
-        uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          paper-path: paper/paper.md
-      - name: Upload compiled pdf
-        uses: actions/upload-artifact@v1
-        with:
-          name: PDF
-          path: paper/paper.pdf
