### 2021-08-17 208a978
moving example figures for docs from docs/figures/ -> docs/figures/examples
