+++ docs/conf.py
@@ -85,7 +85,7 @@ html_theme = 'sphinx_rtd_theme'
 
 # The name of an image file (relative to this directory) to place at the top
 # of the sidebar.
-html_logo = "figures/pysyd_logo_rtd.png"
+html_logo = "figures/misc/pysyd_logo_rtd.png"
 
 # The name of an image file (within the static path) to use as favicon of the
 # docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
