### 2021-08-27 d31800a
Major changes in implementation and functionality, including: - can fix white noise now - can fix nlaws (without errors) - background-fitting only - global fits on raw power spectrum - echelle diagram plotting, including offset in double ED and number of orders plotted (--noy works now) - the way containers (dictionaries) are saved and loaded in (in pysyd/dicts), which will download as package data and users do not need to touch this
