+++ setup.py
@@ -35,6 +35,7 @@ setuptools.setup(
     ],
     install_requires=reqs,
     packages=setuptools.find_packages(),
+    package_data={"": ["dicts/*.dict"]},
     entry_points={'console_scripts':['pysyd=pysyd.cli:main']},
     python_requires=">=3.6",
 )
