### 2021-08-30 6f7a13f
Docs updates (CLI)
