+++ pysyd/cli.py
@@ -7,7 +7,11 @@ from pysyd import TODODIR, INFODIR, INPDIR, OUTDIR
 
 
 def main():
-    # Properties inherent to both modules
+
+#####################################################################
+# Initiate parser
+#
+
     parser = argparse.ArgumentParser(
                                      description="pySYD: Automated Extraction of Global Asteroseismic Parameters", 
                                      prog='pySYD',
@@ -18,10 +22,13 @@ def main():
                         help="Print version number and exit.",
     )
 
-    # In the parent parser, we define arguments and options common to all subcommands
+#####################################################################
+# Parent parser contains arguments and options common to all modes
+#
+
     parent_parser = argparse.ArgumentParser(add_help=False)
 
-    parent_parser.add_argument('--cli', 
+    parent_parser.add_argument('--cli', '-c',
                                dest='cli',
                                help='This option should not be adjusted for current users',
                                default=True,
@@ -62,7 +69,10 @@ def main():
                                action='store_true',
     )
 
-    # Main options 
+#####################################################################
+# Higher-level functionality relevant to any fitting routine
+#
+
     main_parser = argparse.ArgumentParser(add_help=False)
 
     main_parser.add_argument('--bg', '--background', '-b', 
@@ -109,7 +119,7 @@ def main():
                              default=False, 
                              action='store_true',
     )
-    main_parser.add_argument('--save',
+    main_parser.add_argument('--save', '-d',
                              dest='save',
                              help='Do not save output figures and results.',
                              default=True, 
@@ -142,7 +152,11 @@ def main():
                              action='store_false',
     )
 
-    # CLI relevant for finding power excess
+#####################################################################
+# CLI options relevant for estimating an approximate numax
+# (optional routine)
+#
+
     excess = main_parser.add_argument_group('(CRUDE) EXCESS FIT')
 
     excess.add_argument('--bin', '--binning',
@@ -171,7 +185,6 @@ def main():
                         metavar='freq', 
                         dest='lower_ex',
                         help='Lower frequency limit of PS',
-                        nargs='*',
                         default=1.0,
                         type=float,
     )
@@ -179,8 +192,7 @@ def main():
                         metavar='freq', 
                         dest='upper_ex',
                         help='Upper frequency limit of PS',
-                        nargs='*',
-                        default=None,
+                        default=6000.0,
                         type=float,
     )
     excess.add_argument('--step', '--steps', 
@@ -196,14 +208,16 @@ def main():
                         type=int,
     )
 
-    # CLI relevant for background fitting
+#####################################################################
+# CLI options relevant to the background-fitting
+#
+
     background = main_parser.add_argument_group('BACKGROUND-RELATED')
 
     background.add_argument('--lb', '--lowerb', 
                             metavar='freq', 
                             dest='lower_bg',
                             help='Lower frequency limit of PS',
-                            nargs='*',
                             default=1.0,
                             type=float,
     )
@@ -211,8 +225,7 @@ def main():
                             metavar='freq', 
                             dest='upper_bg',
                             help='Upper frequency limit of PS',
-                            nargs='*',
-                            default=None,
+                            default=6000.0,
                             type=float,
     )
     background.add_argument('--iw', '--indwidth',
@@ -269,7 +282,10 @@ def main():
                             action='store_true',
     )
 
-    # CLI relevant for the global parameters
+#####################################################################
+# CLI options related to estimating numax
+#
+
     numax = main_parser.add_argument_group('NUMAX-RELATED')
 
     numax.add_argument('--sm', '--smpar',
@@ -311,6 +327,10 @@ def main():
[…]
