### 2021-08-30 2c86cd9
minor documentation changes
