+++ docs/overview.rst
@@ -3,6 +3,24 @@
 Software Overview
 ####################
 
+
+``pySYD`` is a python-based implementation of the IDL-based SYD pipeline 
+`(Huber et al. 2009) <https://ui.adsabs.harvard.edu/abs/2009CoAst.160...74H/abstract>`_, 
+which was extensively used to measure asteroseismic parameters for Kepler stars. 
+Papers based on asteroseismic parameters measured using the ``SYD`` pipeline include 
+`Huber et al. 2011 <https://ui.adsabs.harvard.edu/abs/2011ApJ...743..143H/abstract>`_, 
+`Chaplin et al. 2014 <https://ui.adsabs.harvard.edu/abs/2014ApJS..210....1C/abstract>`_, 
+`Serenelli et al. 2017 <https://ui.adsabs.harvard.edu/abs/2017ApJS..233...23S/abstract>`_ 
+and `Yu et al. 2018 <https://ui.adsabs.harvard.edu/abs/2018ApJS..236...42Y/abstract>`_.
+
+``pySYD`` adapts the well-tested methodology from ``SYD`` while also improving these 
+existing analyses and expanding upon numerous new (optional) features. Improvements include:
+
+- Automated best-fit background model selection
+- Parallel processing
+- Easily accessible + command-line friendly interface
+- Ability to save samples for further analyses
+
 .. _modes:
 
 Modes
@@ -24,6 +42,8 @@ There are currently four main ``pySYD`` operational modes:
 #. ``parallel`` : Operates the same way as the previous mode, but processes stars simultaneously in parallel. Based on the number of threads
    available, stars are separated into groups (where the number of groups is exactly equal to the number of threads). This mode uses all CLI
    options, including the number of threads to use for parallelization (:ref:`see here<parallel>`).
+
+#. ``test`` : Currently under development.
 
 =========================
 
