### 2021-08-30 e690909
minor documentation changes
