### 2021-08-30 64336fb
Optional ntrials for find numax and alternatively, provide own estimate.
