+++ pysyd/cli.py
@@ -28,7 +28,7 @@ def main():
 
     parent_parser = argparse.ArgumentParser(add_help=False)
 
-    parent_parser.add_argument('--cli', '-c',
+    parent_parser.add_argument('-c', '--cli', 
                                dest='cli',
                                help='This option should not be adjusted for current users',
                                default=True,
@@ -62,7 +62,7 @@ def main():
                                type=str,
                                default=OUTDIR,
     )
-    parent_parser.add_argument('--verbose', '-v', 
+    parent_parser.add_argument('-v', '--verbose', 
                                dest='verbose',
                                help='Turn on verbose output',
                                default=False, 
@@ -70,30 +70,30 @@ def main():
     )
 
 #####################################################################
-# Higher-level functionality relevant to any fitting routine
+# Initial and/or final data treatment + related processes
 #
 
     main_parser = argparse.ArgumentParser(add_help=False)
 
-    main_parser.add_argument('--bg', '--background', '-b', 
+    main_parser.add_argument( '-b', '--bg', '--background',
                              dest='background',
                              help='Turn off the automated background fitting routine',
                              default=True, 
                              action='store_false',
     )
-    main_parser.add_argument('--globe', '--global', '-g', 
+    main_parser.add_argument('-d', '--show', '--display',
+                             dest='show',
+                             help='Show output figures',
+                             default=False, 
+                             action='store_true',
+    )
+    main_parser.add_argument('-g', '--globe', '--global', 
                              dest='globe',
                              help='Do not estimate global asteroseismic parameters (i.e. numax or dnu)',
                              default=True, 
                              action='store_false',
     )
-    main_parser.add_argument('--par', '--parallel', '-p',
-                             dest='parallel',
-                             help='Use parallel processing for data analysis',
-                             default=False,
-                             action='store_true',
-    )
-    main_parser.add_argument('--kc', '--kep_corr', '-k', 
+    main_parser.add_argument('-k', '--kc', '--kep_corr', 
                              dest='kep_corr',
                              help='Turn on the Kepler short-cadence artefact correction routine',
                              default=False, 
@@ -113,23 +113,23 @@ def main():
                              type=int,
                              default=None,
     )
-    main_parser.add_argument('--over', '--overwrite', '-o',
+    main_parser.add_argument('-o', '--over', '--overwrite',
                              dest='overwrite',
                              help='Overwrite existing files with the same name/path',
                              default=False, 
                              action='store_true',
     )
-    main_parser.add_argument('--save', '-d',
+    main_parser.add_argument('-p', '--par', '--parallel', 
+                             dest='parallel',
+                             help='Use parallel processing for data analysis',
+                             default=False,
+                             action='store_true',
+    )
+    main_parser.add_argument('-s', '--save',
                              dest='save',
                              help='Do not save output figures and results.',
                              default=True, 
                              action='store_false',
-    )
-    main_parser.add_argument('--show', '-s', 
-                             dest='show',
-                             help='Show output figures',
-                             default=False, 
-                             action='store_true',
     )
     main_parser.add_argument('--star', '--stars',
                              metavar='star',
@@ -139,13 +139,13 @@ def main():
                              nargs='*',
                              default=None,
     )
-    main_parser.add_argument('--test', '-t', 
+    main_parser.add_argument('-t', '--test', 
                              dest='testing',
                              help='Extra verbose output for testing functionality',
                              default=False, 
                              action='store_true',
     )
-    main_parser.add_argument('--ex', '--excess', '-x', 
+    main_parser.add_argument('-x', '--ex', '--excess',
[…]
