+++ pysyd/utils.py
@@ -188,7 +188,7 @@ def get_groups(args):
 #
 
 def get_excess_params(args, n_trials=3, step=0.25, binning=0.005, smooth_width=20.0, 
-                      mode='mean', lower_ex=1.0, upper_ex=8000.,):
+                      mode='mean', lower_ex=1.0, upper_ex=8000., ask=False,):
     """
     Get the parameters for the find excess routine.
 
@@ -196,14 +196,16 @@ def get_excess_params(args, n_trials=3, step=0.25, binning=0.005, smooth_width=2
     ----------
     args : argparse.Namespace
         the command line arguments
+    ask : bool, optional
+        If `True`, it will ask which trial to use as the estimate for numax.
+    n_trials : int, optional
+        the number of trials. Default value is `3`.
     step : float, optional
         TODO: Write description. Default value is `0.25`.
     binning : float, optional
         logarithmic binning width. Default value is `0.005`.
     mode : {'mean', 'median', 'gaussian'}
         mode to use when binning
-    n_trials : int, optional
-        the number of trials. Default value is `3`.
 
     Returns
     -------
@@ -213,11 +215,11 @@ def get_excess_params(args, n_trials=3, step=0.25, binning=0.005, smooth_width=2
         the parameters of the find excess routine
 
     """
-    vars = ['step', 'binning', 'mode', 'smooth_width', 'n_trials', 'lower_ex', 'upper_ex', 'results']
+    vars = ['step', 'binning', 'mode', 'smooth_width', 'ask', 'n_trials', 'lower_ex', 'upper_ex', 'results']
     if args.cli:
-        vals = [args.step, args.binning, args.mode, args.smooth_width, args.n_trials, args.lower_ex, args.upper_ex, {}]
+        vals = [args.step, args.binning, args.mode, args.smooth_width, args.ask, args.n_trials, args.lower_ex, args.upper_ex, {}]
     else:
-        vals = [step, binning, mode, smooth_width, n_trials, lower_ex, upper_ex, {}]
+        vals = [step, binning, mode, smooth_width, ask, n_trials, lower_ex, upper_ex, {}]
     args.excess = dict(zip(vars,vals))
     return args
 
@@ -907,7 +909,7 @@ def check_input_data(args, star, note):
 # Sets data up for first optional module
 #
 
-def get_estimates(star):
+def get_estimates(star, max_trials=6):
     """
     Parameters used with the first module, which is automated method to identify
     power excess due to solar-like oscillations.
@@ -945,10 +947,12 @@ def get_estimates(star):
             upper = star.nyquist
     star.freq = star.frequency[(star.frequency >= lower)&(star.frequency <= upper)]
     star.pow = star.power[(star.frequency >= lower)&(star.frequency <= upper)]
+    if star.excess['n_trials'] > max_trials:
+        star.excess['n_trials'] = max_trials
     if (star.params[star.name]['numax'] is not None and star.params[star.name]['numax'] <= 500.) or (star.nyquist is not None and star.nyquist <= 300.):
-        star.boxes = np.logspace(np.log10(0.5), np.log10(25.), star.excess['n_trials'])*1.
+        star.boxes = np.logspace(np.log10(0.5), np.log10(25.), star.excess['n_trials'])
     else:
-        star.boxes = np.logspace(np.log10(50.), np.log10(500.), star.excess['n_trials'])*1.
+        star.boxes = np.logspace(np.log10(50.), np.log10(500.), star.excess['n_trials'])
     return star
 
 #####################################################################
@@ -1489,6 +1493,53 @@ def bin_data(x, y, width, log=False, mode='mean'):
 # Other random tools, including a container class of known (mostly
 # solar) physical values in cgs.
 #
+
[…]

[…]
