+++ docs/cli.rst
@@ -77,6 +77,12 @@ the data as well as which modules to use.
    * type = ``bool``
    * default = ``True``
    * action = ``store_false``
+- ``-d``, ``--show``, ``--display``
+   * dest = ``args.show``
+   * help = show output figures (note: this is not recommended if running many stars)
+   * type = ``bool``
+   * default = ``False``
+   * action = ``store_true``
 - ``-g``, ``--globe``, ``--global``
    * dest = ``args.globe``
    * help = Do not estimate global asteroseismic parameters numax and dnu
@@ -117,24 +123,18 @@ the data as well as which modules to use.
    * type = ``bool``
    * default = ``True``
    * action = ``store_false``
-- ``-d``, ``--show``, ``--display``
-   * dest = ``args.show``
-   * help = show output figures (note: this is not recommended if running many stars)
-   * type = ``bool``
-   * default = ``False``
-   * action = ``store_true``
 - ``--star``, ``--stars``
    * dest = ``args.star``
    * help = List of stars to process. Default is ``None``, which will read in the star list from ``args.file``.
    * nargs = '*'
    * default = ``None``
-- ``--test``, ``-t``
+- ``-t``, ``--test``
    * dest = ``args.test``
    * help = Extra verbose output for testing functionality (not currently implemented)
    * type = ``bool``
    * default = ``False``
    * action = ``store_true``
-- ``--ex``, ``--excess``, ``-x``
+- ``-x``, ``--ex``, ``--excess``
    * dest = ``args.background``
    * help = turn off find excess module
    * type = ``bool``
@@ -143,11 +143,17 @@ the data as well as which modules to use.
 
 
 Estimating numax
-++++++++++++++++++++
+******************
 
 The following options are relevant for the first, optional module that is designed
 to estimate numax if it is not known: 
 
+- ``-a``, ``--ask``
+   * dest = ``args.ask``
+   * help = Ask which trial (or estimate) to use
+   * type = ``bool``
+   * default = ``False``
+   * action = ``store_true``
 - ``--bin``, ``--binning``
    * dest = ``args.binning``
    * help = Interval for binning of spectrum in log(muHz) (bins equally in logspace).
@@ -202,6 +208,39 @@ Granulation background
 
 Below is a complete list of parameters relevant to the background-fitting routine:
 
+- ``--basis``
+   * dest = ``args.basis``
+   * help = Which basis to use for background fit (i.e. 'a_b', 'pgran_tau', 'tau_sigma'), *** NOT operational yet ***
+   * type = str
+   * default = `'tau_sigma'`
+- ``--bf``, ``--box``, ``--boxfilter``
+   * dest = ``args.box_filter``
+   * help = Box filter width (in muHz) for plotting the power spectrum
+   * type = ``float``
+   * default = `1.0`
+   * unit = muHz
+- ``-f``, ``--fix``, ``--fixwn``
+   * dest = ``args.fix``
+   * help = Fix the white noise level
+   * type = ``bool``
+   * default = ``False``
+   * action = ``store_true``
+- ``-i``, ``--include``
+   * dest = ``args.include``
+   * help = Include metric values in verbose output, default is `False`.
+   * type = ``bool``
+   * default = ``False``
+   * action = ``store_true``
+- ``--iw``, ``--indwidth``
+   * dest = ``args.ind_width``
+   * help = Width of binning for power spectrum (in muHz)
+   * type = ``float``
+   * default = `20.0`
+- ``--laws``, ``--nlaws``
+   * dest = ``args.n_laws``
+   * help = Force the number of red-noise component(s)
+   * type = int
+   * default = `None`
 - ``--lb``, ``--lowerb``
    * dest = ``args.lower_bg``
[…]
