+++ docs/examples.rst
@@ -10,19 +10,21 @@ for three KIC stars in the **data/** directory. If so, then you are ready to tes
 ====================
 
 High SNR Examples
-*****************
+===================
 
 Below are examples of medium to high signal-to-noise (SNR) detections in three stars of different evolutionary states. The first example includes a brief description of the output plots.
 
 KIC 1435467
-+++++++++++
+*************
 
 KIC 1435467 is our least evolved example star, with numax ~1400 muHz. The following command:
 
-.. code-block::
 
-    $ pysyd run --star 1435467 --ux 5000 --ie -sv
+.. code-block:: bash
 
+    $ pysyd run --star 1435467 -dv
+    
+    
     ------------------------------------------------------
     Target: 1435467
     ------------------------------------------------------
@@ -32,36 +34,38 @@ KIC 1435467 is our least evolved example star, with numax ~1400 muHz. The follow
     # PS is oversampled by a factor of 5
     # PS resolution: 0.426868 muHz
     ------------------------------------------------------
-    Running find_excess module:
-    PS binned to 97 datapoints
-    power excess trial 1: numax = 1424.92 +/- 82.81
-    S/N: 2.85
-    power excess trial 2: numax = 1464.24 +/- 68.35
-    S/N: 6.27
-    power excess trial 3: numax = 1441.51 +/- 98.61
-    S/N: 22.32
-    selecting model 3
+    Estimating numax:
+    PS binned to 189 datapoints
+    Numax estimate 1: 1430.02 +/- 72.61
+    S/N: 2.43
+    Numax estimate 2: 1479.46 +/- 60.64
+    S/N: 4.87
+    Numax estimate 3: 1447.42 +/- 93.31
+    S/N: 13.72
+    Selecting model 3
     ------------------------------------------------------
-    Running fit_background module:
+    Determining background model:
     PS binned to 419 data points
-    Comparing 3 different models:
-    Model 0: Flat white-noise term only
-    Model 1: White noise term + 1 Harvey-like component(s)
-    Model 2: White noise term + 2 Harvey-like component(s)
-    Based on BIC statistic: model 1
+    Comparing 6 different models:
+    Model 0: 0 Harvey-like component(s) + white noise fixed
+    Model 1: 0 Harvey-like component(s) + white noise term
+    Model 2: 1 Harvey-like component(s) + white noise fixed
+    Model 3: 1 Harvey-like component(s) + white noise term
+    Model 4: 2 Harvey-like component(s) + white noise fixed
+    Model 5: 2 Harvey-like component(s) + white noise term
+    Based on BIC statistic: model 2
      **background-corrected PS saved**
-    ----------------------------------------------------
+    ------------------------------------------------------
     Output parameters:
-    numax_smooth: 1307.75 muHz
-    A_smooth: 1.77 ppm^2/muHz
-    numax_gauss: 1364.81 muHz
-    A_gauss: 1.45 ppm^2/muHz
-    FWHM: 280.26 muHz
-    dnu: 70.71 muHz
-    white: 1.42 ppm^2/muHz
-    tau_1: 219.15 s
-    sigma_1: 89.21 ppm
-    ----------------------------------------------------
+    tau_1: 233.71 s
+    sigma_1: 87.45 ppm
+    numax_smooth: 1299.56 muHz
+    A_smooth: 1.75 ppm^2/muHz
+    numax_gauss: 1345.03 muHz
+    A_gauss: 1.49 ppm^2/muHz
+    FWHM: 291.32 muHz
+    dnu: 70.63 muHz
+    ------------------------------------------------------
      - displaying figures
      - press RETURN to exit
      - combining results into single csv file
@@ -76,24 +80,24 @@ Additional commands used in this example (and what they each mean):
[…]
