+++ docs/quickstart.rst
@@ -150,79 +150,96 @@ this figure:
   :width: 680
   :alt: Global fit for KIC 1435467
 
-To estimate uncertainties in the derived parameters, set ``-mc`` to a number sufficient for bootstrap sampling. In the previous 
-example, ``-mc`` was not specified and is 1 by default (for 1 iteration). Below shows the same example with the
+To estimate uncertainties in the derived parameters, set ``--mc`` to a number sufficient for bootstrap sampling. In the previous 
+example, ``--mc`` was not specified and is 1 by default (for 1 iteration). Below shows the same example with the
 sampling enabled, including the verbose output you should see if your software was installed successfully.
 
 .. code-block:: bash
 
-    $ pysyd run -star 1435467 -show -verbose -mc 200
+    $ pysyd run -star 1435467 -dv --mc 200
+        
     
-    ----------------------------------------------------
+    ------------------------------------------------------
     Target: 1435467
-    ----------------------------------------------------
+    ------------------------------------------------------
     # LIGHT CURVE: 37919 lines of data read
     # Time series cadence: 59 seconds
     # POWER SPECTRUM: 99518 lines of data read
     # PS is oversampled by a factor of 5
     # PS resolution: 0.426868 muHz
-    # For first iteration: using oversampled PS [of 5]
-    ----------------------------------------------------
-    Running find_excess module:
-    PS binned to 319 datapoints
-    power excess trial 1: numax = 1438.23 +/- 71.21
-    S/N: 2.95
-    power excess trial 2: numax = 1473.82 +/- 64.02
-    S/N: 5.79
-    power excess trial 3: numax = 1449.98 +/- 97.23
-    S/N: 23.96
-    selecting model 3
-    ----------------------------------------------------
-    Running fit_background module:
-    PS binned to 1709 data points
-    Comparing 4 different models:
-    1: one harvey model w/ white noise free parameter
-    2: one harvey model w/ white noise fixed
-    3: two harvey model w/ white noise free parameter
-    4: two harvey model w/ white noise fixed
-    Based on reduced chi-squared statistic: model 4
-    ----------------------------------------------------
+    ------------------------------------------------------
+    Estimating numax:
+    PS binned to 189 datapoints
+    Numax estimate 1: 1430.02 +/- 72.61
+    S/N: 2.43
+    Numax estimate 2: 1479.46 +/- 60.64
+    S/N: 4.87
+    Numax estimate 3: 1447.42 +/- 93.31
+    S/N: 13.72
+    Selecting model 3
+    ------------------------------------------------------
+    Determining background model:
+    PS binned to 419 data points
+    Comparing 6 different models:
+    Model 0: 0 Harvey-like component(s) + white noise fixed
+    Model 1: 0 Harvey-like component(s) + white noise term
+    Model 2: 1 Harvey-like component(s) + white noise fixed
+    Model 3: 1 Harvey-like component(s) + white noise term
+    Model 4: 2 Harvey-like component(s) + white noise fixed
+    Model 5: 2 Harvey-like component(s) + white noise term
+    Based on BIC statistic: model 2
+     **background-corrected PS saved**
+    ------------------------------------------------------
     Running sampling routine:
-    100%|████████████████████████████████████████████████████████████████████████████████| 200/200 [00:16<00:00, 11.83it/s]
-
[…]
