+++ pysyd/functions.py
@@ -1,364 +1,0 @@
-import numpy as np
-import pandas as pd
-from collections import deque
-from astropy.convolution import Box1DKernel, Gaussian1DKernel, convolve, convolve_fft
-
-
-
-def set_seed(star, lower=1, upper=10**7, size=1):
-    """
-    For Kepler targets that require a correction via CLI (--kc), a random seed is generated
-    from U~[1,10^7] and stored in stars_info.csv for reproducible results in later runs.
-
-    Parameters
-    ----------
-    star : target.Target
-        the pySYD pipeline object
-    lower : int 
-        lower limit for random seed value. Default value is `1`.
-    upper : int
-        upper limit for random seed value. Default value is `10**7`.
-    size : int
-        number of seed values returned. Default value is `1`.
-
-    Returns
-    -------
-    star : target.Target
-        the pySYD pipeline object
-        
-    """
-
-    seed = list(np.random.randint(lower,high=upper,size=size))
-    df = pd.read_csv(star.params['info'])
-    stars = [str(each) for each in df.stars.values.tolist()]
-    idx = stars.index(star.name)
-    df.loc[idx,'seed'] = int(seed[0])
-    star.params[star.name]['seed'] = seed[0]
-    df.to_csv(star.params['info'],index=False)
-    return star
-
-
-def remove_artefact(star, lcp=1.0/(29.4244*60*1e-6), lf_lower=[240.0,500.0], lf_upper =[380.0,530.0], 
-                    hf_lower = [4530.0,5011.0,5097.0,5575.0,7020.0,7440.0,7864.0],
-                    hf_upper = [4534.0,5020.0,5099.0,5585.0,7030.0,7450.0,7867.0],):
-    """
-    Removes SC artefacts in Kepler power spectra by replacing them with noise (using linear interpolation)
-    following a chi-squared distribution. 
-
-    Known artefacts are:
-    1) 1./LC harmonics
-    2) high frequency artefacts (>5000 muHz)
-    3) low frequency artefacts 250-400 muHz (mostly present in Q0 and Q3 data)
-
-    Parameters
-    ----------
-    star : target.Target
-        the pySYD pipeline object
-    lcp : float
-        long cadence period in Msec
-    lf_lower : List[float]
-        lower limit of low frequency artefact
-    lf_upper : List[float]
-        upper limit of low frequency artefact
-    hf_lower : List[float]
-        lower limit of high frequency artefact
-    hf_upper : List[float]
-        upper limit of high frequency artefact
-    Returns
-    -------
-    star : target.Target
-        the pySYD pipeline object
-    """
-
-    if star.params[star.name]['seed'] is None:
-        star = set_seed(star)
-    # LC period in Msec -> 1/LC ~muHz
-    artefact = (1.0+np.arange(14))*lcp
-    # Estimate white noise
-    white = np.mean(star.power[(star.frequency >= max(star.frequency)-100.0)&(star.frequency <= max(star.frequency)-50.0)])
-
-    np.random.seed(int(star.params[star.name]['seed']))
-    # Routine 1: remove 1/LC artefacts by subtracting +/- 5 muHz given each artefact
-    for i in range(len(artefact)):
-        if artefact[i] < np.max(star.frequency):
-            mask = np.ma.getmask(np.ma.masked_inside(star.frequency, artefact[i]-5.0*star.resolution, artefact[i]+5.0*star.resolution))
-            if np.sum(mask) != 0:
-                star.power[mask] = white*np.random.chisquare(2,np.sum(mask))/2.0
-
-    np.random.seed(int(star.params[star.name]['seed']))
-    # Routine 2: fix high frequency artefacts
-    for lower, upper in zip(hf_lower, hf_upper):
[…]
