+++ docs/quickstart.rst
@@ -1,14 +1,14 @@
 .. _quickstart:
 
 Getting Started
-###############
+##################
 
 Jump down to :ref:`summary` to get asteroseismic parameters for a star in less than a minute!
 
 .. _installation:
 
 Installation
-************
+***************
 
 Install ``pysyd`` using pip:
 
@@ -24,12 +24,12 @@ top-level ``pysyd`` directory and try running the following command:
 
     $ python setup.py install
 
-You may test your installation by using ``pysyd --help`` to see available command-line options:
+You may test your installation by using ``pysyd --help`` to see the available commands or modes:
 
 .. code-block::
 		
     $ pysyd --help
-    usage: pySYD [-h] [-version] {setup,load,run,parallel} ...
+    usage: pySYD [-h] [-version] {load,parallel,run,setup,test} ...
 
     pySYD: Automated Extraction of Global Asteroseismic Parameters
 
@@ -44,6 +44,151 @@ You may test your installation by using ``pysyd --help`` to see available comman
         run                 Run the main pySYD pipeline
         setup               Easy setup of relevant directories and files
         test                Test different utilities (currently under development)
+	
+	
+A majority of the data analyses and subsequent tools are used in the most common mode, ``run``,
+and therefore, using the following command::
+
+    $ pysyd run --help                     
+    usage: pySYD run [-h] [-c] [--file path] [--in path] [--info path]
+                     [--out path] [-v] [-b] [-d] [-g] [-k] [--ofa n] [--ofn n]
+                     [-o] [-p] [-s] [--star [star [star ...]]] [-t] [-x] [-a]
+                     [--bin value] [--bm mode] [--lx freq] [--step value]
+                     [--trials n] [--sw value] [--ux freq] [--basis str]
+                     [--bf value] [-f] [-i] [--iw value] [--laws n] [--lb freq]
+                     [--metric metric] [--rms n] [--ub freq] [--ew value]
+                     [--lp [freq [freq ...]]] [--numax [value [value ...]]]
+                     [--sm value] [--up [freq [freq ...]]]
+                     [--dnu [value [value ...]]] [--method method] [--peak n]
+                     [--sp value] [--thresh value] [--ce cmap] [--cv value] [-y]
+                     [-e] [--le [freq [freq ...]]] [--notch] [--nox n] [--noy n]
+                     [--se value] [--ue [freq [freq ...]]] [--mc n] [-m]
+
+    optional arguments:
+      -h, --help            show this help message and exit
+      -c, --cli             This option should not be adjusted for current users
+      --file path, --list path, --todo path
+                            List of stars to process
+      --in path, --input path, --inpdir path
+                            Input directory
+      --info path, --information path
+                            Path to star info
+      --out path, --outdir path, --output path
+                            Output directory
+      -v, --verbose         Turn on verbose output
+      -b, --bg, --background
+                            Turn off the automated background fitting routine
+      -d, --show, --display
+                            Show output figures
+      -g, --globe, --global
+                            Do not estimate global asteroseismic parameters (i.e.
+                            numax or dnu)
+      -k, --kc, --kep_corr  Turn on the Kepler short-cadence artefact correction
+                            routine
+      --ofa n, --of_actual n
+                            The oversampling factor (OF) of the input PS
+      --ofn n, --of_new n   The OF to be used for the first iteration
+      -o, --over, --overwrite
+                            Overwrite existing files with the same name/path
+      -p, --par, --parallel
+                            Use parallel processing for data analysis
+      -s, --save            Do not save output figures and results.
+      --star [star [star ...]], --stars [star [star ...]]
+                            List of stars to process
+      -t, --test            Extra verbose output for testing functionality
[…]
