### 2021-08-31 080e08c
Update examples.rst
