+++ docs/examples.rst
@@ -2,7 +2,7 @@
 
 
 Examples
-########
+############
 
 If ``pysyd setup`` was successfully executed, there should now be light curves and power spectra 
 for three KIC stars in the **data/** directory. If so, then you are ready to test out the software!
@@ -12,7 +12,8 @@ for three KIC stars in the **data/** directory. If so, then you are ready to tes
 High SNR Examples
 ===================
 
-Below are examples of medium to high signal-to-noise (SNR) detections in three stars of different evolutionary states. The first example includes a brief description of the output plots.
+Below are three examples of medium to high signal-to-noise (SNR) detections for stars of different evolutionary states. See the 
+:ref:`description<description>` for more details about the output figures.
 
 KIC 1435467
 *************
@@ -70,15 +71,18 @@ KIC 1435467 is our least evolved example star, with numax ~1400 muHz. The follow
     ------------------------------------------------------
 
 
-runs KIC 1435467 using the default method, which first runs ``find_excess`` followed by ``fit_global``.
+runs KIC 1435467 using the default method, which first estimates numax before deriving any parameters. 
+The reason for this is that the frequency region with power excess is masked out to better estimate 
+background parameters.
 
 Additional commands used in this example (and what they each mean):
+
  - ``--ux 5000`` is the upper frequency bound of the power spectrum used during the first module 
    (i.e. 'x' for excess, ``--lx`` would be the same but the lower bound for this module). These bounds  
    are used strictly for computational purposes and do not alter or change the power spectrum in any way.
  - ``--ie`` turns the bicubic interpolation on when plotting the \'echelle diagram. This is 
    particularly helpful for lower SNR examples like this. 
- - ``-dv`` == `-d` + `-v` -> single hashes are reserved for boolean arguments, which correspond to 
+ - ``-dv`` == ``-d`` + ``-v`` -> single hashes are reserved for boolean arguments, which correspond to 
    ``display`` and ``verbose``, respectively. Since ``pySYD`` is optimized for many stars, both of these
    options are ``False`` by default.
    
@@ -98,9 +102,7 @@ The derived parameters from the global fit are summarized below:
   :alt: Global fit of KIC 1435467
 
 
-.. note::
-
-    For a breakdown of what each panel in each figure means, please see ref for more details.
+**For a breakdown of what each panel in each figure means, please see :ref:<description> for more details.**
   
   
 The derived parameters are saved to an output csv file but also printed at the end of the verbose output.
@@ -160,18 +162,23 @@ To quantify uncertainties in these parameters, we need to turn on the Monte Carl
     ------------------------------------------------------
     
  
-
 where the first 2/3 of the output is (and should be) identical to the first example. By default, 
 ``--mc == 1`` since you should always check your results first before running ``pySYD`` for
 several iterations! The method used to derive the uncertainties is similar to a 
-bootstrapping technique and typically n=200 is more than sufficient. You may also use the ``--samples``
-option if you would like to save the posteriors of the parameters for later use.
+bootstrapping technique and typically n=200 is more than sufficient. 
 
-The Monte Carlo ``sampling`` results:
+Parameter posteriors:
 
 .. image:: figures/examples/1435467_samples.png
   :width: 680
-  :alt: Parameter posteriors for KIC 1435467.
+  :alt: Parameter posteriors for KIC 1435467
+  
+  
+.. note::
+
[…]
