### 2021-08-31 e052059
Update performance.rst
