+++ docs/performance.rst
@@ -1,7 +1,14 @@
 .. _performance:
 
+=============
 Performance
-###########
+=============
+
+‣ hey
+‣ ashley
+‣ tf you doin
+
+idk.
 
 .. _comparison:
 
@@ -10,7 +17,7 @@ The resulting values are compared for the two methods below for the frequency of
 (left) and the large frequency separation (Dnu) on the right. For reference,
 ``SYD == pySYD`` and ``IDL == SYD``.
 
-.. image:: figures/new.png
+.. image:: figures/performance/comparison.png
 
 There residuals show no strong systematics to within <0.5% in Dnu and <~1% in numax, which is smaller than the typical 
 random uncertainties. This confirms that the open-source python package pySYD provides consistent results with the legacy 
