### 2021-08-31 e22af60
Update cli.rst
