### 2021-08-31 52eaa4f
Update performance.rst
