+++ docs/performance.rst
@@ -4,12 +4,6 @@
 Performance
 =============
 
-‣ hey
-‣ ashley
-‣ tf you doin
-
-idk.
-
 .. _comparison:
 
 We ran pySYD on ~100 Kepler legacy stars observed in short-cadence and compared the output to IDL SYD results from `Serenelli et al. 2017 <https://ui.adsabs.harvard.edu/abs/2017ApJS..233...23S/abstract>`_. The same time series and power spectra were used for both analyses.
