+++ docs/utils.rst
@@ -2,9 +2,9 @@
 
 Utilities
 =========
-Utilities are mostly non-science related functions and capabilities. Examples include
-functions to load and check input data, reading in appropriate information 
-dictionaries for relevant stars, etc.
+Container classes, parameter dictionaries, functions related to file loading and/or
+saving as well as various data manipulation methods (i.e. correcting artefacts, 
+binning data, etc.).
 
 .. module:: pysyd
 
