### 2021-08-31 4627bf8
Delete functions.rst
