+++ docs/functions.rst
@@ -1,12 +1,0 @@
-.. _functions:
-
-Functions
-=========
-Data manipulation methods (i.e. smoothing functions, binning data) or other 
-science-related functions (i.e. scaling relations) which are directly implemented 
-in the main pipeline.
-
-.. module:: pysyd
-
-.. automodule:: pysyd.functions
-   :members:
