+++ docs/quickstart.rst
@@ -45,157 +45,7 @@ You may test your installation by using ``pysyd --help`` to see the available co
         run                 Run the main pySYD pipeline
         setup               Easy setup of relevant directories and files
         test                Test different utilities (currently under development)
-	
-	
-A majority of the data analyses and subsequent tools are used in the most common mode, ``run``. To see what
-options and parameters are available for the main pipeline execution, we'll use the following command:
 
-.. code-block::
-
-    $ pysyd run --help                     
-    usage: pySYD run [-h] [-c] [--file path] [--in path] [--info path]
-                     [--out path] [-v] [-b] [-d] [-g] [-k] [--ofa n] [--ofn n]
-                     [-o] [-p] [-s] [--star [star [star ...]]] [-t] [-x] [-a]
-                     [--bin value] [--bm mode] [--lx freq] [--step value]
-                     [--trials n] [--sw value] [--ux freq] [--basis str]
-                     [--bf value] [-f] [-i] [--iw value] [--laws n] [--lb freq]
-                     [--metric metric] [--rms n] [--ub freq] [--ew value]
-                     [--lp [freq [freq ...]]] [--numax [value [value ...]]]
-                     [--sm value] [--up [freq [freq ...]]]
-                     [--dnu [value [value ...]]] [--method method] [--peak n]
-                     [--sp value] [--thresh value] [--ce cmap] [--cv value] [-y]
-                     [-e] [--le [freq [freq ...]]] [--notch] [--nox n] [--noy n]
-                     [--se value] [--ue [freq [freq ...]]] [--mc n] [-m]
-
-    optional arguments:
-      -h, --help            show this help message and exit
-      -c, --cli             This option should not be adjusted for current users
-      --file path, --list path, --todo path
-                            List of stars to process
-      --in path, --input path, --inpdir path
-                            Input directory
-      --info path, --information path
-                            Path to star info
-      --out path, --outdir path, --output path
-                            Output directory
-      -v, --verbose         Turn on verbose output
-      -b, --bg, --background
-                            Turn off the automated background fitting routine
-      -d, --show, --display
-                            Show output figures
-      -g, --globe, --global
-                            Do not estimate global asteroseismic parameters (i.e.
-                            numax or dnu)
-      -k, --kc, --kep_corr  Turn on the Kepler short-cadence artefact correction
-                            routine
-      --ofa n, --of_actual n
-                            The oversampling factor (OF) of the input PS
-      --ofn n, --of_new n   The OF to be used for the first iteration
-      -o, --over, --overwrite
-                            Overwrite existing files with the same name/path
-      -p, --par, --parallel
-                            Use parallel processing for data analysis
-      -s, --save            Do not save output figures and results.
-      --star [star [star ...]], --stars [star [star ...]]
-                            List of stars to process
-      -t, --test            Extra verbose output for testing functionality
-      -x, --ex, --excess    Turn off the find excess routine
-
-    Estimate numax:
-      -a, --ask             Ask which trial to use
-      --bin value, --binning value
-                            Binning interval for PS (in muHz)
-      --bm mode, --mode mode, --bmode mode
-                            Binning mode
-      --lx freq, --lowerx freq
-                            Lower frequency limit of PS
-      --step value, --steps value
-      --trials n, --ntrials n
-      --sw value, --smoothwidth value
-                            Box filter width (in muHz) for smoothing the PS
-      --ux freq, --upperx freq
-                            Upper frequency limit of PS
-
-    Background Fit:
-      --basis str           Which basis to use for background fit (i.e. 'a_b',
-                            'pgran_tau', 'tau_sigma'), *** NOT operational yet ***
-      --bf value, --box value, --boxfilter value
-                            Box filter width [in muHz] for plotting the PS
[…]
