### 2021-08-31 b8e1773
Update cli.rst
