+++ docs/conf.py
@@ -68,7 +68,7 @@ extensions = [
     'sphinx.ext.intersphinx',
     'sphinx.ext.todo',
     'sphinx.ext.coverage',
-    'sphinx.ext.pngmath',
+#    'sphinx.ext.pngmath',
     'sphinx.ext.ifconfig',
     'matplotlib.sphinxext.only_directives',
     'matplotlib.sphinxext.plot_directive',
