+++ README.md
@@ -19,7 +19,7 @@ Written & developed by Ashley Chontos, Maryum Sayeed, Daniel Huber, and Pavadol 
 We thank Dennis Stello, Jie Yu, Marc Hon, Yifan Chen, Yaguang Li, and other `pySYD` users for 
 discussion and suggestions which helped with the development of this code.
 
-Please cite our [JOSS paper](https://arxiv.org/abs/2108.00582) if you make use of `pySYD` in your work. The recommended BibTeX entry for this citation is::
+Please cite our [JOSS paper](https://arxiv.org/abs/2108.00582) if you make use of `pySYD` in your work. The recommended BibTeX entry for this citation is:
 
     {@ARTICLE{2021arXiv210800582C,
            author = {{Chontos}, Ashley and {Huber}, Daniel and {Sayeed}, Maryum and {Yamsiri}, Pavadol},
