+++ docs/index.rst
@@ -42,7 +42,7 @@ Bug reports and feature requests should be posted to the `GitHub issue tracker <
 .. toctree::
    :maxdepth: 1
    :hidden: 
-   :caption: Getting Started
+   :caption: Intro
 
    quickstart
    overview
