### 2021-08-31 6839965
Update cli.rst
