### 2021-08-31 e06f248
Update cli.rst
