### 2021-08-31 b69353f
Update cli.rst
