### 2021-08-31 10627fd
Update cli.rst
