### 2021-08-31 2d64d60
Update cli.rst
