### 2021-08-31 42f85ea
Update cli.rst
