### 2021-09-01 59e0b4f
Update examples.rst
