### 2021-09-01 cb8c1a5
Update cli.rst
