### 2021-09-01 95034f7
Update cli.rst
