### 2021-10-01 #3331 @ashleychontos
Thank you for clarifying @mbobra - the primary three developers for `pySYD` are tied up at the moment (which includes myself, currently busy with postdoc applications), hence the delay. I think we still want to pursue it and also agree with all the reviewers comments, but a likely time scale when I can revisit this is ~2 months? Is that an ok amount of time to be paused?
