### 2021-10-01 #3331 @danielskatz
In my opinion, this is getting close to the far edge of acceptable, but if the reviewers are willing to wait and then take another look at that time, we can accept it.
