### 2021-10-01 #3331 @benjaminpope
Fine by me!

———————————————
Dr Benjamin Pope (he/him)
Lecturer in Astrophysics
University of Queensland
benjaminpope.github.io
