+++ README.md
@@ -7,6 +7,7 @@
 [![Documentation](https://readthedocs.org/projects/pysyd/badge/?version=latest)](https://pysyd.readthedocs.io/en/latest/?badge=latest)
 [![License: MIT](https://img.shields.io/badge/License-MIT-orange.svg)](https://opensource.org/licenses/MIT)
 [![JOSS](https://joss.theoj.org/papers/6465a9dd3141c207175f200c7f891f1e/status.svg)](https://joss.theoj.org/papers/6465a9dd3141c207175f200c7f891f1e)
+<a href="https://ascl.net/2111.017"><img src="https://img.shields.io/badge/ascl-2111.017-blue.svg?colorB=262255" alt="ascl:2111.017" /></a>
 
 </div>
 
