+++ CITATION.cff
@@ -10,8 +10,9 @@ authors:
 - family-names: "Sayeed"
   given-names: "Maryum"
   orcid: "https://orcid.org/0000-0001-6180-8482"
+- family-names: "Yamsiri"
+  given-names: Pavadol
 title: "pySYD"
 version: 1.6.6
-doi: 10.5281/zenodo.5140574
 date-released: 2021-07-28
-url: "https://github.com/ashleychontos/pySYD"
+url: "https://ui.adsabs.harvard.edu/abs/2021arXiv210800582C"
