### 2022-02-03 #3331 @mbobra
@ashleychontos How is it going? Do you think this submission is ready to undergo the rest of the review process?

If not, I suggest you withdraw the submission and submit again. The review process is not really a long-term holding place, and we're at the 8-month mark here so it is time to wrap up or withdraw. However, there's absolutely no issue whatsoever with re-submitting -- and if you already have willing reviewers at the ready, the editor will be more than happy to re-assign them to the review. And I'm also happy to act as the editor again! You can request me when you re-submit. So basically you lose nothing and gain the time you need to polish this up. Please let me know what you think!
