### 2022-02-11 #3331 @benjaminpope
Certainly!

———————————————
Dr Benjamin Pope (he/him)
Lecturer in Astrophysics
University of Queensland
benjaminpope.github.io
