### 2022-03-04 #3331 @ashleychontos
Great, thank you @mbobra!

@MaryumSayeed and I just finished chatting about the action items for `pySYD` and both agreed that 2 weeks seems like a reasonable amount of time to finish up the remaining tasks (also paging @danxhuber). _**However,**_ her and I will reconvene again in a week to make sure we are still good for time. I have a general question (that would possibly affect the timeline) -- does the JOSS submission depend on any current repo "issues"?

Again, many thanks all (@benjaminpope @danhey) for being incredibly patient! We are actively working on addressing the original comments/suggestions/etc. now :)
