+++ CONTRIBUTING.md
@@ -1,0 +1,160 @@
+# Contribute to `pySYD`
+
+We welcome pull requests to `pySYD`. If you'd like to add a new symbol, or try to
+tackle adding a larger feature, keep reading. If you have any questions, or want
+help solving a problem, feel free to stop by our [GitHub Discussions](https://github.com/KaTeX/KaTeX/discussions).
+
+## Helpful contributions
+
+If you'd like to contribute, try contributing new symbols or functions that
+KaTeX doesn't currently support. The documentation has pages listing
+[supported functions](https://katex.org/docs/supported.html) and
+[functions that KaTeX supports and some that it doesn't support](https://katex.org/docs/support_table.html).
+You can check them to see if we don't support a function you like, or try your
+function in the interactive demo at [http://katex.org/](http://katex.org/).
+The wiki has a page that describes how to [examine TeX commands and where to find
+rules](https://github.com/KaTeX/KaTeX/wiki/Examining-TeX) which can be quite
+useful when adding new commands.
+
+#### Single symbols
+
+There are many individual symbols that KaTeX doesn't yet support. Read through
+the [symbols.js](src/symbols.js) file for more information on how to add a
+symbol.
+
+To figure out the unicode symbol for the symbol you are trying to add, try using
+the symbol in MathJax to see what unicode symbol it outputs. An interactive
+MathJax shell can be found [here](http://fiddle.jshell.net/YpqVp/41/show/).
+
+To figure out what group your symbol falls into, look through the symbols list
+to find other symbols of a similar kind. (e.g. if you were adding `\neq`, look
+for `=`). If you cannot find anything similar, or are unsure, you can try using
+your symbol in TeX surrounded by other different kinds of symbols, and seeing
+whether your spacing matches the spacing that TeX produces.
+
+Once your symbol works, check the JavaScript console to make sure you don't get
+a message like "Can't find character metrics for \_" when you render your symbol.
+If you do, check out [extract_ttfs.py](metrics/extract_ttfs.py).
+
+#### Adding new functions
+
+New functions should be added in [src/functions](src/functions) using
+`defineFunction` from [defineFunction.js](src/defineFunction.js).  Read the
+comments in this file to get started.  Look at
+[phantom.js](src/functions/phantom.js) and
+[delimsizing.js](src/functions/delimsizing.js) as examples of how to use
+`defineFunction`.  Notice how delimsizing.js groups several related functions
+together in a single call to `defineFunction`.
+
+The new method of defining functions combines methods that were previously
+spread out over three different files [functions.js](src/functions.js),
+[buildHTML.js](src/buildHTML.js), [buildMathML.js](src/buildMathML.js) into a
+single file.  The goal is to have all functions use this new system.
+
+#### Macros
+
+Macros should be added in [src/macros.js](src/macros.js) using `defineMacro`.
+They are expanded in the "gullet" (`MacroExpander`).
+
+## Testing
+
+Local testing can be done by running the webpack-dev-server using configuration
+`webpack.dev.js`. Run `yarn` to install dependencies, and then `yarn start`
+to start the server.
+
+This will host an interactive editor at
+[http://localhost:7936/](http://localhost:7936/) to play around with and test
+changes.
+
+#### Jest tests
+
+The JavaScript parser and some of the HTML and MathML tree
+builders are tested with Jest. These tests can be run using node with
+`yarn test:jest`.  If you need to debug the tests see
+[https://facebook.github.io/jest/docs/troubleshooting.html](https://facebook.github.io/jest/docs/troubleshooting.html)
+
+The interactive editor can also be used for debugging tests in the browser by
+copy/pasting the test case to be debugged into the editor.  The permalink option
[…]
