+++ .github/issue_template.md
@@ -1,0 +1,27 @@
+## Title (aka "Issue")
+
+### Description
+
+A clear and concise description of what the issue/bug is
+
+### Reproducibility
+
+In order to reproduce the issue from our end, please provide *both* the light curve and power spectrum for the relevant target
+
+### Computer (please complete the following):
+
+ - **OS:** [e.g. Linux, macOS, Microsoft Windows,]
+ - **Version:** [e.g. v2.1.2]
+
+### Screenshots
+
+If applicable, add screenshots to help explain your problem
+
+### Additional notes
+Add any other information here 
+
+### Checklist -
+
+**Required Materials** | <br/> <ul> <li>- [ ] Description</li> <li>- [ ] Power spectrum</li> <li>- [ ] Light curve</li> <li>- [ ] Computer OS</li> <li>- [ ] Software version</li> </ul> 
+:--------------- | :--------------------------------------------------------------------------------
+**Extra Information** | <ul><li>- [ ] Screenshots</li><li>- [ ] Notes</li></ul>     
