+++ CONTRIBUTING.md
@@ -1,42 +1,186 @@
-# Contribute to `pySYD`
+# Contributing to `pySYD`
 
-We welcome pull requests to `pySYD`. If you'd like to add a new symbol, or try to
-tackle adding a larger feature, keep reading. If you have any questions, or want
-help solving a problem, feel free to stop by our [GitHub Discussions](https://github.com/KaTeX/KaTeX/discussions).
+We would love for you to contribute to `pySYD` and help make asteroseismology even more accessible than it is today! 
+For any general support questions or help solving a problem, feel free to stop by our [Discussion Page](). If you 
+have ideas or suggestions for improvements, we welcome any and all pull requests! As a contributor, here are the 
+guidelines we ask you to follow:
 
-## Helpful contributions
+- [Question or problem?](#question)
+- [Issues & bugs](#issue)
+- [New ideas](#feature)
+- [Guidelines](#submit)
+- [Style guide](#style)
+- [Testing](#test)
 
-If you'd like to contribute, try contributing new symbols or functions that
-KaTeX doesn't currently support. The documentation has pages listing
-[supported functions](https://katex.org/docs/supported.html) and
-[functions that KaTeX supports and some that it doesn't support](https://katex.org/docs/support_table.html).
-You can check them to see if we don't support a function you like, or try your
-function in the interactive demo at [http://katex.org/](http://katex.org/).
-The wiki has a page that describes how to [examine TeX commands and where to find
-rules](https://github.com/KaTeX/KaTeX/wiki/Examining-TeX) which can be quite
-useful when adding new commands.
 
-#### Single symbols
+## <a name="question"></a> Have a general question or problem?
 
-There are many individual symbols that KaTeX doesn't yet support. Read through
-the [symbols.js](src/symbols.js) file for more information on how to add a
-symbol.
+We may already have the answer for you! Please visit our [discussion forum]() to see if your question has already been answered in one of the threads. You can also help us keep this up-to-date and the most helpful for the community by adding relevant tags wherever possible.
 
-To figure out the unicode symbol for the symbol you are trying to add, try using
-the symbol in MathJax to see what unicode symbol it outputs. An interactive
-MathJax shell can be found [here](http://fiddle.jshell.net/YpqVp/41/show/).
+Please do *not* open issues for general support questions as we want to preserve GitHub issues for bug reports and feature requests. You've got a much better chance of getting your question(s) answered in our [discussion forum](stackoverflow.com/questions/tagged/angular-devtools), with a chance that it may have already been answered before! 
 
-To figure out what group your symbol falls into, look through the symbols list
-to find other symbols of a similar kind. (e.g. if you were adding `\neq`, look
-for `=`). If you cannot find anything similar, or are unsure, you can try using
-your symbol in TeX surrounded by other different kinds of symbols, and seeing
-whether your spacing matches the spacing that TeX produces.
+To save everyone some time, we will be systematically closing all issues that are general support questions and redirecting people to the discussion board.
 
-Once your symbol works, check the JavaScript console to make sure you don't get
-a message like "Can't find character metrics for \_" when you render your symbol.
-If you do, check out [extract_ttfs.py](metrics/extract_ttfs.py).
+If you would like to chat about the question in real-time, please contact Ashley via email to set up a zoom call.
 
-#### Adding new functions
+
+## <a name="issue"></a> Did you find a bug?
+
+If you find a bug in the source code, you can help us by
+[submitting an issue](#submit-issue) to our [GitHub Repository][github]. Even better, you can
+[submit a Pull Request](#submit-pr) with a fix.
+
+
+
[…]
