+++ AUTHORS.md
@@ -1,0 +1,7 @@
+**Main author:** Ashley Chontos @ashleychontos achontos@hawaii.edu
+
+**Other contributors:**
+- Maryum Sayeed
+- Daniel Huber
+- Pavadol Yamsiri
+- 
