+++ AUTHORS.md
@@ -1,7 +1,7 @@
-**Main author:** Ashley Chontos @ashleychontos achontos@hawaii.edu
+**Main author:** Ashley Chontos ([@ashleychontos](https://github.com/ashleychontos>))
 
-**Other contributors:**
-- Maryum Sayeed
-- Daniel Huber
-- Pavadol Yamsiri
+**All contributors:**
+- Huber, Daniel ([@danxhuber](https://github.com/danxhuber))
+- Sayeed, Maryum ([@MaryumSayeed](https://github.com/MaryumSayeed))
+- Yamsiri, Pavadol ([]())
 - 
