+++ AUTHORS.md
@@ -1,6 +1,6 @@
 **Main author:** Ashley Chontos ([@ashleychontos](https://github.com/ashleychontos>))
 
-**All contributors:**
+**All contributors:** (sorted alphabetically)
 - Huber, Daniel ([@danxhuber](https://github.com/danxhuber))
 - Sayeed, Maryum ([@MaryumSayeed](https://github.com/MaryumSayeed))
 - Yamsiri, Pavadol ([]())
