### 2022-03-06 e7cd41f
Create CODE_OF_CONDUCT.md
