### 2022-03-06 3662a68
Delete CITATION.cff
