+++ CITATION.cff
@@ -1,18 +1,0 @@
-cff-version: 1.2.0
-message: "If you use this software, please cite it as below."
-authors:
-- family-names: "Chontos"
-  given-names: "Ashley"
-  orcid: "https://orcid.org/0000-0003-1125-2564"
-- family-names: "Huber"
-  given-names: "Daniel"
-  orcid: "https://orcid.org/0000-0001-8832-4488"
-- family-names: "Sayeed"
-  given-names: "Maryum"
-  orcid: "https://orcid.org/0000-0001-6180-8482"
-- family-names: "Yamsiri"
-  given-names: Pavadol
-title: "pySYD"
-version: 1.6.6
-date-released: 2021-07-28
-url: "https://ui.adsabs.harvard.edu/abs/2021arXiv210800582C"
