+++ .github/ISSUE_TEMPLATE/custom.md
@@ -1,10 +1,12 @@
 ---
-name: Custom issue template
-about: Describe this issue template's purpose here.
+name: Are we missing something?
+about: Please let us know.
 title: ''
 labels: ''
 assignees: ''
 
 ---
 
+### Description
 
+Concisely summarize what you think can be added to make your experience better.
