+++ .github/ISSUE_TEMPLATE/custom.md
@@ -1,12 +1,0 @@
----
-name: Are we missing something?
-about: Please let us know.
-title: ''
-labels: ''
-assignees: ''
-
----
-
-### Description
-
-Concisely summarize what you think can be added to make your experience better.
