+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -7,14 +7,22 @@ assignees: ''
 
 ---
 
-**Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+## Request Checklist
 
-**Describe the solution you'd like**
-A clear and concise description of what you want to happen.
+Is this a major request? | <br/><ul><li>- [ ] Yes </li><li>- [ ] No</li></ul>
+:---- | :----
+Do you have a proposed <br/> solution? | <br/><ul><li>- [ ] Yes </li><li>- [ ] No</li></ul>
+If so, will you be submitting <br/> a pull request? | <br/><ul><li>- [ ] Yes </li><li>- [ ] No</li><li>- [ ] N/A</li></ul>
+What type of request is it? | <br/><ul><li>- [ ] New feature </li><li>- [ ] Modification </li><li>- [ ] Suggested update</li><li>- [ ] Other? (fill answer here)</li></ul>
+Why is it needed? | <br/><ul><li>- [ ] General improvement</li><li>- [ ] Solves a problem </li><li>- [ ] Relevant to an open issue </li><li>- [ ] Increased efficiency and/or speed</li><li>- [ ] Other? (fill answer here)</li></ul>
 
-**Describe alternatives you've considered**
-A clear and concise description of any alternative solutions or features you've considered.
+## Information & supporting files
 
-**Additional context**
-Add any other context or screenshots about the feature request here.
+### Description
+
+Provide a clear and concise explanation for what you would like changed, including why you would like to see it implemented. For example, is it related to a problem or open issue? If so, please explain in detail and provide any other relevant information (i.e. links, screenshots, etc.). Are there any alternative solutions you've considered?
+
+### Other
+
+Please add other context or any relevant files to support the request.
+
