+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -7,28 +7,34 @@ assignees: ''
 
 ---
 
+## Checklist
+
+**Required**<br/>**Materials** | <br/> <ul> <li>- [ ] Description</li> <li>- [ ] Power spectrum</li> <li>- [ ] Light curve</li> <li>- [ ] Computer OS</li> <li>- [ ] Software version</li> </ul> 
+:--------------- | :--------------------------------------------------------------------------------
+**Extra**<br/>**Information** | <br/> <ul><li>- [ ] Screenshots</li><li>- [ ] Notes</li></ul>
+
+## Information
+
 ### Description
 
-A clear and concise description of what the issue/bug is
+A clear description of what the issue is. Some talking points include:
+- what's happening? i.e. is the code crashing completely or are you getting unexpected output?
+- if it's an error, what does it say (screenshots are helpful here)?
+- at which step does it occur?
+- have you identified a workaround?
+- do you have a proposed solution?
 
-### Reproducibility
+### Computing
 
-In order to reproduce the issue from our end, please provide *both* the light curve and power spectrum for the relevant target
-
-### Computer (please complete the following):
+Please complete by editing the following:
 
  - **OS:** [e.g. Linux, macOS, Microsoft Windows,]
  - **Version:** [e.g. v2.1.2]
 
-### Screenshots
+### Relevant files
 
-If applicable, add screenshots to help explain your problem
+**In order to reproduce the issue from our end, please provide *both* the light curve and power spectrum for the relevant target!**<br/>
+Attach the required data files here and if applicable, any figures or screenshots to help explain your problem.
 
-### Additional notes
-Add any other information here 
-
-### Checklist -
-
-**Required Materials** | <br/> <ul> <li>- [ ] Description</li> <li>- [ ] Power spectrum</li> <li>- [ ] Light curve</li> <li>- [ ] Computer OS</li> <li>- [ ] Software version</li> </ul> 
-:--------------- | :--------------------------------------------------------------------------------
-**Extra Information** | <br/> <ul><li>- [ ] Screenshots</li><li>- [ ] Notes</li></ul>
+### Notes
+Add any other relevant information here (e.g., links to issues or solutions)
