+++ .github/pull-request-template.md
@@ -1,15 +1,22 @@
-### Pull request summary
-
-**Some talking points as a guide:**
-- brief description of what was done 
-  - is this a modification to existing software 
-  - does it include any new features or functions
-- why or what prompted the change
-- if changes are substantial or include any new functions, are they properly commented throughout the code and documentation? please see [here](CONTRIBUTING.md) for more details.
-- reference or link if the pull request directly addresses an existing issue
-
-#### Checklist
+## Request Checklist
 
 * [ ] Did you test your changes? (please do ***NOT*** submit the pull request until you have)
-* [ ] Did you add other relevant materials (e.g., docstrings, documentation)?
+* [ ] Is your code compliant with our recommended style guide?
+* [ ] Is the code complete with docstrings and other relevant documentation?
+* [ ] Will it require any changes from our end? If so, please explain.
+* [ ] Did you add other relevant materials or files?
 * [ ] Did you provide links for any address issues?
+
+## Information & Supporting Files
+
+### Description
+
+Brief summary of the pull request content i.e. is this a modification to existing software or does it include any new features or functions? Why or what prompted the change? Are changes substantial or include any new functions? If so, are they properly commented throughout the code and documentation? Please see [here](https://github.com/ashleychontos/pySYD/blob/master/CONTRIBUTING.md) for more details.
+
+### Attachments
+
+Please provide any relevant files i.e. data if adding a new target to a test or as an example
+
+### Other
+
+Any other relevant points i.e. reference/link if directly addressing an existing issue
