+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -7,29 +7,23 @@ assignees: ''
 
 ---
 
-## Report Checklist
+------
 
-**Required**<br/>**Materials** | <br/> <ul> <li>- [ ] Description</li> <li>- [ ] Power spectrum</li> <li>- [ ] Light curve</li> <li>- [ ] Computer OS</li> <li>- [ ] Software version</li> </ul> 
-:--------------- | :--------------------------------------------------------------------------------
-**Extra**<br/>**Information** | <br/> <ul><li>- [ ] Screenshots</li><li>- [ ] Notes</li></ul>
-
-## Information & Supporting Files
+## Relevant Materials
 
 ### Description
 
-A clear description of what the issue is. Some talking points include:
-- what's happening? i.e. is the code crashing completely or are you getting unexpected output?
-- if it's an error, what does it say (screenshots are helpful here)?
-- at which step does it occur?
-- have you identified a workaround?
-- do you have a proposed solution?
+What's happening -- is the code crashing completely or are you seeing unexpected output? If it's an error, what does it say (feel free to provide screenshots)? At what step does the problem occur? When this happens, are you using the software via command line or within scripts? Is the issue specific to the software version? In other words, has it worked for you before? If so, do you know what version it last worked with (approximate is ok too)? 
+
+Is this already a known problem? If so, has there already been a proposed workaround or solution? Did you happen to identify a temporary workaround or solution? If yes, will you be submitting a pull request to address this? (please link the two if you do!) 
 
 ### Computing
 
-Please complete by editing the following:
+Please provide (by editing) the following information:
 
- - **OS:** [e.g. Linux, macOS, Microsoft Windows,]
- - **Version:** [e.g. v2.1.2]
+ - **OS:** [e.g., Linux, macOS, Microsoft Windows,]
+ - **Version:** [e.g., v2.1.2]
+ - **Implementation:** [e.g., command line, Jupyter, scripts,]
 
 ### Attachments
 
@@ -37,4 +31,20 @@ Please complete by editing the following:
 Attach the required data files here and if applicable, any figures or screenshots to help explain your problem.
 
 ### Other
-Add any other relevant information here (e.g., links to issues or solutions)
+Add any other relevant information here (e.g., links to related issues or solutions)
+
+------
+
+## Summary Checklist:
+
+- How were you using `pySYD` when you experienced the issue? <ul><li>- [ ] CLI</li><li>- [ ] API</li><li>- [ ] jupyter</li></ul>
+- Did you provide all the relevant information and files so that we can reproduce the error? <ul><li>- [ ] Yes</li><li>- [ ] No</li></ul>
+- Did you check if the issue has already been reported? <ul><li>- [ ] Yes <br/> &rightarrow; Was there a proposed workaround or solution? <ul><li>- [ ] Yes</li><li>- [ ] No</li><li>- [ ] Idk</li></ul></li><li>- [ ] No </li></ul>
+- Do you know if the issue is version specific? <ul><li>- [ ] Yes* </li><li>- [ ] No</li><li>- [ ] Idk</li></ul>
+- Have you identified a workaround or solution? <ul><li>- [ ] Yes* </li><li>- [ ] No</li></ul>
+- Will you be submitting a pull request associated with this issue? <ul><li>- [ ] Yes </li><li>- [ ] No</li><li>- [ ] Idk</li></ul>
+
+<br/>
+* If you selected 'Yes' for any of these points, please expand on each of them in more detail below.<br/> Note: If you found any relevant threads, questions, solutions or are submitting a pull request, please remember to provide links for any/all as these are incredibly helpful for future users. 
+
+------
