### 2022-03-06 42aa1b6
Update and rename api.rst to user_guide.rst
