+++ docs/api.rst
@@ -1,14 +1,0 @@
-.. _api:
-
-API
-===
-
-.. toctree::
-   :maxdepth: 2
-
-   pipeline
-   target
-   functions
-   models
-   utils	      
-   plots
