+++ docs/user_guide.rst
@@ -1,0 +1,14 @@
+.. _user_guide:
+
+User Guide
+===========
+
+.. toctree::
+   :maxdepth: 1
+
+   CLI
+   pipeline
+   target
+   models
+   utils	      
+   plots
