### 2022-03-06 e16af18
Update user_guide.rst
