### 2022-03-06 ca474e1
Update cli.rst
