+++ docs/conf.py
@@ -68,10 +68,8 @@ extensions = [
     'sphinx.ext.intersphinx',
     'sphinx.ext.todo',
     'sphinx.ext.coverage',
-#    'sphinx.ext.pngmath',
+    'sphinx.ext.panels',
     'sphinx.ext.ifconfig',
-#    'matplotlib.sphinxext.only_directives',
-#    'matplotlib.sphinxext.plot_directive',
  ]
 
 #extensions.append('sphinx.ext.pngmath')
