+++ docs/conf.py
@@ -1,13 +1,9 @@
 # -*- coding: utf-8 -*-
 #
-# pySYD documentation buld configuration file, created by
-# sphinx-quickstart on 
-
-# Configuration file for the Sphinx documentation builder.
-#
 # This file only contains a selection of the most common options. For a full
 # list see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
+
 
 # -- Path setup --------------------------------------------------------------
 
@@ -33,6 +29,7 @@ nbsphinx_allow_errors = True
 
 # -- Project information -----------------------------------------------------
 
+
 project = u'pySYD'
 copyright = u'2021, Ashley Chontos, Daniel Huber, and Maryum Sayeed'
 author = u'Ashley Chontos, Daniel Huber, and Maryum Sayeed'
@@ -44,6 +41,7 @@ release = pysyd.__version__
 
 
 # -- General configuration ---------------------------------------------------
+
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
@@ -80,23 +78,27 @@ exclude_patterns = ['_build']
 
 # -- Options for HTML output -------------------------------------------------
 
+
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-#
+
 html_theme = 'sphinx_rtd_theme'
 
 # The name of an image file (relative to this directory) to place at the top
 # of the sidebar.
+
 html_logo = "figures/misc/pysyd_logo_rtd.png"
 
 # The name of an image file (within the static path) to use as favicon of the
 # docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
 # pixels large.
+
 #html_favicon = "figures/pysyd_logo_favicon.ico"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
+
 html_static_path = ['_static']
 
 latex_logo = '_static/latex.png'
