+++ docs/conf.py
@@ -55,8 +55,9 @@ extensions = [
     'sphinx.ext.graphviz',
     'sphinx.ext.ifconfig',
     'sphinx.ext.intersphinx',
+    'sphinx.ext.panels',
     'sphinx.ext.todo',
-    'sphinx_panels',
+#    'sphinx_panels',
  ]
 
 # The suffix(es) of source filesnames.
