+++ docs/conf.py
@@ -12,14 +12,14 @@
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
 
-#import mock
 import os
 import sys
+import mock
 sys.path.insert(0, os.path.abspath('.'))
 
-#autodoc_mock_imports = ['scipy', 'pandas', 'numpy', 'nbsphinx']
-#for mod_name in autodoc_mock_imports:
-#    sys.modules[mod_name] = mock.Mock()
+autodoc_mock_imports = ['scipy', 'pandas', 'numpy', 'sphinx_panels']
+for mod_name in autodoc_mock_imports:
+    sys.modules[mod_name] = mock.Mock()
 
 import pysyd
 import sphinx_rtd_theme
@@ -48,7 +48,6 @@ release = pysyd.__version__
 # ones.
 
 extensions = [
-    'sphinx_panels',
     'sphinx.ext.autodoc',
     'sphinx.ext.autosummary',
     'sphinx.ext.coverage',
