+++ docs/installation.rst
@@ -11,7 +11,7 @@ Installation
 
 
 With pip [recommended]
-########################
+***********************
 
 Install the latest stable version using pip:
 
@@ -28,7 +28,7 @@ top-level ``pysyd`` directory and try running the following command:
     python setup.py install
 
 From Source
-###############
+************
 
 The latest development version can be cloned from GitHub using git:
 
@@ -45,7 +45,7 @@ Then to build and install the project use:
 from inside the cloned ``pySYD`` directory.
 
 Testing
-#########
+********
 
 You can test your installation by using the help command: 
     
@@ -69,7 +69,7 @@ You can test your installation by using the help command:
 
 
 Setup
-******
+########
 
 The easiest way to start using the ``pySYD`` package is by running our ``setup`` feature
 from a convenient directory:
@@ -127,14 +127,14 @@ which will print the absolute paths of all directories that are created during s
 .. _installation/dependencies:
 
 Python Dependencies
-*********************
+####################
 
 List of dependencies.
 
 .. _installation/example:
 
 Example 
-*********
+#########
 
 The ``setup`` feature downloads data for three example stars (to data/) provided: 1435467 (the least evolved), 
 2309595 (~SG), and 11618103 (RGB). To run a single star, execute the main script with the following command:
@@ -285,7 +285,7 @@ as well as descriptions about what the plots are showing.
 .. _installation/quickstart:
 
 Quickstart
-************
+############
 
 To get started right away, use the following commands:
 
