### 2022-03-07 4d71fa2
Update examples.rst
