+++ docs/conf.py
@@ -56,7 +56,7 @@ extensions = [
 # The suffix(es) of source filesnames.
 # You can specify multiple suffix as a list of string:
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ['.rst', '.md']
 
 # The master toctree document.
 master_doc = 'index'
