### 2022-03-08 407de31
Update and rename AUTHORS.md to AUTHORS.rst
