+++ AUTHORS.md
@@ -1,7 +1,0 @@
-**Main author:** Ashley Chontos ([@ashleychontos](https://github.com/ashleychontos>))
-
-**All contributors:** (sorted alphabetically)
-- Grayson, Sam ([@charmoniumQ](https://github.com/charmoniumQ))
-- Huber, Daniel ([@danxhuber](https://github.com/danxhuber))
-- Sayeed, Maryum ([@MaryumSayeed](https://github.com/MaryumSayeed))
-- Yamsiri, Pavadol ([@pavyamsiri](https://github.com/pavyamsiri))
