+++ AUTHORS.rst
@@ -1,0 +1,9 @@
+**Main author:** Ashley Chontos (`git <https://github.com/ashleychontos>`_ | `email <achontos@hawaii.edu>`_ | `website <https://ashleyin.space>`_)
+
+**All contributors:** (sorted alphabetically)
+
+* Chontos, Ashley (`@ashleychontos <https://github.com/ashleychontos>`_)
+* Grayson, Sam (`@charmoniumQ <https://github.com/charmoniumQ>`_)
+* Huber, Daniel (`@danxhuber <https://github.com/danxhuber>`_)
+* Sayeed, Maryum (`@MaryumSayeed <https://github.com/MaryumSayeed>`_)
+* Yamsiri, Pavadol (`@pavyamsiri <https://github.com/pavyamsiri>`_)
