### 2022-03-08 6a5dbf6
Update and rename docs/performance.rst to docs/introduction/performance.rst
