### 2022-03-08 e9ea274
Update and rename docs/faq.rst to docs/advanced/faq.rst
