+++ docs/advanced/faq.rst
@@ -1,8 +1,8 @@
-.. _faq:
+.. _faq/index:
 
-############
+****
 FAQs
-############
+****
 
 
 
