### 2022-03-08 f1bca70
Update cli.rst
