+++ docs/advanced/guide/models.rst
@@ -5,7 +5,5 @@ Models
 Frequency domain distributions (i.e. Gaussian, Lorentzian, Harvey, etc.)
 used in the pySYD pipeline.
 
-.. module:: pysyd
-
 .. automodule:: pysyd.models
    :members:
