### 2022-03-08 ef99a50
Update and rename docs/pipeline.rst to docs/advanced/guide/pipeline.rst
