+++ docs/advanced/guide/pipeline.rst
@@ -6,7 +6,5 @@ Used with the command line interface (CLI), which will accept the command line a
 and initialize the appropriate module. This includes the main pipeline initialization, which
 now has parallelization capabilities.
 
-.. module:: pysyd
-
 .. automodule:: pysyd.pipeline
    :members:
