### 2022-03-08 7f595af
Update and rename docs/plots.rst to docs/advanced/guide/plots.rst
