+++ docs/advanced/guide/plots.rst
@@ -4,7 +4,5 @@ Plots
 =====
 All plotting routines.
 
-.. module:: pysyd
-
 .. automodule:: pysyd.plots
    :members:
