+++ docs/advanced/guide/target.rst
@@ -4,7 +4,5 @@ Target
 ======
 The main pipeline class object that is initialized for each processed star.
 
-.. module:: pysyd
-
 .. automodule:: pysyd.target
    :members:
