+++ docs/introduction/attribution.rst
@@ -1,0 +1,4 @@
+***********
+Attribution
+***********
+
