+++ docs/introduction/contribution.rst
@@ -1,0 +1,3 @@
+************
+Contribution
+************
