+++ docs/index.rst
@@ -42,9 +42,9 @@ Our list of contributors continues to grow!
 Citations
 =========
 
-Check out what ``pySYD`` has been used for already! 
+Check out `projects <https://ui.adsabs.harvard.edu/abs/2021arXiv210800582C/citations>`_ that have already used ``pySYD``! 
 
-Please see this page for information on how to properly cite the relevant sources and software.
+Please see :ref:`this page <introduction/attribution>` for information on how to properly cite the relevant sources and software.
 
 .. toctree::
    :maxdepth: 1
