+++ docs/index.rst
@@ -11,7 +11,7 @@ pySYD: |br| Automated Measurements |br| of Global Asteroseismic Parameters
 Asteroseismology, or the study of stellar oscillations, is a powerful tool
 for studying the internal structure of stars and determining their fundamental 
 properties. Specifically for stars similar to the Sun, turbulent near-surface 
-convection excites sound waves that propagate within the stellar cavity, hence
+convection excites sound waves that propagate within the stellar cavity, and hence
 providing powerful constraints on stellar interiors that are inaccessible by 
 any other means. While it is well-established and widely-accepted as the gold 
 standard for characterizing fundamental stellar properties (e.g., masses, radii,
