+++ docs/introduction/attribution.rst
@@ -1,4 +1,0 @@
-***********
-Attribution
-***********
-
