+++ docs/introduction/citing.rst
@@ -1,0 +1,4 @@
+*********
+Citations
+*********
+
