### 2022-03-08 d8f9c4d
Update citing.rst
