### 2022-03-08 f379e56
Update AUTHORS.rst
