+++ docs/COLLABORATORS.rst
@@ -1,0 +1,5 @@
+**All collaborators:** 
+
+* Tim Bedding
+* Marc Hon
+* Dennis Stello
