+++ docs/CONTRIBUTORS.rst
@@ -1,9 +1,9 @@
 **Main author:** Ashley Chontos (`email <achontos@hawaii.edu>`_ | `website <https://ashleyin.space>`_)
 
-**All contributors:** (sorted alphabetically)
+**All contributors:**
 
-* Chontos, Ashley (`@ashleychontos <https://github.com/ashleychontos>`_)
-* Grayson, Sam (`@charmoniumQ <https://github.com/charmoniumQ>`_)
-* Huber, Daniel (`@danxhuber <https://github.com/danxhuber>`_)
-* Sayeed, Maryum (`@MaryumSayeed <https://github.com/MaryumSayeed>`_)
-* Yamsiri, Pavadol (`@pavyamsiri <https://github.com/pavyamsiri>`_)
+* Ashley Chontos (`@ashleychontos <https://github.com/ashleychontos>`_)
+* Sam Grayson (`@charmoniumQ <https://github.com/charmoniumQ>`_)
+* Daniel Huber (`@danxhuber <https://github.com/danxhuber>`_)
+* Maryum Sayeed (`@MaryumSayeed <https://github.com/MaryumSayeed>`_)
+* Pavadol Yamsiri (`@pavyamsiri <https://github.com/pavyamsiri>`_)
