+++ docs/introduction/attribution.rst
@@ -4,7 +4,7 @@ Citing
 
 .. _citing-citations:
 
-If you found `pySYD` useful and make use of it in your work, please cite our `JOSS paper <https://arxiv.org/abs/2108.00582>`_:
+If you found ``pySYD`` useful and make use of it in your work, please cite our `JOSS paper <https://arxiv.org/abs/2108.00582>`_:
 
 .. code-block::
 
