+++ docs/citing.rst
@@ -1,0 +1,62 @@
+******
+Citing
+******
+
+.. _citing/main:
+
+If you found ``pySYD`` useful and make use of it in your work, please cite our `JOSS paper <https://arxiv.org/abs/2108.00582>`_:
+
+.. code-block::
+
+    @ARTICLE{2021arXiv210800582C,
+           author = {{Chontos}, Ashley and {Huber}, Daniel and {Sayeed}, Maryum and {Yamsiri}, Pavadol},
+            title = "{$\texttt{pySYD}$: Automated measurements of global asteroseismic parameters}",
+          journal = {arXiv e-prints},
+         keywords = {Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+             year = 2021,
+            month = aug,
+              eid = {arXiv:2108.00582},
+            pages = {arXiv:2108.00582},
+    archivePrefix = {arXiv},
+           eprint = {2108.00582},
+     primaryClass = {astro-ph.SR}, 
+           adsurl = {https://ui.adsabs.harvard.edu/abs/2021arXiv210800582C},
+          adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    }
+    
+If applicable, please also use our `ASCL listing <https://ui.adsabs.harvard.edu/abs/2021ascl.soft11017C>`_ as a software citation:
+
+.. code-block::
+
+    @MISC{2021ascl.soft11017C,
+           author = {{Chontos}, Ashley and {Huber}, Daniel and {Sayeed}, Maryum and {Yamsiri}, Pavadol},
+            title = "{pySYD: Measuring global asteroseismic parameters}",
+         keywords = {Software},
+             year = 2021,
+            month = nov,
+              eid = {ascl:2111.017},
+            pages = {ascl:2111.017},
+    archivePrefix = {ascl},
+           eprint = {2111.017},
+           adsurl = {https://ui.adsabs.harvard.edu/abs/2021ascl.soft11017C},
+          adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    }
+
+.. note::
+
+    ``pySYD`` is only possible thanks to many powerful Python libraries -- therefore
+    we strongly encourage you to also cite :ref:`its dependencies <installation-dependencies>`.
+
+
+----
+
+.. _citing/projects:
+
+Projects w/ ``pySYD``
+=====================
+
+We added a project thread in the public repository called `Projects w/ pySYD <https://github.com/ashleychontos/pySYD/discussions/28>`_.
+Please consider adding to this in any way -- anything from early results, figures, student projects to published manuscripts, posters,
+or any other ``pySYD``-related shoutouts. We would love to see firsthand what you are using ``pySYD`` for!
+
+-----
