+++ docs/questions.rst
@@ -1,0 +1,56 @@
+.. _faq/index:
+
+****
+FAQs
+****
+
+
+
+#. I'm getting an error. What do I do?
+	
+	Please submit a pull request `here <https://github.com/ashleychontos/pySYD/pulls>`_. Be sure to include:
+	  1. the version of ``pySYD`` used, 
+	  2. the full command that was used, 
+	  3. a screenshot of the error,
+	  4. the times series and/or power spectrum that was used,
+	  5. and any other relevant details you can provide. 
+	The GitHub page for ``pySYD`` is: _`https://github.com/ashleychontos/pySYD`_.
+
+
+#. There are a lot of upper and lower bound options, what do they all mean? 
+	
+	- ``--lb``/``--ub``: limits the frequency range of the power spectrum that is used to derive the granulation background and global parameters 
+	- ``--le``/``--ue``: the folded frequency range containing mixed modes to "whiten" and thus better estimate dnu (requires an estimate for dnu as well)
+	- ``--lp``/``--up``: limits the region around the power excess that is used to calculate the autocorrelation function (ACF)
+	- ``--lx``/``--ux``: limits the region used to search for the power excess and estimate numax
+
+
+#. What's the difference between ``--bf``, ``--sw``, ``--sm``, ``--sp`` and ``--se``?
+	
+	- ``--bf``: smooths the power spectrum to calculate red noise components *after* the power excess has been masked out.
+	- ``--sw``: smooths the power spectrum in the first module when ``pySYD`` searches for the power excess. This value does not affect the power spectrum in the second module. The default value is 50 muHz.
+	- ``--sm``: smooths the power spectrum in the second module when ``pySYD`` measures numax as the center of the Gaussian around the power excess.
+	- ``--sp``: smooths the power spectrum in panel 5 that is used to calculate the ACF in panel 6. The default value is 2.5 muHz.
+	- ``--se``: smooths the echelle diagram in panel 8. To use this argument, specify the smoothing box filter in muHz.
+
+#. I'm not happy with the dnu that ``pySYD`` has measured. How can I change dnu?
+	
+	The best way to change the dnu measurement is to specify a value of dnu using the ``--dnu`` flag (ie. ``--dnu 42.1``). If the ACF contains many peaks around its higher amplitude peaks due to low signal-to-noise data, you can smooth the power spectrum that is used to calculate the ACF by changing the ``--sp`` flag (ie. ``--sp 3.5``). Alternatively, if the peak corresponding to the best dnu measurement is not one of the five highest peaks, you can change the number of peaks found in the ACF by specifying ``--npeaks`` (ie. ``--npeaks 10``). 
+	
+	If the power spectrum contains mixed modes that mask the dnu measurement, you can remove these mixed modes by specifying the lower and upper bound of where the mixed modes are found in the echelle diagram using the following flags: ``--le`` and ``--ue``. Refer to the advanced usage section below for an example on how to remove mixed modes. **This feature is coming soon!**
+
+	Lastly, you can also use ``--xx`` flag to find the dnu manually. This uses the `Echelle <https://github.com/danhey/echelle>`_ package to faciliate dnu measurement. **This feature is coming soon!** 
+
+#. There's an artefact in my data preventing an accurate numax and/or dnu measurement. What can I do?
+	
[…]
