+++ docs/overview.rst
@@ -1,0 +1,172 @@
+.. _overview/index:
+
+********
+Overview
+********
+
+
+``pySYD`` is a python-based implementation of the IDL-based SYD pipeline 
+`(Huber et al. 2009) <https://ui.adsabs.harvard.edu/abs/2009CoAst.160...74H/abstract>`_, 
+which was extensively used to measure asteroseismic parameters for Kepler stars. 
+Papers based on asteroseismic parameters measured using the ``SYD`` pipeline include 
+`Huber et al. 2011 <https://ui.adsabs.harvard.edu/abs/2011ApJ...743..143H/abstract>`_, 
+`Chaplin et al. 2014 <https://ui.adsabs.harvard.edu/abs/2014ApJS..210....1C/abstract>`_, 
+`Serenelli et al. 2017 <https://ui.adsabs.harvard.edu/abs/2017ApJS..233...23S/abstract>`_ 
+and `Yu et al. 2018 <https://ui.adsabs.harvard.edu/abs/2018ApJS..236...42Y/abstract>`_.
+
+``pySYD`` adapts the well-tested methodology from ``SYD`` while also improving these 
+existing analyses and expanding upon numerous new (optional) features. Improvements include:
+
+- Automated best-fit background model selection
+- Parallel processing
+- Easily accessible + command-line friendly interface
+- Ability to save samples for further analyses
+
+
+.. _modes:
+
+Modes
+========
+
+There are currently five operational ``pySYD`` modes: 
+
+#. ``setup`` : Initializes ``pysyd.pipeline.setup`` for quick and easy setup of directories, files and examples. This mode only
+   inherits higher level functionality and has limited CLI (see :ref:`parent parser<parentparse>` below). Using this feature will
+   set up the paths and files consistent with what is recommended and discussed in more detail below.
+
+#. ``load`` : Loads in data for a single target through ``pysyd.pipeline.load``. Because this does handle data, this has 
+   full access to both the :ref:`parent<parentparse>` and :ref:`main parser<mainparse>`.
+
+#. ``run`` : The main pySYD pipeline function is initialized through ``pysyd.pipeline.run`` and runs the two core modules 
+   (i.e. ``find_excess`` and ``fit_background``) for each star consecutively. This mode operates using most CLI options, inheriting
+   both the :ref:`parent<parentparse>` and :ref:`main parser<mainparse>` options.
+
+#. ``parallel`` : Operates the same way as the previous mode, but processes stars simultaneously in parallel. Based on the number of threads
+   available, stars are separated into groups (where the number of groups is exactly equal to the number of threads). This mode uses all CLI
+   options, including the number of threads to use for parallelization (:ref:`see here<parallel>`).
+
+#. ``test`` : Currently under development.
+
+=========================
+
+.. _structure:
+
+Structure
+=============
+
+We recommend using the following structure under three main directories, which is discussed 
+in more detail below:
+
+#. **info/** : [optional input] directory to provide prior information on the processed stars
+#. **data/** : [required input] directory containing the light curves and power spectra
+#. **results/** : [optional output] directory for result figures and files
+
+Input
+********
+
+Info/
+++++++++
+
+There are two main files provided:
+
+#. **info/todo.txt** : text file with one star ID per line, which must match the data ID. If no stars are specified via command line, the star(s) listed in the text file will be processed by default. This is recommended when running a large number of stars.
+
+#. **info/star_info.csv** : contains individual star information. Star IDs are crossmatched with this list and therefore, do not need to be in any particular order. In order to read information in properly, it **must** contain the following columns:
+
+   * "stars" [int] : star IDs that should exactly match the star provided via command line or in todo.txt
[…]
