+++ docs/examples.rst
@@ -1,0 +1,354 @@
+********
+Examples
+********
+
+If you ran the ``pysyd setup`` was successfully executed, there should now be light curves and power spectra 
+for three KIC stars in the **data/** directory. If so, then you are ready to test out the software!
+
+High SNR Examples
+===================
+
+We will start with some of the easiest examples, which have very high signal-to-noise (SNR) detections and then ease into
+the more difficult detections and/or situations. 
+
+Below are three examples of medium to high signal-to-noise (SNR) detections for stars of different evolutionary states. See the 
+:ref:`examples/description` for more details about each panel in the output figures.
+
+KIC 1435467
+*************
+
+KIC 1435467 is our least evolved example star, with numax ~1400 muHz. The following command:
+
+
+    pysyd run --star 1435467 -dv
+    
+    
+    ------------------------------------------------------
+    Target: 1435467
+    ------------------------------------------------------
+    # LIGHT CURVE: 37919 lines of data read
+    # Time series cadence: 59 seconds
+    # POWER SPECTRUM: 99518 lines of data read
+    # PS is oversampled by a factor of 5
+    # PS resolution: 0.426868 muHz
+    ------------------------------------------------------
+    Estimating numax:
+    PS binned to 189 datapoints
+    Numax estimate 1: 1430.02 +/- 72.61
+    S/N: 2.43
+    Numax estimate 2: 1479.46 +/- 60.64
+    S/N: 4.87
+    Numax estimate 3: 1447.42 +/- 93.31
+    S/N: 13.72
+    Selecting model 3
+    ------------------------------------------------------
+    Determining background model:
+    PS binned to 419 data points
+    Comparing 6 different models:
+    Model 0: 0 Harvey-like component(s) + white noise fixed
+    Model 1: 0 Harvey-like component(s) + white noise term
+    Model 2: 1 Harvey-like component(s) + white noise fixed
+    Model 3: 1 Harvey-like component(s) + white noise term
+    Model 4: 2 Harvey-like component(s) + white noise fixed
+    Model 5: 2 Harvey-like component(s) + white noise term
+    Based on BIC statistic: model 2
+     **background-corrected PS saved**
+    ------------------------------------------------------
+    Output parameters:
+    tau_1: 233.71 s
+    sigma_1: 87.45 ppm
+    numax_smooth: 1299.56 muHz
+    A_smooth: 1.75 ppm^2/muHz
+    numax_gauss: 1345.03 muHz
+    A_gauss: 1.49 ppm^2/muHz
+    FWHM: 291.32 muHz
+    dnu: 70.63 muHz
+    ------------------------------------------------------
+     - displaying figures
+     - press RETURN to exit
+     - combining results into single csv file
+    ------------------------------------------------------
+
+
+runs KIC 1435467 using the default method, which first estimates numax before deriving any parameters. 
+The reason for this is that the frequency region with power excess is masked out to better estimate 
+background parameters.
+
+Additional commands used in this example (and what they each mean):
+
+ - ``--ux 5000`` is the upper frequency bound of the power spectrum used during the first module 
+   (i.e. 'x' for excess, ``--lx`` would be the same but the lower bound for this module). These bounds  
+   are used strictly for computational purposes and do not alter or change the power spectrum in any way.
+ - ``--ie`` turns the bicubic interpolation on when plotting the \'echelle diagram. This is 
+   particularly helpful for lower SNR examples like this. 
+ - ``-dv`` == ``-d`` + ``-v`` -> single hashes are reserved for boolean arguments, which correspond to 
+   ``display`` and ``verbose``, respectively. Since ``pySYD`` is optimized for many stars, both of these
[…]
