### 2022-03-08 4de2e52
Create cli.rst
