+++ docs/guide/index.rst
@@ -1,0 +1,12 @@
+User Guide
+===========
+
+.. toctree::
+   :maxdepth: 1
+
+   cli
+   pipeline
+   target
+   models
+   utils	      
+   plots
