+++ docs/guide/models.rst
@@ -1,0 +1,7 @@
+Models
+======
+Frequency domain distributions (i.e. Gaussian, Lorentzian, Harvey, etc.)
+used in the pySYD pipeline.
+
+.. automodule:: pysyd.models
+   :members:
