### 2022-03-08 fe5cfbd
Create pipeline.rst
