+++ docs/guide/models.rst
@@ -1,5 +1,7 @@
-Models
-======
+*************************
+Models [``pysyd.models``]
+*************************
+
 Frequency domain distributions (i.e. Gaussian, Lorentzian, Harvey, etc.)
 used in the pySYD pipeline.
 
