### 2022-03-08 825b7b2
Create plots.rst
