+++ docs/guide/plots.rst
@@ -1,0 +1,8 @@
+***********************
+Plots [``pysyd.plots``]
+***********************
+
+All plotting routines.
+
+.. automodule:: pysyd.plots
+   :members:
