+++ docs/guide/target.rst
@@ -1,0 +1,8 @@
+*************************
+Target [``pysyd.target``]
+*************************
+
+The main pipeline class object that is initialized for each processed star.
+
+.. automodule:: pysyd.target
+   :members:
