+++ docs/advanced/guide/pipeline.rst
@@ -1,10 +1,0 @@
-.. _pipeline:
-
-Pipeline
-========
-Used with the command line interface (CLI), which will accept the command line arguments
-and initialize the appropriate module. This includes the main pipeline initialization, which
-now has parallelization capabilities.
-
-.. automodule:: pysyd.pipeline
-   :members:
