+++ docs/advanced/guide/index.rst
@@ -1,14 +1,0 @@
-.. _user_guide:
-
-User Guide
-===========
-
-.. toctree::
-   :maxdepth: 1
-
-   cli
-   pipeline
-   target
-   models
-   utils	      
-   plots
