+++ docs/advanced/guide/target.rst
@@ -1,8 +1,0 @@
-.. _target:
-
-Target
-======
-The main pipeline class object that is initialized for each processed star.
-
-.. automodule:: pysyd.target
-   :members:
