+++ docs/advanced/advanced.rst
@@ -1,131 +1,0 @@
-.. _advanced/index:
-.. role:: bash(code)
-   :language: bash
-
-**************
-Advanced Usage
-**************
-
-
-Below are examples of how to use specific ``pySYD`` features, as well as plots showing results before and after their usage.
-
-
-``--dnu: force dnu``
-++++++++
-
-+-------------------------------------------------+---------------------------------------------------------+
-| Before                                          | After                                                   |
-+=================================================+=========================================================+
-| Fix the dnu value if the desired dnu is not automatically selected by `pySYD`.                            |
-+-------------------------------------------------+---------------------------------------------------------+
-|:bash:`pysyd run --star 9512063 --numax 843`     |:bash:`pysyd run --star 9512063 --numax 843 --dnu 49.54` |
-+-------------------------------------------------+---------------------------------------------------------+
-| .. figure:: figures/advanced/9512063_before.png | .. figure:: figures/advanced/9512063_after.png          |
-|    :width: 600                                  |    :width: 600                                          |
-+-------------------------------------------------+---------------------------------------------------------+
-
-
-``--ew: excess width``
-++++++++
-
-+------------------------------------------------------------------+------------------------------------------------------------------+
-| Before                                                           | After                                                            |
-+==================================================================+==================================================================+
-| Changed the excess width in the background corrected power spectrum used to calculate the ACF (and hence dnu).                      |
-+------------------------------------------------------------------+------------------------------------------------------------------+
-| :bash:`pysyd run --star 9542776 --numax 900`                     | :bash:`pysyd run --star 9542776 --numax 900 --ew 1.5`            |
-+------------------------------------------------------------------+------------------------------------------------------------------+
-| .. figure:: figures/advanced/9542776_before.png                  | .. figure:: figures/advanced/9542776_after.png                   |
-|    :width: 600                                                   |    :width: 600                                                   |
-+------------------------------------------------------------------+------------------------------------------------------------------+
-
-
-``--ie: smooth echelle``
-++++++++
-
-+------------------------------------------------------------------+------------------------------------------------------------------+
-| Before                                                           | After                                                            |
-+==================================================================+==================================================================+
-| Smooth echelle diagram by turning on the interpolation, in order to distinguish the modes for low SNR cases.                        |
-+------------------------------------------------------------------+------------------------------------------------------------------+
-| :bash:`pysyd run 3112889 --numax 871.52 --dnu 53.2`              | :bash:`pysyd run --star 3112889 --numax 871.52 --dnu 53.2 --ie`  |
-+------------------------------------------------------------------+------------------------------------------------------------------+
-| .. figure:: figures/advanced/3112889_before.png                  | .. figure:: figures/advanced/3112889_after.png                   |
-|    :width: 600                                                   |    :width: 600                                                   |
-+------------------------------------------------------------------+------------------------------------------------------------------+
-
-
-``--kc: Kepler correction``
-++++++++
-
-+------------------------------------------------------------------+------------------------------------------------------------------+
-| Before                                                           | After                                                            |
-+==================================================================+==================================================================+
-| Remove *Kepler* artefacts from the power spectrum for an accurate numax estimate.                                                   |
-+------------------------------------------------------------------+------------------------------------------------------------------+
-| :bash:`pysyd run --star 8045442 --numax 550`                     | :bash:`pysyd run --star 8045442 --numax 550 --kc`                |
-+------------------------------------------------------------------+------------------------------------------------------------------+
-| .. figure:: figures/advanced/8045442_before.png                  | .. figure:: figures/advanced/8045442_after.png                   |
-|    :width: 600                                                   |    :width: 600                                                   |
-+------------------------------------------------------------------+------------------------------------------------------------------+
-
-
-``--lp: lower frequency of power excess``
-++++++++
-
[…]
