+++ docs/advanced/guide/utils.rst
@@ -1,10 +1,0 @@
-.. _utils:
-
-Utilities
-=========
-Container classes, parameter dictionaries, functions related to file loading and/or
-saving as well as various data manipulation methods (i.e. correcting artefacts, 
-binning data, etc.).
-
-.. automodule:: pysyd.utils
-   :no-inheritance-diagram:
