+++ docs/advanced/guide/plots.rst
@@ -1,8 +1,0 @@
-.. _plots:
-
-Plots
-=====
-All plotting routines.
-
-.. automodule:: pysyd.plots
-   :members:
