+++ docs/advanced/guide/models.rst
@@ -1,9 +1,0 @@
-.. _models:
-
-Models
-======
-Frequency domain distributions (i.e. Gaussian, Lorentzian, Harvey, etc.)
-used in the pySYD pipeline.
-
-.. automodule:: pysyd.models
-   :members:

[…]
