+++ docs/introduction/contributing.rst
@@ -1,3 +1,0 @@
-************
-Contributing
-************

[…]
