+++ docs/glossary.rst
@@ -1,3 +1,53 @@
-********
-Glossary
-********
+*****************
+Glossary of terms
+*****************
+
+asteroseismology
+    the study of oscillations in stars
+
+
+solar-like oscillations (p-mode oscillations)
+    implied in the name, these oscillations are driven by the same mechanism as that observed in the Sun, which is
+    due to turbulent, near-surface convection. They are also sometimes referred to as p-mode oscillations, after the
+    pressure-driven (or acoustic sound) waves that are resonating in the stellar cavity.
+
+
+global asteroseismic parameters
+    the term 'global' is used to describe the general properties of the observed oscillations and is not associated with
+    the detailed frequency analysis of individual oscillation modes, a process referred to as peakbagging. Traditionally
+    the two main global asteroseismic properties are numax and dnu, both of which are described in more detail in their
+    respective entry.
+    
+    there are two distinct features of solar-like oscillations that enable the measurement of the two main global 
+    properties, numax and dnu. The stochastic nature of convection leads to oscillation modes over a range of frequencies, 
+    where the envelope of the observed modes is approximately Gaussian and the frequency corresponding to the middle of
+    peak of this Gaussian-like envelope is referred to as numax. The second feature is the comb pattern or regular spacing
+    between different modes, which is referred to as the characteristic frequency spacing or dnu. Therefore, the term 'global' 
+    is used to describe the general properties of the oscillations, like the center and amplitude of the Gaussian-like envelope. The second 
+    distinct feature is the comb pattern or regular spacing between different modes, which is
+    referred to as the characteristic frequency spacing or dnu. For purposes of our analyses, global asteroseismic 
+    parameters regular spacing or combP-mode oscillations
+    In addition 
+    to the center of the frequency range (numax), there is a regular spacing or comb pattern between the observed modes 
+    that is referred to as the characteristic spacing or dnu.
+
+
+frequency of maximum power (numax)
+    the frequency corresponding to maximum power, which is roughly the center of the Gaussian-like envelope of oscillations
+    
+    scales with evolutionary state, logg, acoustic cutoff
+
+characteristic frequency separation (dnu)
+    the comb patter or regular spacing observed for solar-like oscillations for different modes
+    
+    scales with mean stellar density
+
+
+*Kepler* legacy sample
+    a sample of well-studied, oscillating *Kepler* stars (cite)
+
+
+``SYD``
+    the well-known IDL-based asteroseismic pipeline created by Dan Huber during his PhD in Sydney (hence SYD). ``SYD``
+    has been extensively tested and benchmarked to other closed-source asteroseismic tools on *Kepler* stars.
+
