### 2022-03-08 b76819a
Update citing.rst
