### 2022-03-08 5bb0319
Update citing.rst
