### 2022-03-08 89b39c4
Update citing.rst
