### 2022-03-09 #3331 @mbobra
Thanks for the update, @ashleychontos! To answer your question -- it depends.

The JOSS reviewers will go through the checklist above and to determine whether pySYD satisfies the review criteria. If pySYD does not satisfy some given criterion, then they will open an issue on the repo. Those issues must be resolved before pySYD passes review. However, you absolutely do not need to resolve other issues on the repo unrelated to the review criteria. Does that answer the question?
