+++ docs/installation.rst
@@ -9,10 +9,6 @@
 Installation
 ************
 
-.. #############
-   Miscellaneous
-   #############
-
 
 With ``pip``
 ************
@@ -70,9 +66,9 @@ from inside the cloned ``pySYD`` directory.
 
 .. _installation/dependencies:
 
-############
+
 Dependencies
-############
+************
 
 This package has the following dependencies:
 
@@ -88,7 +84,7 @@ and `setup.cfg <https://github.com/ashleychontos/pySYD/setup.cfg>`_. However, us
 ``conda`` should install and enforce these versions automatically. 
 
 Optional
-********
+########
 
 If using the sampling feature and you want a progress bar, you'll need to install:
 
@@ -98,15 +94,14 @@ If using the sampling feature and you want a progress bar, you'll need to instal
 
 .. _installation/testing:
 
-#######
 Testing
-#######
+*******
 
 You can test your installation by using the help command: 
     
 .. dropdown:: pysyd --help
     
-    | usage: pySYD [-h] [-version] {load,parallel,run,setup,test} ...
+    |`` usage: pySYD [-h] [-version] {load,parallel,run,setup,test} ...
     | 
     | pySYD: Automated Extraction of Global Asteroseismic Parameters
     |
@@ -120,15 +115,14 @@ You can test your installation by using the help command:
     |     parallel            Run pySYD in parallel
     |     run                 Run the main pySYD pipeline
     |     setup               Easy setup of relevant directories and files
-    |     test                Test different utilities (currently under development)
+    |     test                Test different utilities (currently under development)``
 
 -----
 
 .. _installation/setup:
 
-#####
 Setup
-#####
+*****
 
 The easiest way to start using the ``pySYD`` package is by running our ``setup`` feature
 from a convenient directory:
@@ -147,39 +141,39 @@ The optional verbose command can be called with the setup feature:
 
 .. dropdown:: pysyd setup --verbose
     
-    | Downloading relevant data from source directory:
+    |`` Downloading relevant data from source directory:
     | 
     | /Users/ashleychontos/Desktop/info
     |   % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
-    |                                                                   Dload  Upload   Total   Spent    Left  Speed
+    |                                    Dload  Upload   Total   Spent    Left  Speed
     | 100    25  100    25    0     0     49      0 --:--:-- --:--:-- --:--:--    49
     |   % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
-    |                                                                   Dload  Upload   Total   Spent    Left  Speed
+    |                                    Dload  Upload   Total   Spent    Left  Speed
     | 100   239  100   239    0     0    508      0 --:--:-- --:--:-- --:--:--   508
     |   % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
-    |                                                                   Dload  Upload   Total   Spent    Left  Speed
+    |                                    Dload  Upload   Total   Spent    Left  Speed
     | 100 1518k  100 1518k    0     0  1601k      0 --:--:-- --:--:-- --:--:-- 1601k
     |   % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
-    |                                                                   Dload  Upload   Total   Spent    Left  Speed
+    |                                    Dload  Upload   Total   Spent    Left  Speed
     | 100 3304k  100 3304k    0     0  2958k      0  0:00:01  0:00:01 --:--:-- 2958k
     |   % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
-    |                                                                   Dload  Upload   Total   Spent    Left  Speed
[…]
