+++ docs/index.rst
@@ -4,9 +4,9 @@
    
 .. raw:: html
 
-     <style> .font {font-family:'Courier New'} .red {color:red} </style>
+     <style> .font {font-family:'Courier New'} </style><style> .red {color:red} </style>
 
-.. role:: redfont
+.. role:: font
 
 .. module:: pysyd
 
@@ -19,7 +19,7 @@ for studying the internal structure of stars and determining their fundamental
 properties. Specifically for stars similar to the Sun, turbulent near-surface 
 convection excites sound waves that propagate within the stellar cavity, and hence
 providing powerful constraints on stellar interiors that are inaccessible by 
-any other means. :redfont:`While it is well-established and widely-accepted as the gold 
+any other means. :font:`While it is well-established and widely-accepted as the gold 
 standard for characterizing fundamental stellar properties (e.g., masses, radii,
 ages, etc.), asteroseismology is still mostly a niche field and closed-door science.` 
 
