### 2022-03-09 b149246
Update and rename citing.rst to attribution.rst
