+++ docs/installation.rst
@@ -1,7 +1,7 @@
 .. link-button:: installation/quickstart
     :type: ref
     :text: Jump down to quickstart to get started right away!
-    :classes: btn-outline-primary btn-block
+    :classes: btn-outline-secondary btn-block
 
 .. _installation/index:
 
@@ -78,17 +78,11 @@ This package has the following dependencies:
 * `Matplotlib <https://matplotlib.org/index.html#module-matplotlib>`_
 * `Astropy <https://www.astropy.org>`_
 * `scipy <https://docs.scipy.org/doc/>`_
+* `tqdm <https://tqdm.github.io>`_
 
 Explicit version requirements are specified in the project `requirements.txt <https://github.com/ashleychontos/pySYD/requirements.txt>`_ 
 and `setup.cfg <https://github.com/ashleychontos/pySYD/setup.cfg>`_. However, using ``pip`` or 
 ``conda`` should install and enforce these versions automatically. 
-
-Optional
-########
-
-If using the sampling feature and you want a progress bar, you'll need to install:
-
-* `tqdm <https://tqdm.github.io>`_
 
 -----
 
@@ -101,21 +95,21 @@ You can test your installation by using the help command:
     
 .. dropdown:: pysyd --help
     
-    |`` usage: pySYD [-h] [-version] {load,parallel,run,setup,test} ...
-    | 
-    | pySYD: Automated Extraction of Global Asteroseismic Parameters
+    |`` usage: pySYD [-h] [-version] {load,parallel,run,setup,test} ...``
     |
-    | optional arguments:
-    |   -h, --help            show this help message and exit
-    |   -version, --version   Print version number and exit.
+    |`` pySYD: Automated Extraction of Global Asteroseismic Parameters``
+    |
+    |`` optional arguments:``
+    |``   -h, --help            show this help message and exit``
+    |``   -version, --version   Print version number and exit.``
     | 
-    | pySYD modes:
-    |   {load,parallel,run,setup,test}
-    |     load                Load in data for a given target
-    |     parallel            Run pySYD in parallel
-    |     run                 Run the main pySYD pipeline
-    |     setup               Easy setup of relevant directories and files
-    |     test                Test different utilities (currently under development)``
+    |`` pySYD modes:``
+    |``   {load,parallel,run,setup,test}``
+    |``     load                Load in data for a given target``
+    |``     parallel            Run pySYD in parallel``
+    |``     run                 Run the main pySYD pipeline``
+    |``     setup               Easy setup of relevant directories and files``
+    |``     test                Test different utilities (currently under development)``
 
 -----
 
@@ -141,39 +135,39 @@ The optional verbose command can be called with the setup feature:
 
 .. dropdown:: pysyd setup --verbose
     
-    |`` Downloading relevant data from source directory:
+    |`` Downloading relevant data from source directory:``
     | 
-    | /Users/ashleychontos/Desktop/info
-    |   % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
-    |                                    Dload  Upload   Total   Spent    Left  Speed
-    | 100    25  100    25    0     0     49      0 --:--:-- --:--:-- --:--:--    49
-    |   % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
-    |                                    Dload  Upload   Total   Spent    Left  Speed
-    | 100   239  100   239    0     0    508      0 --:--:-- --:--:-- --:--:--   508
-    |   % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
[…]
