### 2022-04-26 #3331 @ashleychontos
Without further ado!

@mbobra, @benjaminpope, @danhey,

Thank you all for volunteering your time to review the software, and *we appreciate your patience*. Since both reviewers followed similar formats, we decided to group comments together by sections, with the hope that it will tell a more cohesive story. Since a long time has passed since our original submission, we have included the reviewers comments as a reference.
______________________________________________________________________________

## Intro Comments
- **[BP:]** Thanks for inviting me to review the pySYD release - this is an important piece of software that will bring asteroseismology from the closed-source dark ages to an open source renaissance (I hope!). IDL-SYD is well-used and well-known, but together with some other major pipelines has not been available open-source, which means that asteroseismology as a field suffers from reproducibility and inaccessibility. pySYD will be an important step forward that, I am pretty sure, will enable a great deal of exciting science.
- **[BP:]** The release of pySYD as documented so far isn't quite where we need to be for JOSS or for optimal use for the community. In particular, there are two high level issues:
- **[BP:]** The testing of the pySYD software against IDL-SYD is not documented in this paper, or cited from elsewhere, so there is actually a gap in reproducibility introduced.
- **[BP:]** The documentation of pySYD is not quite at the stage where it will be optimally user-friendly, and we probably need to have a bit more hand-holding with the examples to help people like me pick it up and play.
It is also not clear to me whether this is intended as a purely command-line tool, or whether the API is sufficiently useful for working with other code, for example in a Jupyter notebook or as part of a larger script. If this is beyond the scope of the project it is fine, but I didn't get a sense from the paper or the documentation what kind of workflow was expected, supported, or possible.
- **[DH:]** Hi all, this has been a pleasure to review! It's a great piece of software, and I have no doubt it will be extremely useful for asteroseismology. Overall, I believe pySYD satisfies the JOSS criteria for publication after a few points have been addressed.
- **[DH:]** I agree with all of Ben's comments, especially regarding the distinction between a command-line vs. API tool. It appears to me that the functionality is there for using pySYD from say, a Jupyter notebook, but this has not been documented.
- **[DH:]** I will try to structure my review in the same format as Ben’s.

**Response:** We understand your confusion(s) re: the distinction between being a command-line tool vs. something more modular (using the API as a guide) and have therefore clarified this better in both the documentation and the manuscript. `pySYD` was initially developed to be a command-line tool but became more modular organically with time. Therefore, we have added some [Jupyter notebooks](https://pysyd.readthedocs.io/en/latest/usage/nb/index.html) to demonstrate this usage as well. I will address the `pySYD` vs `SYD` comments in more detail in the last section, where other specific comments and questions were made regarding the comparison.
______________________________________________________________________________

## Installation
- **[BP:]** pip install works, and pysyd setup works too.
- **[BP:]** Could pysyd setup put data in an install directory, rather than in the current directory?
- **[DH:]** Installation is fine from a fresh Conda environment. However, tqdm appears to be a requirement that is not listed in pySYD.

**Response:** The `tqdm` package was an oversight and we have now added the package to the relevant locations (i.e. documentation and Github repo) to avoid any future confusion. There was already an optional flag that existed for putting data from setup into an arbitrary directory and have now implemented an even more efficient way to do this via ``utils._fix_init(new_path)``. So in the event that you would want to change it, run the setup feature with the ``-f`` (or ``fix``) flag first with your desired path — which if `None` is provided, it will locate the install directory and use that as the new default path.
______________________________________________________________________________

## Unit tests
- **[BP:]** As far as I can tell, there are no unit tests or continuous integration solutions, at least as-is on the GitHub repo. The JOSS review requires automated tests.
- **[DH:]** I agree with Ben here, there don't appear to be any unit tests! It's important to have unit tests, especially for a tool that will hopefully become a standard in asteroseismology. I suggest at the minimum, you include tests for the core functionality (i.e., input a star -> ensure the outputs are exactly the same). If you need some inspiration on how testing is done for an astronomy Python package, Lightkurve does a very good job of it. Check out their [tests](https://github.com/lightkurve/lightkurve/tree/main/tests), and their [GitHub actions workflow](https://github.com/lightkurve/lightkurve/blob/main/.github/workflows/lightkurve-tests.yml).

**Response:** The comments made here are great points but beyond the scope of the project at this juncture, although it is definitely a long-term goal of the package. The features currently in place do test the core functionality of the software by exactly reproducing values and uncertainties (to the hundredth decimal place) for global parameters numax and dnu for our three example stars and hence, should suffice according to JOSS standards. We understand that this was partly due to the confusion of the examples for our original submission, but we have since addressed this by implementing and saving seeds for every random process, ensuring reproducible results.
______________________________________________________________________________

## Repository
- **[BP:]** There are no community guidelines as required by JOSS.
- **[DH:]** No community guidelines.
- **[DH:]** Perhaps a minor oversight but pySYD doesn't seem to specify which Python version(s) it is compatible with. I know many astronomers still use Python 2.7 for compatibility, will pySYD work with that?

**Response:** We have added community guidelines (using GitHub’s recommended standards as an example) to the GitHub repo and the documentation. While we are fairly certain that the Python version (>=3) was specified and enforced in the pip install (via setup.py), we have added it to the [requirements.txt](https://github.com/ashleychontos/pySYD/requirements.txt) and explicitly state its [dependency](https://pysyd.readthedocs.io/en/latest/installation.html) in the documentation.
______________________________________________________________________________

### Examples & functionality
- **[BP:]** We get nice and fast performance from pysyd run -star 1435467 -show -verbose as expected from the Getting Started / Example Fit part of the readthedocs. On this page, you should include the expected output though - so we know if the graphs are working ok.
- **[BP:]** The output from this example differs from the one on the page! I understand this is because of the differing -mc flag, but I'm not sure why one command is shown in full (with no output), and the other output is shown in full (but no command).
- **[BP:]** The Quickstart example should have an example output. Also not clear how this is different from Example Fit.
The examples are nice graphs... but supposing we did pip install, it's not immediately obvious where the data live, and no commands are given to reproduce the examples. So after "you are ready to test out the software!" there should be a few worked examples showing commands and outputs!
- **[BP:]** The figures take a very long time to render and close, at least on my MacBook Pro. Perhaps the examples might like to have -save flags by default so that these don't cause an issue.
- **[DH:]** The examples all work fine for me.
- **[DH:]** it is not clear in the documentation that the plots must be closed for the pipeline to proceed. i.e., when I ran the “pysyd run -star 1435467 -show -verbose” command, I had to close the plots as they appeared for the pipeline to continue. They should either automatically close or a prompt should be displayed for continuing, or the pipeline should just continue and create new plots as it goes.
- **[DH:]** Would it be prudent to include an example for the TESS data? And on that note, how does PySyd deal with significant gaps in the data? This is kind of a new issue because the Kepler data was so good! I’ve noticed that there’s a pysyd.utils.stitch_data() function for gapped data, perhaps a warning should also be thrown when this function is used? This could be fixed in one step: provide an example for a TESS star that has a large data gap in it!

**Response:** Our apologies for the confusion surrounding the examples and expected outputs. As I mentioned before, `pySYD` now automatically implements and saves seeds for any locally-processed stars that undergo any random process in order to prevent this from happening in the future. The goal is to eventually include all the “verbose output” in a single config file (and do away with the long command-line output) — the config file which will contain any information needed to reproduce the results (e.g., something similar to what [radvel](https://github.com/California-Planet-Search/radvel) does with the configparser function). That will likely take care of a majority of the confusion mentioned in this section. Finally, with regard to the new problem of TESS data gaps, the software will process a star the same way irrespective of the instrument (i.e it will not use the stitching function), but of course any power spectrum computed from a time series with large gaps will take longer — which we currently do provide a warning for. In addition, we also print a warning if a person opts to use the data stitching function.
______________________________________________________________________________

## Manuscript
- **[BP:]** Summary - this probably needs to be reworked. Currently it is a summary of asteroseismology, but not of pySYD. You can cut down some of the explanation of physics if necessary, but certainly need to add a bit about what pySYD is and what you use it for.
- **[BP:]** In first paragraphs - there is a comment that SYD has been tested against closed-source pipelines. It is my understanding that SYD itself is not yet open source, and that this very paper is the open-source release - so you want to clarify whether SYD too is closed-source.
- **[BP:]** Should cite Borucki and Ricker papers for Kepler and TESS citations.
- **[BP:]** The comparison of pySYD and IDL-SYD shows nonzero scatter... but this is ostensibly a translation of the IDL-SYD pipeline, so I would expect them to be even closer. This is important because if reproducibility of legacy Kepler results is a goal, systematics in IDL-SYD should also be reproduced! If it is a reworking of the same ideas, but not a translation, this will differ. There should probably be some comments about this. 0.5% is not quite trivial, even if the uncertainties are typically of order ~ %, if you average over many stars.
- **[BP:]** The IDL-SYD to pySYD benchmarks are not open source - are these linked anywhere? This comparison is not reproducible - we don't know which stars these are, or what the exact time baselines are (3-11 months is quite a range). I think this comparison is an important part of the paper, given pySYD is about reproducibility and extension of the IDL-SYD pipeline.
- **[BP:]** I think it should say earlier in the paper that this is about solar-like oscillations, as opposed to (say) classical pulsations.
- **[DH:]** Overall the manuscript is well written and clear, I like it!
- **[DH:]** The scatter in the IDL-SYD and pySYD comparison plots also concerns me a bit. If the goal is to completely reproduce the functionality of the IDL version then there should be little to no scatter. Do we know what causes these deviations, or, are they purely the result of random sampling? And which pipeline output would we expect to give the more accurate result?
- **[DH:]** Regarding the open sourcing of the benchmarks:
Only a suggestion, but would it be easier to instead write a small utility function in pySYD that calculates the power spectrum given an input light curve? This would also be useful for the general functionality of PySYD, allowing people to supply a light curve. It would also benefit reproducibility: I have seen a lot of different power spectra normalizations in my time …

We have addressed all minor text comments in the manuscript. Regarding the comparison between `SYD` and `pySYD` — `pySYD` was initially *meant* to be a direct 1:1 translation but because it comes with many enhancements — the answers will not match exactly. We improved the comparison by re-running the sample and setting parameters more closely to those used in the `SYD` analyses and found much better agreement, with no significant offset in the median values. We also added a [folder](https://github.com/ashleychontos/pySYD/tree/master/tests/benchmark) to the GitHub page that contains all the relevant information to recreate the comparison, including the i) list of targets, ii) a shell script of all the executed commands from the pySYD end, iii) results from each of the two pipelines and iv) scripts to regenerate the figure. Finally, we also provide instructions for how to access and process the data, the light curves which are all publicly available on [KASOC](https://kasoc.phys.au.dk/).
______________________________________________________________________________

Of course there is always more that can be done but we have finally reached a point which we are confident about and therefore, encourage you all to take a look at the new versions (`v.5.9.4`) and get back to us if you have any questions!

Cheers,

@ashleychontos, @danxhuber, @MaryumSayeed & @pavyamsiri
