### 2022-04-27 #3331 @benjaminpope
Dear @ashleychontos,

Thanks for the update.

I like the new Jupyter notebook demos and explanation in the body of the paper.

Before I delve into the review - the paper as rendered above doesn't contain any figures, but they are referred to in the text. There is a broken hyperref at line 78 " (section )" and that entire section seems incomplete.

Perhaps the latest version of the paper has not yet been uploaded?

As another comment, re unit tests, it is a requirement of JOSS that you have automated testing. I regret that we can't say it's beyond the scope of the paper - I have to be able to tick the box from JOSS. In your comment above, I don't know which features you're referring to when you say you have tests implemented - on the repo, I see plenty of unit tests in the `tests` folder but no GitHub action to automate these (which would hopefully be easy to add!).

Apologies,

Ben
