### 2022-04-27 #3331 @ashleychontos
Thanks @benjaminpope --

I was going to say that I just double-checked the PDF @editorialbot generated and it is definitely an old, unfinished version. The paper lives (and compiles perfectly fine) in the source repo but I guess it has been so long since the original submission that I thought that's how JOSS papers were generated. Are there instructions for uploading a resubmission to JOSS that I've somehow missed?
