### 2022-04-27 #3331 @ashleychontos
@benjaminpope regarding the automated testing -- I visited the JOSS documentation [here](https://joss.readthedocs.io/en/latest/review_checklist.html) and it says:

- **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

What we have right now currently satisfies the latter of the two, which is what our original response above was referring to.
