### 2022-04-27 #3331 @benjaminpope
As far as manual steps go, I guess that is ok in principle - though I should add it would be very easy to include these as a set of automated tests using a github action to execute the example.

The example you have linked, however, doesn't really have an up to date test functionality of the existing code. The vibe they want in an open source software project is that you should be able to test the software straightforwardly and have it come out with a benchmark answer - it isn't very helpful to have to compare this by eye to a list of numbers when this could be done automatically on the machine. I can't verify what you mean for the hundredth decimal place, either, and it's not described on that page (which isn't itself described as a test page - it is essential that there literally be some page that leads you through tests, at bare minimum, called tests, or billed as tests or something).
