### 2022-04-27 #3331 @benjaminpope
What I mean is - if you really don't want to do unit tests, you could at least make a script that is executed automatically or otherwise, that checks all output numbers against a precomputed list of numbers that it should return.
