### 2022-04-27 #3331 @ashleychontos
I appreciate all the feedback/suggestions @benjaminpope

That makes sense! I was actually in the process of testing that very thing but thought that what we had completed had already met the requirements. FWIW: it's not that I _don't_ want to do unit tests -- I have never done them before and I'm assuming the learning curve starts out a bit more steep, which would therefore take much longer than someone with more experience such as you or @danhey.
