### 2022-04-27 #3331 @benjaminpope
I appreciate that! If you don't want to go all the way to really serious unit tests, a good start would be to do a couple end-to-end where you check outputs of a whole script. As you say, you have code that does this by hand.

The easy way to do this is to set up a `test_end_to_end.py` file under `./tests/` with this script as a function `def test_end_to_end`, or something like that. The function could just execute your script, and compare it to some values you have manually entered as correct - and return True if they are within some tolerance, and False otherwise.

Then you make a `test.yml` file in the directory `.github/workflows/` and go to the Actions tab on the github page to set it up to run whenever you push. See this one my student did for one of our projects: https://github.com/SharmaLlama/ticktack/blob/main/.github/workflows/tests.yml. This is a relatively small package and you can hack this as an easy example.

All it has to include is instructions for installing your code, and where to find the unit tests. You can edit the appropriate lines to point to the appropriate directories. Then it can generate a badge for the README.md front page, just like on that repo I linked - and it will come up green if everything passes.

Really the reason you want automated unit tests is so that third-party contributors can make changes or additions to your software and everyone can be sure it won't break anything. Which is why this is required for JOSS - because you wouldn't submit just any research code to JOSS, but specifically packages you want to develop and use collaboratively in the community. Really the gold standard for this is function by function tests, and I don't know how tightly JOSS specifies what sort of unit testing is required (@mbobra ?) but when I am not developing for wide use, I still like to do a few end-to-end tests to check overall functionality.
