### 2022-04-27 #3331 @benjaminpope
Christina Hedges has some good docs here: https://christinahedges.github.io/astronomy_workflow/notebooks/2.0-testing/pytest.html
