### 2022-05-05 #3331 @danhey
Hi all, sorry for my fiddling with editorialbot. It doesn't seem to want to output the latest paper version.

In any case, @ashleychontos, many thanks for addressing the majority of our issues. I do believe that unit tests are a crucial aspect of software development, more so for what I expect to be a widely used package. Right now, it looks like PySyd has a tests/ folder, but these are copied from another project. Regardless of how development proceeds, these copied tests should be deleted to prevent confusion.

I copy below the JOSS guidelines on testing:
```
- Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)
- OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)
- Bad (not acceptable): No way for you, the reviewer, to objectively assess whether the software works
```

Right now, I see PySyd testing being between OK and Bad. The setup script and example gives us a manual test. But really, all the required functionality is here for automated testing: you can run these scripts (`pysyd run --star 1435467 -dv --ux 5000 --mc 200` etc) inside a test and assert that the output matches what you expect. What I (and Ben!) really want is the ability to clone the repo, `cd` to the `tests/` folder, run `pytest`, and get a simple pass/fail output.  The link that Ben posted is a great resource to understand this.

Otherwise, great work on the software. I want to commend the authors for such a comprehensive documentation, including a glossary(!) and library of related software.
