### 2022-05-09 #3331 @mbobra
@danielskatz Any idea why the editorialbot is not generating the correct pdf (from here: https://github.com/ashleychontos/pySYD/tree/master/paper)?
