### 2022-05-09 #3331 @mbobra
@ashleychontos Have you seen [this astropy developer guide](https://docs.astropy.org/en/latest/development/testguide.html) on unit tests? Check out the simple example under the 'Writing Tests' section. You can write some really simple, one-line tests.

For example: Let's say a function outputs a value for longitude. You can write a unit test to check that the output is between 0 and 360. If it is, the test passes. Otherwise, the test fails. That's it.

Could you add a couple simple unit tests like these? There's no need to go overboard -- a couple tests is good enough, and maybe some documentation with ideas for future tests sometime down the road.
