### 2022-05-10 #3331 @xuanxu
There are two different `paper.md` files in the repository ([here](https://github.com/ashleychontos/pySYD/blob/master/paper/paper.md) and [here](https://github.com/ashleychontos/pySYD/blob/master/paper/copy/paper.md)), the bot is creating the pdf based on the first it finds.
