### 2022-08-19 #3331 @mbobra
👋 @ashleychontos I hope (and assume) the defense went well! I totally understand wanting to take some time afterward -- but could you update us with an expected timeline to close out this review? It doesn't have to be immediate but I think it would help everyone to have a short update.
