### 2022-08-20 #3331 @ashleychontos
Hello @mbobra indeed it did, many thanks!

My start date at Princeton is 9/1. How does one final push for the month of September sound? I know it's not immediate but I am trying to be realistic while I settle in to the new place. I've actually had a couple meetings about the package recently so I've been beginning to ramp this back up again anyway!
