### 2022-10-21 #3331 @ashleychontos
my sincere apologies @mbobra -- this has unintentionally taken [much] longer than expected.

I am currently working through a few updates to the package, which is my **_top priority_** at the moment. I have definitely not forgotten nor plan to let it slip through the cracks. I'd say about another week or two and it will be all set for resubmission!
