### 2022-10-30 #3331 @ashleychontos
Hi all! I'm pleased to say that there is a handy new test feature available, which tests the full suite of methods (i.e. beginning to end of sampling) on 3 example stars and asserts that all answers for our main asteroseismic observables, numax and dnu, are exactly equal to their known values. This can easily be done by:

```
pip install --upgrade pysyd (v6.2.0)
pysyd test
```

Additionally, I've updated the relevant parts of the documentation to reflect these new updates.

I'm almost positive that everything else had already been addressed, that the delay was primarily due to the implementation of tests. Please let me know if there is anything I've managed to overlook, thanks!
