### 2022-10-31 #3331 @benjaminpope
Hi Ashley,

Thanks for getting to this! Glad that you have a test suite going.

I strongly recommend that you use a GitHub Action or similar for applying these tests rather than requiring the user to do so.

There are [good instructions](https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python) from GitHub for doing this, or you can look at existing open source packages that use GitHub Actions for testing.

All the best,

Ben
