### 2022-11-04 #3331 @ashleychontos
Thank you @benjaminpope and @danhey for the help and recommendations, all which have significantly improved the software! This was the first time I've implemented any testing, etc. so it took a bit more time than I expected so I do appreciate everyone's (including @mbobra) patience.

That said, I am happy to report that I have added a number of tests and set them up as a GitHub Action that will trigger any time there is a push/pull request. I think there are 17 tests at the moment, a few which test the full functionality of pySYD from start to finish -- which is why they take ~30 seconds to complete.

Not sure how either of you plan to run it but I just tested a new git clone (**v6.10.0**), setup install and -m pytest -- and everything looks great from my end so I think it's good to go now (yay)!

Thanks again,

Ashley
