### 2022-11-07 #3331 @mbobra
Thank you both so very much for your thorough reviews, @danhey and @benjaminpope! Your volunteer time keeps JOSS running and the JOSS team appreciates it very much ☀️
