### 2022-11-07 #3331 @mbobra
@ashleychontos It looks like the editorialbot identified some missing DOIs. I took a look and they seem correct to me. Do you agree? If so, do you mind updating paper.bib?
