### 2022-11-07 #3331 @ashleychontos
I agree re: DOIs and have updated the .bib file :)
