### 2022-11-07 #3331 @mbobra
Congratulations, @ashleychontos 🥳 It was a long ride but I hope you feel the JOSS review process improved the software. I think the final result turned out awesome. The EiC team @openjournals/joss-eics will take it from here!
