### 2022-11-07 #3331 @ashleychontos
Oh absolutely - I learned more about software and development throughout this process than I ever knew pre-JOSS times combined. Thank you again SO much for everything, especially for riding this out with me!
