### 2022-11-07 #3331 @dfm
@ashleychontos — I've opened a PR with a few formatting edits to the bibliography, but we should be good to go after that: https://github.com/ashleychontos/pySYD/pull/42
