### 2022-11-07 #3331 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
