### 2022-11-07 #3331 @dfm
@danhey, @benjaminpope — many thanks for your reviews here and to @mbobra for editing this submission! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨

@ashleychontos — your paper is now accepted and published in JOSS :zap::rocket::boom:
